@@ -1,22 +1,25 @@
 name: flutter_webrtc
 description: Flutter WebRTC plugin for iOS/Android/Destkop/Web, based on GoogleWebRTC.
-version: 0.9.0-prerelease
+version: 0.11.7
 homepage: https://github.com/cloudwebrtc/flutter-webrtc
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.3.0 <4.0.0'
   flutter: '>=1.22.0'
 
 dependencies:
-  dart_webrtc: ^1.0.4
+  collection: ^1.17.0
+  dart_webrtc: ^1.4.9
   flutter:
     sdk: flutter
   path_provider: ^2.0.2
-  webrtc_interface: ^1.0.2
-  
+  web: ^1.0.0
+  webrtc_interface: ^1.2.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
   import_sorter: ^4.6.0
+  lints: ^4.0.0
   pedantic: ^1.11.1
   test: any
 
@@ -31,4 +34,8 @@ flutter:
       macos:
         pluginClass: FlutterWebRTCPlugin
       windows:
+        pluginClass: FlutterWebRTCPlugin
+      linux:
+        pluginClass: FlutterWebRTCPlugin
+      elinux:
         pluginClass: FlutterWebRTCPlugin
