@@ -1,31 +1,42 @@
 package com.cloudwebrtc.webrtc;
 
+import static com.cloudwebrtc.webrtc.utils.MediaConstraintsUtils.parseMediaConstraints;
+
 import android.app.Activity;
 import android.content.Context;
+import android.content.pm.PackageManager;
 import android.graphics.SurfaceTexture;
 import android.hardware.Camera;
 import android.hardware.Camera.CameraInfo;
 import android.media.MediaRecorder;
+import android.media.AudioAttributes;
+import android.media.AudioDeviceInfo;
+import android.os.Build;
 import android.util.Log;
 import android.util.LongSparseArray;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 
+import com.cloudwebrtc.webrtc.audio.AudioDeviceKind;
+import com.cloudwebrtc.webrtc.audio.AudioSwitchManager;
+import com.cloudwebrtc.webrtc.audio.AudioUtils;
 import com.cloudwebrtc.webrtc.record.AudioChannel;
 import com.cloudwebrtc.webrtc.record.FrameCapturer;
 import com.cloudwebrtc.webrtc.utils.AnyThreadResult;
+import com.cloudwebrtc.webrtc.utils.Callback;
 import com.cloudwebrtc.webrtc.utils.ConstraintsArray;
 import com.cloudwebrtc.webrtc.utils.ConstraintsMap;
 import com.cloudwebrtc.webrtc.utils.EglUtils;
 import com.cloudwebrtc.webrtc.utils.ObjectType;
-import com.cloudwebrtc.webrtc.SimulcastVideoEncoderFactoryWrapper;
+import com.cloudwebrtc.webrtc.utils.PermissionUtils;
+import com.cloudwebrtc.webrtc.utils.Utils;
+import com.twilio.audioswitch.AudioDevice;
 
 import org.webrtc.AudioSource;
 import org.webrtc.AudioTrack;
 import org.webrtc.CryptoOptions;
-import org.webrtc.DefaultVideoDecoderFactory;
-import org.webrtc.DefaultVideoEncoderFactory;
 import org.webrtc.DtmfSender;
 import org.webrtc.EglBase;
 import org.webrtc.IceCandidate;
@@ -49,6 +60,7 @@ import org.webrtc.PeerConnection.TcpCandidatePolicy;
 import org.webrtc.PeerConnectionFactory;
 import org.webrtc.PeerConnectionFactory.InitializationOptions;
 import org.webrtc.PeerConnectionFactory.Options;
+import org.webrtc.RtpCapabilities;
 import org.webrtc.RtpSender;
 import org.webrtc.SdpObserver;
 import org.webrtc.SessionDescription;
@@ -56,10 +68,12 @@ import org.webrtc.SessionDescription.Type;
 import org.webrtc.VideoTrack;
 import org.webrtc.audio.AudioDeviceModule;
 import org.webrtc.audio.JavaAudioDeviceModule;
+import org.webrtc.video.CustomVideoDecoderFactory;
+import org.webrtc.video.CustomVideoEncoderFactory;
 
 import java.io.File;
-import java.io.UnsupportedEncodingException;
 import java.nio.ByteBuffer;
+import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -75,33 +89,17 @@ import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.view.TextureRegistry;
 import io.flutter.view.TextureRegistry.SurfaceTextureEntry;
 
-import static com.cloudwebrtc.webrtc.utils.MediaConstraintsUtils.parseMediaConstraints;
-
 public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
-  interface AudioManager {
-
-    void onAudioManagerRequested(boolean requested);
-
-    void setMicrophoneMute(boolean mute);
-
-    void setSpeakerphoneOn(boolean on);
-
-
-  }
-
   static public final String TAG = "FlutterWebRTCPlugin";
 
   private final Map<String, PeerConnectionObserver> mPeerConnectionObservers = new HashMap<>();
-  private BinaryMessenger messenger;
-  private Context context;
+  private final BinaryMessenger messenger;
+  private final Context context;
   private final TextureRegistry textures;
-
   private PeerConnectionFactory mFactory;
-
   private final Map<String, MediaStream> localStreams = new HashMap<>();
   private final Map<String, MediaStreamTrack> localTracks = new HashMap<>();
-
-  private LongSparseArray<FlutterRTCVideoRenderer> renders = new LongSparseArray<>();
+  private final LongSparseArray<FlutterRTCVideoRenderer> renders = new LongSparseArray<>();
 
   /**
    * The implementation of {@code getUserMedia} extracted into a separate file in order to reduce
@@ -109,18 +107,20 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
    */
   private GetUserMediaImpl getUserMediaImpl;
 
-  private final AudioManager audioManager;
-
   private AudioDeviceModule audioDeviceModule;
+
+  private FlutterRTCFrameCryptor frameCryptor;
 
   private Activity activity;
 
-  MethodCallHandlerImpl(Context context, BinaryMessenger messenger, TextureRegistry textureRegistry,
-                        @NonNull AudioManager audioManager) {
+  private CustomVideoEncoderFactory videoEncoderFactory;
+
+  private CustomVideoDecoderFactory videoDecoderFactory;
+
+  MethodCallHandlerImpl(Context context, BinaryMessenger messenger, TextureRegistry textureRegistry) {
     this.context = context;
     this.textures = textureRegistry;
     this.messenger = messenger;
-    this.audioManager = audioManager;
   }
 
   static private void resultError(String method, String error, Result result) {
@@ -130,10 +130,22 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
   }
 
   void dispose() {
+    for (final MediaStream mediaStream : localStreams.values()) {
+      streamDispose(mediaStream);
+      mediaStream.dispose();
+    }
+    localStreams.clear();
+    for (final MediaStreamTrack track : localTracks.values()) {
+      track.dispose();
+    }
+    localTracks.clear();
+    for (final PeerConnectionObserver connection : mPeerConnectionObservers.values()) {
+      peerConnectionDispose(connection);
+    }
     mPeerConnectionObservers.clear();
   }
-
-  private void ensureInitialized(boolean bypassVoiceProcessing) {
+  private void initialize(boolean bypassVoiceProcessing, int networkIgnoreMask, boolean forceSWCodec, List<String> forceSWCodecList,
+  @Nullable ConstraintsMap androidAudioConfiguration) {
     if (mFactory != null) {
       return;
     }
@@ -143,53 +155,147 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
                     .setEnableInternalTracer(true)
                     .createInitializationOptions());
 
-    // Initialize EGL contexts required for HW acceleration.
-    EglBase.Context eglContext = EglUtils.getRootEglBaseContext();
-
     getUserMediaImpl = new GetUserMediaImpl(this, context);
 
-    if (bypassVoiceProcessing) {
-        audioDeviceModule =
-                JavaAudioDeviceModule.builder(context)
-                        .setUseHardwareAcousticEchoCanceler(false)
+    frameCryptor = new FlutterRTCFrameCryptor(this);
+
+    AudioAttributes audioAttributes = null;
+    if (androidAudioConfiguration != null) {
+      Integer usageType = AudioUtils.getAudioAttributesUsageTypeForString(
+              androidAudioConfiguration.getString("androidAudioAttributesUsageType"));
+      Integer contentType = AudioUtils.getAudioAttributesContentTypeFromString(
+              androidAudioConfiguration.getString("androidAudioAttributesContentType"));
+
+      // Warn if one is provided without the other.
+      if (usageType == null ^ contentType == null) {
+          Log.w(TAG, "usageType and contentType must both be provided!");
+      }
+
+      if (usageType != null && contentType != null) {
+          audioAttributes = new AudioAttributes.Builder()
+                  .setUsage(usageType)
+                  .setContentType(contentType)
+                  .build();
+      }
+    }
+    JavaAudioDeviceModule.Builder audioDeviceModuleBuilder = JavaAudioDeviceModule.builder(context);
+
+    if(bypassVoiceProcessing) {
+      audioDeviceModuleBuilder.setUseHardwareAcousticEchoCanceler(false)
                         .setUseHardwareNoiseSuppressor(false)
                         .setUseStereoInput(true)
                         .setUseStereoOutput(true)
-                        .setAudioSource(MediaRecorder.AudioSource.MIC)
-                        .setSamplesReadyCallback(getUserMediaImpl.inputSamplesInterceptor)
-                        .createAudioDeviceModule();
+                        .setAudioSource(MediaRecorder.AudioSource.MIC);
     } else {
-        audioDeviceModule =
-                JavaAudioDeviceModule.builder(context)
-                        .setUseHardwareAcousticEchoCanceler(true)
+      audioDeviceModuleBuilder.setUseHardwareAcousticEchoCanceler(true)
                         .setUseHardwareNoiseSuppressor(true)
-                        .setSamplesReadyCallback(getUserMediaImpl.inputSamplesInterceptor)
-                        .createAudioDeviceModule();
+                        .setSamplesReadyCallback(getUserMediaImpl.inputSamplesInterceptor);
     }
+    
+    if (audioAttributes != null) {
+      audioDeviceModuleBuilder.setAudioAttributes(audioAttributes);
+    }
+
+    audioDeviceModule = audioDeviceModuleBuilder.createAudioDeviceModule();
 
     getUserMediaImpl.audioDeviceModule = (JavaAudioDeviceModule) audioDeviceModule;
 
-    mFactory = PeerConnectionFactory.builder()
-            .setOptions(new Options())
-            .setVideoEncoderFactory(new SimulcastVideoEncoderFactoryWrapper(eglContext, true, false))
-            .setVideoDecoderFactory(new DefaultVideoDecoderFactory(eglContext))
+    final Options options = new Options();
+    options.networkIgnoreMask = networkIgnoreMask;
+
+    final PeerConnectionFactory.Builder factoryBuilder = PeerConnectionFactory.builder()
+            .setOptions(options);
+
+    // Initialize EGL contexts required for HW acceleration.
+    EglBase.Context eglContext = EglUtils.getRootEglBaseContext();
+
+    videoEncoderFactory = new CustomVideoEncoderFactory(eglContext, true, true);
+    videoDecoderFactory = new CustomVideoDecoderFactory(eglContext);
+
+    factoryBuilder
+            .setVideoEncoderFactory(videoEncoderFactory)
+            .setVideoDecoderFactory(videoDecoderFactory);
+
+    videoDecoderFactory.setForceSWCodec(forceSWCodec);
+    videoDecoderFactory.setForceSWCodecList(forceSWCodecList);
+    videoEncoderFactory.setForceSWCodec(forceSWCodec);
+    videoEncoderFactory.setForceSWCodecList(forceSWCodecList);
+
+    mFactory = factoryBuilder
             .setAudioDeviceModule(audioDeviceModule)
             .createPeerConnectionFactory();
+
   }
 
   @Override
   public void onMethodCall(MethodCall call, @NonNull Result notSafeResult) {
+
     final AnyThreadResult result = new AnyThreadResult(notSafeResult);
     switch (call.method) {
-      case "initialize":
+      case "initialize": {
+        int networkIgnoreMask = Options.ADAPTER_TYPE_UNKNOWN;
         Map<String, Object> options = call.argument("options");
+        ConstraintsMap constraintsMap = new ConstraintsMap(options);
+        if (constraintsMap.hasKey("networkIgnoreMask")
+                && constraintsMap.getType("networkIgnoreMask") == ObjectType.Array) {
+          final ConstraintsArray ignoredAdapters = constraintsMap.getArray("networkIgnoreMask");
+          if (ignoredAdapters != null) {
+            for (Object adapter : ignoredAdapters.toArrayList()) {
+              switch (adapter.toString()) {
+                case "adapterTypeEthernet":
+                  networkIgnoreMask += Options.ADAPTER_TYPE_ETHERNET;
+                  break;
+                case "adapterTypeWifi":
+                  networkIgnoreMask += Options.ADAPTER_TYPE_WIFI;
+                  break;
+                case "adapterTypeCellular":
+                  networkIgnoreMask += Options.ADAPTER_TYPE_CELLULAR;
+                  break;
+                case "adapterTypeVpn":
+                  networkIgnoreMask += Options.ADAPTER_TYPE_VPN;
+                  break;
+                case "adapterTypeLoopback":
+                  networkIgnoreMask += Options.ADAPTER_TYPE_LOOPBACK;
+                  break;
+                case "adapterTypeAny":
+                  networkIgnoreMask += Options.ADAPTER_TYPE_ANY;
+                  break;
+              }
+            }
+
+          }
+        }
+        boolean forceSWCodec = false;
+        if (constraintsMap.hasKey("forceSWCodec")
+                && constraintsMap.getType("forceSWCodec") == ObjectType.Boolean) {
+          final boolean v = constraintsMap.getBoolean("forceSWCodec");
+          forceSWCodec = v;
+        }
+        List<String> forceSWCodecList = new ArrayList<>();
+        if(constraintsMap.hasKey("forceSWCodecList")
+                && constraintsMap.getType("forceSWCodecList") == ObjectType.Array) {
+          final List<Object> array = constraintsMap.getListArray("forceSWCodecList");
+          for(Object v : array) {
+            forceSWCodecList.add(v.toString());
+          }
+        } else {
+          // disable HW Codec for VP9 by default.
+          forceSWCodecList.add("VP9");
+        }
+
+        ConstraintsMap androidAudioConfiguration = null;
+        if (constraintsMap.hasKey("androidAudioConfiguration")
+                && constraintsMap.getType("androidAudioConfiguration") == ObjectType.Map) {
+            androidAudioConfiguration = constraintsMap.getMap("androidAudioConfiguration");
+        }
         boolean enableBypassVoiceProcessing = false;
         if(options.get("bypassVoiceProcessing") != null) {
           enableBypassVoiceProcessing = (boolean)options.get("bypassVoiceProcessing");
         }
-        ensureInitialized(enableBypassVoiceProcessing);
+        initialize(enableBypassVoiceProcessing, networkIgnoreMask, forceSWCodec, forceSWCodecList, androidAudioConfiguration);
         result.success(null);
-      break;
+        break;
+      }
       case "createPeerConnection": {
         Map<String, Object> constraints = call.argument("constraints");
         Map<String, Object> configuration = call.argument("configuration");
@@ -293,7 +399,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
           RtpSender audioSender = null;
           for (RtpSender sender : peerConnection.getSenders()) {
 
-            if (sender.track().kind().equals("audio")) {
+            if (sender != null && sender.track() != null && sender.track().kind().equals("audio")) {
               audioSender = sender;
             }
           }
@@ -328,20 +434,15 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
       }
       case "dataChannelSend": {
         String peerConnectionId = call.argument("peerConnectionId");
-        int dataChannelId = call.argument("dataChannelId");
+        String dataChannelId = call.argument("dataChannelId");
         String type = call.argument("type");
         Boolean isBinary = type.equals("binary");
         ByteBuffer byteBuffer;
         if (isBinary) {
           byteBuffer = ByteBuffer.wrap(call.argument("data"));
         } else {
-          try {
             String data = call.argument("data");
-            byteBuffer = ByteBuffer.wrap(data.getBytes("UTF-8"));
-          } catch (UnsupportedEncodingException e) {
-            resultError("dataChannelSend", "Could not encode text string as UTF-8.", result);
-            return;
-          }
+            byteBuffer = ByteBuffer.wrap(data.getBytes(StandardCharsets.UTF_8));
         }
         dataChannelSend(peerConnectionId, dataChannelId, byteBuffer, isBinary);
         result.success(null);
@@ -349,21 +450,22 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
       }
       case "dataChannelClose": {
         String peerConnectionId = call.argument("peerConnectionId");
-        int dataChannelId = call.argument("dataChannelId");
+        String dataChannelId = call.argument("dataChannelId");
         dataChannelClose(peerConnectionId, dataChannelId);
         result.success(null);
         break;
       }
       case "streamDispose": {
         String streamId = call.argument("streamId");
-        mediaStreamRelease(streamId);
+        streamDispose(streamId);
         result.success(null);
         break;
       }
       case "mediaStreamTrackSetEnable": {
         String trackId = call.argument("trackId");
         Boolean enabled = call.argument("enabled");
-        mediaStreamTrackSetEnabled(trackId, enabled);
+        String peerConnectionId = call.argument("peerConnectionId");
+        mediaStreamTrackSetEnabled(trackId, enabled, peerConnectionId);
         result.success(null);
         break;
       }
@@ -372,8 +474,8 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         String trackId = call.argument("trackId");
         mediaStreamAddTrack(streamId, trackId, result);
         for (int i = 0; i < renders.size(); i++) {
-          FlutterRTCVideoRenderer renderer = renders.get(i);
-          if (renderer.checkMediaStream(streamId)) {
+          FlutterRTCVideoRenderer renderer = renders.valueAt(i);
+          if (renderer.checkMediaStream(streamId, "local")) {
             renderer.setVideoTrack((VideoTrack) localTracks.get(trackId));
           }
         }
@@ -383,17 +485,12 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         String streamId = call.argument("streamId");
         String trackId = call.argument("trackId");
         mediaStreamRemoveTrack(streamId, trackId, result);
-        for (int i = 0; i < renders.size(); i++) {
-          FlutterRTCVideoRenderer renderer = renders.get(i);
-          if (renderer.checkVideoTrack(trackId)) {
-            renderer.setVideoTrack(null);
-          }
-        }
+        removeStreamForRendererById(streamId);
         break;
       }
       case "trackDispose": {
         String trackId = call.argument("trackId");
-        localTracks.remove(trackId);
+        trackDispose(trackId);
         result.success(null);
         break;
       }
@@ -451,6 +548,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         int textureId = call.argument("textureId");
         String streamId = call.argument("streamId");
         String ownerTag = call.argument("ownerTag");
+        String trackId = call.argument("trackId");
         FlutterRTCVideoRenderer render = renders.get(textureId);
         if (render == null) {
           resultError("videoRendererSetSrcObject", "render [" + textureId + "] not found !", result);
@@ -462,7 +560,11 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         } else {
           stream = getStreamForId(streamId, ownerTag);
         }
-        render.setStream(stream);
+        if (trackId != null && !trackId.equals("0")){
+          render.setStream(stream, trackId, ownerTag);
+        } else {
+          render.setStream(stream, ownerTag);
+        }
         result.success(null);
         break;
       }
@@ -477,6 +579,12 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         getUserMediaImpl.setTorch(trackId, torch, result);
         break;
       }
+      case "mediaStreamTrackSetZoom": {
+        String trackId = call.argument("trackId");
+        double zoomLevel = call.argument("zoomLevel");
+        getUserMediaImpl.setZoom(trackId, zoomLevel, result);
+        break;
+      }
       case "mediaStreamTrackSwitchCamera": {
         String trackId = call.argument("trackId");
         getUserMediaImpl.switchCamera(trackId, result);
@@ -485,20 +593,54 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
       case "setVolume": {
         String trackId = call.argument("trackId");
         double volume = call.argument("volume");
-        mediaStreamTrackSetVolume(trackId, volume);
+        String peerConnectionId = call.argument("peerConnectionId");
+        mediaStreamTrackSetVolume(trackId, volume, peerConnectionId);
         result.success(null);
+        break;
+      }
+      case "selectAudioOutput": {
+        String deviceId = call.argument("deviceId");
+        AudioSwitchManager.instance.selectAudioOutput(AudioDeviceKind.fromTypeName(deviceId));
+        result.success(null);
+        break;
+      }
+      case "clearAndroidCommunicationDevice": {
+        AudioSwitchManager.instance.clearCommunicationDevice();
         break;
       }
       case "setMicrophoneMute":
         boolean mute = call.argument("mute");
-        audioManager.setMicrophoneMute(mute);
+        AudioSwitchManager.instance.setMicrophoneMute(mute);
         result.success(null);
         break;
+      case "selectAudioInput":
+        if (Build.VERSION.SDK_INT > Build.VERSION_CODES.LOLLIPOP_MR1) {
+          String deviceId = call.argument("deviceId");
+          getUserMediaImpl.setPreferredInputDevice(deviceId);
+          result.success(null);
+        } else {
+          result.notImplemented();
+        }
+        break;
+      case "setAndroidAudioConfiguration": {
+        Map<String, Object> configuration = call.argument("configuration");
+        AudioSwitchManager.instance.setAudioConfiguration(configuration);
+        result.success(null);
+        break;
+      }
       case "enableSpeakerphone":
         boolean enable = call.argument("enable");
-        audioManager.setSpeakerphoneOn(enable);
+        AudioSwitchManager.instance.enableSpeakerphone(enable);
         result.success(null);
         break;
+      case "enableSpeakerphoneButPreferBluetooth":
+        AudioSwitchManager.instance.enableSpeakerButPreferBluetooth();
+        result.success(null);
+        break;
+      case "requestCapturePermission": {
+        getUserMediaImpl.requestCapturePermission(result);
+        break;
+      }
       case "getDisplayMedia": {
         Map<String, Object> constraints = call.argument("constraints");
         ConstraintsMap constraintsMap = new ConstraintsMap(constraints);
@@ -512,8 +654,9 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
           String path = call.argument("path");
           VideoTrack videoTrack = null;
           String videoTrackId = call.argument("videoTrackId");
+          String peerConnectionId = call.argument("peerConnectionId");
           if (videoTrackId != null) {
-            MediaStreamTrack track = getTrackForId(videoTrackId);
+            MediaStreamTrack track = getTrackForId(videoTrackId, peerConnectionId);
             if (track instanceof VideoTrack) {
               videoTrack = (VideoTrack) track;
             }
@@ -539,11 +682,12 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         getUserMediaImpl.stopRecording(recorderId);
         result.success(null);
         break;
-      case "captureFrame":
+      case "captureFrame": {
         String path = call.argument("path");
         String videoTrackId = call.argument("trackId");
+        String peerConnectionId = call.argument("peerConnectionId");
         if (videoTrackId != null) {
-          MediaStreamTrack track = getTrackForId(videoTrackId);
+          MediaStreamTrack track = getTrackForId(videoTrackId, peerConnectionId);
           if (track instanceof VideoTrack) {
             new FrameCapturer((VideoTrack) track, new File(path), result);
           } else {
@@ -553,6 +697,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
           resultError("captureFrame", "Track is null", result);
         }
         break;
+      }
       case "getLocalDescription": {
         String peerConnectionId = call.argument("peerConnectionId");
         PeerConnection peerConnection = getPeerConnection(peerConnectionId);
@@ -563,7 +708,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
           params.putString("type", sdp.type.canonicalForm());
           result.success(params.toMap());
         } else {
-          resultError("getLocalDescription", "peerConnection is nulll", result);
+          resultError("getLocalDescription", "peerConnection is null", result);
         }
         break;
       }
@@ -581,7 +726,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
             result.success(params.toMap());
           }
         } else {
-          resultError("getRemoteDescription", "peerConnection is nulll", result);
+          resultError("getRemoteDescription", "peerConnection is null", result);
         }
         break;
       }
@@ -593,7 +738,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
           peerConnectionSetConfiguration(new ConstraintsMap(configuration), peerConnection);
           result.success(null);
         } else {
-          resultError("setConfiguration", "peerConnection is nulll", result);
+          resultError("setConfiguration", "peerConnection is null", result);
         }
         break;
       }
@@ -607,6 +752,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
       case "removeTrack": {
         String peerConnectionId = call.argument("peerConnectionId");
         String senderId = call.argument("senderId");
+
         removeTrack(peerConnectionId, senderId, result);
         break;
       }
@@ -670,10 +816,11 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         rtpSenderSetTrack(peerConnectionId, rtpSenderId, trackId, false, result);
         break;
       }
-      case "rtpSenderDispose": {
+      case "rtpSenderSetStreams": {
         String peerConnectionId = call.argument("peerConnectionId");
         String rtpSenderId = call.argument("rtpSenderId");
-        rtpSenderDispose(peerConnectionId, rtpSenderId, result);
+        List<String> streamIds = call.argument("streamIds");
+        rtpSenderSetStreams(peerConnectionId, rtpSenderId, streamIds, result);
         break;
       }
       case "getSenders": {
@@ -691,10 +838,134 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         getTransceivers(peerConnectionId, result);
         break;
       }
+      case "setPreferredInputDevice": {
+        if (Build.VERSION.SDK_INT > Build.VERSION_CODES.LOLLIPOP_MR1) {
+          String deviceId = call.argument("deviceId");
+          getUserMediaImpl.setPreferredInputDevice(deviceId);
+          result.success(null);
+        } else {
+          result.notImplemented();
+        }
+        break;
+      }
+      case "getRtpSenderCapabilities": {
+        String kind = call.argument("kind");
+        MediaStreamTrack.MediaType mediaType = MediaStreamTrack.MediaType.MEDIA_TYPE_AUDIO;
+        if (kind.equals("video")) {
+          mediaType = MediaStreamTrack.MediaType.MEDIA_TYPE_VIDEO;
+        }
+        RtpCapabilities capabilities = mFactory.getRtpSenderCapabilities(mediaType);
+        result.success(capabilitiestoMap(capabilities).toMap());
+        break;
+      }
+      case "getRtpReceiverCapabilities": {
+        String kind = call.argument("kind");
+        MediaStreamTrack.MediaType mediaType = MediaStreamTrack.MediaType.MEDIA_TYPE_AUDIO;
+        if (kind.equals("video")) {
+          mediaType = MediaStreamTrack.MediaType.MEDIA_TYPE_VIDEO;
+        }
+        RtpCapabilities capabilities = mFactory.getRtpReceiverCapabilities(mediaType);
+        result.success(capabilitiestoMap(capabilities).toMap());
+        break;
+      }
+      case "setCodecPreferences": {
+        String peerConnectionId = call.argument("peerConnectionId");
+        List<Map<String, Object>> codecs = call.argument("codecs");
+        String transceiverId = call.argument("transceiverId");
+        rtpTransceiverSetCodecPreferences(peerConnectionId, transceiverId, codecs, result);
+        break;
+      }
+      case "getSignalingState": {
+        String peerConnectionId = call.argument("peerConnectionId");
+        PeerConnection pc = getPeerConnection(peerConnectionId);
+        if (pc == null) {
+          resultError("getSignalingState", "peerConnection is null", result);
+        } else {
+          ConstraintsMap params = new ConstraintsMap();
+          params.putString("state", Utils.signalingStateString(pc.signalingState()));
+          result.success(params.toMap());
+        }
+        break;
+      }
+      case "getIceGatheringState": {
+        String peerConnectionId = call.argument("peerConnectionId");
+        PeerConnection pc = getPeerConnection(peerConnectionId);
+        if (pc == null) {
+          resultError("getIceGatheringState", "peerConnection is null", result);
+        } else {
+          ConstraintsMap params = new ConstraintsMap();
+          params.putString("state", Utils.iceGatheringStateString(pc.iceGatheringState()));
+          result.success(params.toMap());
+        }
+        break;
+      }
+      case "getIceConnectionState": {
+       String peerConnectionId = call.argument("peerConnectionId");
+       PeerConnection pc = getPeerConnection(peerConnectionId);
+        if (pc == null) {
+          resultError("getIceConnectionState", "peerConnection is null", result);
+        } else {
+          ConstraintsMap params = new ConstraintsMap();
+          params.putString("state", Utils.iceConnectionStateString(pc.iceConnectionState()));
+          result.success(params.toMap());
+        }
+        break;
+      }
+      case "getConnectionState": {
+        String peerConnectionId = call.argument("peerConnectionId");
+        PeerConnection pc = getPeerConnection(peerConnectionId);
+        if (pc == null) {
+          resultError("getConnectionState", "peerConnection is null", result);
+        } else {
+          ConstraintsMap params = new ConstraintsMap();
+          params.putString("state", Utils.connectionStateString(pc.connectionState()));
+          result.success(params.toMap());
+        }
+        break;
+      }
       default:
+        if(frameCryptor.handleMethodCall(call, result)) {
+          break;
+        }
         result.notImplemented();
         break;
     }
+  }
+
+  private ConstraintsMap capabilitiestoMap(RtpCapabilities capabilities) {
+    ConstraintsMap capabilitiesMap = new ConstraintsMap();
+    ConstraintsArray codecArr = new ConstraintsArray();
+    for(RtpCapabilities.CodecCapability codec : capabilities.codecs){
+      ConstraintsMap codecMap = new ConstraintsMap();
+      codecMap.putString("mimeType", codec.mimeType);
+      codecMap.putInt("clockRate", codec.clockRate);
+      if(codec.numChannels != null)
+        codecMap.putInt("channels", codec.numChannels);
+      List<String> sdpFmtpLineArr = new ArrayList<>();
+      for(Map.Entry<String, String> entry : codec.parameters.entrySet()) {
+        if(entry.getKey().length() > 0) {
+          sdpFmtpLineArr.add(entry.getKey() + "=" + entry.getValue());
+        } else {
+          sdpFmtpLineArr.add(entry.getValue());
+        }
+      }
+      if(sdpFmtpLineArr.size() > 0)
+        codecMap.putString("sdpFmtpLine", String.join(";", sdpFmtpLineArr));
+      codecArr.pushMap(codecMap);
+    }
+    ConstraintsArray headerExtensionsArr = new ConstraintsArray();
+    for(RtpCapabilities.HeaderExtensionCapability headerExtension : capabilities.headerExtensions){
+      ConstraintsMap headerExtensionMap = new ConstraintsMap();
+      headerExtensionMap.putString("uri", headerExtension.getUri());
+      headerExtensionMap.putInt("id", headerExtension.getPreferredId());
+      headerExtensionMap.putBoolean("encrypted", headerExtension.getPreferredEncrypted());
+      headerExtensionsArr.pushMap(headerExtensionMap);
+    }
+    capabilitiesMap.putArray("codecs", codecArr.toArrayList());
+    capabilitiesMap.putArray("headerExtensions", headerExtensionsArr.toArrayList());
+    ConstraintsArray fecMechanismsArr = new ConstraintsArray();
+    capabilitiesMap.putArray("fecMechanisms", fecMechanismsArr.toArrayList());
+    return capabilitiesMap;
   }
 
   private PeerConnection getPeerConnection(String id) {
@@ -851,6 +1122,12 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
       }
     }
 
+    // maxIPv6Networks
+    if (map.hasKey("maxIPv6Networks")
+            && map.getType("maxIPv6Networks") == ObjectType.Number) {
+      conf.maxIPv6Networks = map.getInt("maxIPv6Networks");
+    }
+
     // === below is private api in webrtc ===
 
     // tcpCandidatePolicy (private api)
@@ -962,7 +1239,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     }
     // cryptoOptions
     if (map.hasKey("cryptoOptions")
-          && map.getType("cryptoOptions") == ObjectType.Map) {
+            && map.getType("cryptoOptions") == ObjectType.Map) {
       final ConstraintsMap cryptoOptions = map.getMap("cryptoOptions");
       conf.cryptoOptions = CryptoOptions.builder()
               .setEnableGcmCryptoSuites(cryptoOptions.hasKey("enableGcmCryptoSuites") && cryptoOptions.getBoolean("enableGcmCryptoSuites"))
@@ -970,6 +1247,11 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
               .setEnableEncryptedRtpHeaderExtensions(cryptoOptions.hasKey("enableEncryptedRtpHeaderExtensions") && cryptoOptions.getBoolean("enableEncryptedRtpHeaderExtensions"))
               .setEnableAes128Sha1_32CryptoCipher(cryptoOptions.hasKey("enableAes128Sha1_32CryptoCipher") && cryptoOptions.getBoolean("enableAes128Sha1_32CryptoCipher"))
               .createCryptoOptions();
+    }
+    if (map.hasKey("enableCpuOveruseDetection")
+            && map.getType("enableCpuOveruseDetection") == ObjectType.Boolean) {
+      final boolean v = map.getBoolean("enableCpuOveruseDetection");
+      conf.enableCpuOveruseDetection = v;
     }
     return conf;
   }
@@ -984,21 +1266,25 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
             parseMediaConstraints(constraints),
             observer);
     observer.setPeerConnection(peerConnection);
-    if (mPeerConnectionObservers.size() == 0) {
-      audioManager.onAudioManagerRequested(true);
-    }
     mPeerConnectionObservers.put(peerConnectionId, observer);
     return peerConnectionId;
   }
 
   @Override
-  public Map<String, MediaStream> getLocalStreams() {
-    return localStreams;
+  public boolean putLocalStream(String streamId, MediaStream stream) {
+    localStreams.put(streamId, stream);
+    return true;
   }
 
   @Override
-  public Map<String, MediaStreamTrack> getLocalTracks() {
-    return localTracks;
+  public boolean putLocalTrack(String trackId, MediaStreamTrack track) {
+    localTracks.put(trackId, track);
+    return true;
+  }
+
+  @Override
+  public MediaStreamTrack getLocalTrack(String trackId) {
+    return localTracks.get(trackId);
   }
 
   @Override
@@ -1018,7 +1304,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
 
     do {
       uuid = UUID.randomUUID().toString();
-    } while (getTrackForId(uuid) != null);
+    } while (getTrackForId(uuid, null) != null);
 
     return uuid;
   }
@@ -1026,6 +1312,11 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
   @Override
   public PeerConnectionFactory getPeerConnectionFactory() {
     return mFactory;
+  }
+
+  @Override
+  public PeerConnectionObserver getPeerConnectionObserver(String peerConnectionId) {
+    return mPeerConnectionObservers.get(peerConnectionId);
   }
 
   @Nullable
@@ -1038,6 +1329,11 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
   @Override
   public Context getApplicationContext() {
     return context;
+  }
+
+  @Override
+  public BinaryMessenger getMessenger() {
+    return messenger;
   }
 
   MediaStream getStreamForId(String id, String peerConnectionId) {
@@ -1064,11 +1360,14 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     return stream;
   }
 
-  private MediaStreamTrack getTrackForId(String trackId) {
+  private MediaStreamTrack getTrackForId(String trackId, String peerConnectionId) {
     MediaStreamTrack track = localTracks.get(trackId);
 
     if (track == null) {
       for (Entry<String, PeerConnectionObserver> entry : mPeerConnectionObservers.entrySet()) {
+        if (peerConnectionId != null && entry.getKey().compareTo(peerConnectionId) != 0)
+          continue;
+
         PeerConnectionObserver pco = entry.getValue();
         track = pco.remoteTracks.get(trackId);
 
@@ -1129,12 +1428,41 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
       }
     }
 
-    ConstraintsMap audio = new ConstraintsMap();
-    audio.putString("label", "Audio");
-    audio.putString("deviceId", "audio-1");
-    audio.putString("facing", "");
-    audio.putString("kind", "audioinput");
-    array.pushMap(audio);
+    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) {
+      ConstraintsMap audio = new ConstraintsMap();
+      audio.putString("label", "Audio");
+      audio.putString("deviceId", "audio-1");
+      audio.putString("kind", "audioinput");
+      audio.putString("groupId", "microphone");
+      array.pushMap(audio);
+    } else {
+      android.media.AudioManager audioManager = ((android.media.AudioManager) context
+              .getSystemService(Context.AUDIO_SERVICE));
+      final AudioDeviceInfo[] devices = audioManager.getDevices(android.media.AudioManager.GET_DEVICES_INPUTS);
+      for (int i = 0; i < devices.length; i++) {
+        AudioDeviceInfo device = devices[i];
+        if (device.getType() == AudioDeviceInfo.TYPE_BUILTIN_MIC || device.getType() == AudioDeviceInfo.TYPE_BLUETOOTH_SCO ||
+                device.getType() == AudioDeviceInfo.TYPE_WIRED_HEADSET) {
+          ConstraintsMap audio = new ConstraintsMap();
+          audio.putString("label", AudioUtils.getAudioDeviceLabel(device));
+          audio.putString("deviceId", AudioUtils.getAudioDeviceId(device));
+          audio.putString("groupId", AudioUtils.getAudioGroupId(device));
+          audio.putString("kind", "audioinput");
+          array.pushMap(audio);
+        }
+      }
+    }
+
+    List<? extends AudioDevice> audioOutputs = AudioSwitchManager.instance.availableAudioDevices();
+
+    for (AudioDevice audioOutput : audioOutputs) {
+      ConstraintsMap audioOutputMap = new ConstraintsMap();
+      audioOutputMap.putString("label", audioOutput.getName());
+      audioOutputMap.putString("deviceId", AudioDeviceKind.fromAudioDevice(audioOutput).typeName);
+      audioOutputMap.putString("groupId", "" + AudioDeviceKind.fromAudioDevice(audioOutput).typeName);
+      audioOutputMap.putString("kind", "audiooutput");
+      array.pushMap(audioOutputMap);
+    }
 
     ConstraintsMap map = new ConstraintsMap();
     map.putArray("sources", array.toArrayList());
@@ -1156,25 +1484,22 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     result.success(resultMap);
   }
 
-  public void mediaStreamTrackStop(final String id) {
-    // Is this functionality equivalent to `mediaStreamTrackRelease()` ?
-    // if so, we should merge this two and remove track from stream as well.
-    MediaStreamTrack track = localTracks.get(id);
+  public void trackDispose(final String trackId) {
+    MediaStreamTrack track = localTracks.get(trackId);
     if (track == null) {
-      Log.d(TAG, "mediaStreamTrackStop() track is null");
+      Log.d(TAG, "trackDispose() track is null");
       return;
     }
+    removeTrackForRendererById(trackId);
     track.setEnabled(false);
     if (track.kind().equals("video")) {
-      getUserMediaImpl.removeVideoCapturer(id);
+      getUserMediaImpl.removeVideoCapturer(trackId);
     }
-    localTracks.remove(id);
-    // What exactly does `detached` mean in doc?
-    // see: https://www.w3.org/TR/mediacapture-streams/#track-detached
+    localTracks.remove(trackId);
   }
 
-  public void mediaStreamTrackSetEnabled(final String id, final boolean enabled) {
-    MediaStreamTrack track = getTrackForId(id);
+  public void mediaStreamTrackSetEnabled(final String id, final boolean enabled, String peerConnectionId) {
+    MediaStreamTrack track = getTrackForId(id, peerConnectionId);
 
     if (track == null) {
       Log.d(TAG, "mediaStreamTrackSetEnabled() track is null");
@@ -1185,9 +1510,9 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     track.setEnabled(enabled);
   }
 
-  public void mediaStreamTrackSetVolume(final String id, final double volume) {
-    MediaStreamTrack track = localTracks.get(id);
-    if (track != null && track instanceof AudioTrack) {
+  public void mediaStreamTrackSetVolume(final String id, final double volume, String peerConnectionId) {
+    MediaStreamTrack track = getTrackForId(id, null);
+    if (track instanceof AudioTrack) {
       Log.d(TAG, "setVolume(): " + id + "," + volume);
       try {
         ((AudioTrack) track).setVolume(volume);
@@ -1202,12 +1527,17 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
   public void mediaStreamAddTrack(final String streamId, final String trackId, Result result) {
     MediaStream mediaStream = localStreams.get(streamId);
     if (mediaStream != null) {
-      MediaStreamTrack track = getTrackForId(trackId);//localTracks.get(trackId);
+      MediaStreamTrack track = getTrackForId(trackId, null);//localTracks.get(trackId);
       if (track != null) {
-        if (track.kind().equals("audio")) {
+        String kind = track.kind();
+        if (kind.equals("audio")) {
           mediaStream.addTrack((AudioTrack) track);
-        } else if (track.kind().equals("video")) {
+          result.success(null);
+        } else if (kind.equals("video")) {
           mediaStream.addTrack((VideoTrack) track);
+          result.success(null);
+        } else {
+          resultError("mediaStreamAddTrack", "mediaStreamAddTrack() track [" + trackId + "] has unsupported type: " + kind, result);
         }
       } else {
         resultError("mediaStreamAddTrack", "mediaStreamAddTrack() track [" + trackId + "] is null", result);
@@ -1215,7 +1545,6 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     } else {
       resultError("mediaStreamAddTrack", "mediaStreamAddTrack() stream [" + streamId + "] is null", result);
     }
-    result.success(null);
   }
 
   public void mediaStreamRemoveTrack(final String streamId, final String trackId, Result result) {
@@ -1223,18 +1552,22 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     if (mediaStream != null) {
       MediaStreamTrack track = localTracks.get(trackId);
       if (track != null) {
-        if (track.kind().equals("audio")) {
+        String kind = track.kind();
+        if (kind.equals("audio")) {
           mediaStream.removeTrack((AudioTrack) track);
-        } else if (track.kind().equals("video")) {
+          result.success(null);
+        } else if (kind.equals("video")) {
           mediaStream.removeTrack((VideoTrack) track);
+          result.success(null);
+        } else {
+          resultError("mediaStreamRemoveTrack", "mediaStreamRemoveTrack() track [" + trackId + "] has unsupported type: " + kind, result);
         }
       } else {
-        resultError("mediaStreamRemoveTrack", "mediaStreamAddTrack() track [" + trackId + "] is null", result);
+        resultError("mediaStreamRemoveTrack", "mediaStreamRemoveTrack() track [" + trackId + "] is null", result);
       }
     } else {
-      resultError("mediaStreamRemoveTrack", "mediaStreamAddTrack() stream [" + streamId + "] is null", result);
+      resultError("mediaStreamRemoveTrack", "mediaStreamRemoveTrack() stream [" + streamId + "] is null", result);
     }
-    result.success(null);
   }
 
   public void mediaStreamTrackRelease(final String streamId, final String _trackId) {
@@ -1274,6 +1607,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     params.putString("deviceId", "" + index);
     params.putString("facing", facing);
     params.putString("kind", "videoinput");
+    params.putString("groupId", "camera");
     return params;
   }
 
@@ -1465,11 +1799,14 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     boolean res = false;
     PeerConnection peerConnection = getPeerConnection(id);
     if (peerConnection != null) {
+      int sdpMLineIndex = 0;
+      if (!candidateMap.isNull("sdpMLineIndex")) {
+        sdpMLineIndex = candidateMap.getInt("sdpMLineIndex");
+      }
       IceCandidate candidate = new IceCandidate(
-              candidateMap.getString("sdpMid"),
-              candidateMap.getInt("sdpMLineIndex"),
-              candidateMap.getString("candidate")
-      );
+          candidateMap.getString("sdpMid"),
+          sdpMLineIndex,
+          candidateMap.getString("candidate"));
       res = peerConnection.addIceCandidate(candidate);
     } else {
       resultError("peerConnectionAddICECandidate", "peerConnection is null", result);
@@ -1482,7 +1819,11 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     if (pco == null || pco.getPeerConnection() == null) {
       resultError("peerConnectionGetStats", "peerConnection is null", result);
     } else {
-      pco.getStats(trackId, result);
+      if(trackId == null || trackId.isEmpty()) {
+        pco.getStats(result);
+      } else {
+        pco.getStatsForTrack(trackId, result);
+      }
     }
   }
 
@@ -1506,30 +1847,69 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
 
   public void peerConnectionDispose(final String id) {
     PeerConnectionObserver pco = mPeerConnectionObservers.get(id);
-    if (pco == null || pco.getPeerConnection() == null) {
-      Log.d(TAG, "peerConnectionDispose() peerConnection is null");
+    if (pco != null) {
+      if (peerConnectionDispose(pco)) {
+
+        mPeerConnectionObservers.remove(id);
+      }
     } else {
-      pco.dispose();
-      mPeerConnectionObservers.remove(id);
+      Log.d(TAG, "peerConnectionDispose() peerConnectionObserver is null");
     }
     if (mPeerConnectionObservers.size() == 0) {
-      audioManager.onAudioManagerRequested(false);
+      AudioSwitchManager.instance.stop();
     }
   }
 
-  public void mediaStreamRelease(final String id) {
-    MediaStream mediaStream = localStreams.get(id);
-    if (mediaStream != null) {
-      for (VideoTrack track : mediaStream.videoTracks) {
-        localTracks.remove(track.id());
-        getUserMediaImpl.removeVideoCapturer(track.id());
-      }
-      for (AudioTrack track : mediaStream.audioTracks) {
-        localTracks.remove(track.id());
-      }
-      localStreams.remove(id);
+  public boolean peerConnectionDispose(final PeerConnectionObserver pco) {
+    if (pco.getPeerConnection() == null) {
+      Log.d(TAG, "peerConnectionDispose() peerConnection is null");
     } else {
-      Log.d(TAG, "mediaStreamRelease() mediaStream is null");
+      pco.dispose();
+      return true;
+    }
+    return false;
+  }
+
+  public void streamDispose(final String streamId) {
+    MediaStream stream = localStreams.get(streamId);
+    if (stream != null) {
+      streamDispose(stream);
+      localStreams.remove(streamId);
+      removeStreamForRendererById(streamId);
+    } else {
+      Log.d(TAG, "streamDispose() mediaStream is null");
+    }
+  }
+
+  public void streamDispose(final MediaStream stream) {
+    List<VideoTrack> videoTracks = stream.videoTracks;
+    for (VideoTrack track : videoTracks) {
+      localTracks.remove(track.id());
+      getUserMediaImpl.removeVideoCapturer(track.id());
+      stream.removeTrack(track);
+    }
+    List<AudioTrack> audioTracks = stream.audioTracks;
+    for (AudioTrack track : audioTracks) {
+      localTracks.remove(track.id());
+      stream.removeTrack(track);
+    }
+  }
+
+  private void removeStreamForRendererById(String streamId) {
+    for (int i = 0; i < renders.size(); i++) {
+      FlutterRTCVideoRenderer renderer = renders.valueAt(i);
+      if (renderer.checkMediaStream(streamId, "local")) {
+        renderer.setStream(null, "");
+      }
+    }
+  }
+
+  private void removeTrackForRendererById(String trackId) {
+    for (int i = 0; i < renders.size(); i++) {
+      FlutterRTCVideoRenderer renderer = renders.valueAt(i);
+      if (renderer.checkVideoTrack(trackId, "local")) {
+        renderer.setStream(null, null);
+      }
     }
   }
 
@@ -1546,7 +1926,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     }
   }
 
-  public void dataChannelSend(String peerConnectionId, int dataChannelId, ByteBuffer bytebuffer,
+  public void dataChannelSend(String peerConnectionId, String dataChannelId, ByteBuffer bytebuffer,
                               Boolean isBinary) {
     // Forward to PeerConnectionObserver which deals with DataChannels
     // because DataChannel is owned by PeerConnection.
@@ -1559,7 +1939,7 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     }
   }
 
-  public void dataChannelClose(String peerConnectionId, int dataChannelId) {
+  public void dataChannelClose(String peerConnectionId, String dataChannelId) {
     // Forward to PeerConnectionObserver which deals with DataChannels
     // because DataChannel is owned by PeerConnection.
     PeerConnectionObserver pco
@@ -1632,6 +2012,15 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     }
   }
 
+  public void rtpTransceiverSetCodecPreferences(String peerConnectionId, String transceiverId, List<Map<String, Object>> codecs, Result result) {
+    PeerConnectionObserver pco = mPeerConnectionObservers.get(peerConnectionId);
+    if (pco == null || pco.getPeerConnection() == null) {
+      resultError("setCodecPreferences", "peerConnection is null", result);
+    } else {
+      pco.rtpTransceiverSetCodecPreferences(transceiverId, codecs, result);
+    }
+  }
+
   public void rtpTransceiverGetDirection(String peerConnectionId, String transceiverId, Result result) {
     PeerConnectionObserver pco = mPeerConnectionObservers.get(peerConnectionId);
     if (pco == null || pco.getPeerConnection() == null) {
@@ -1668,15 +2057,6 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     }
   }
 
-  public void rtpSenderDispose(String peerConnectionId, String rtpSenderId, Result result) {
-    PeerConnectionObserver pco = mPeerConnectionObservers.get(peerConnectionId);
-    if (pco == null || pco.getPeerConnection() == null) {
-      resultError("rtpSenderDispose", "peerConnection is null", result);
-    } else {
-      pco.rtpSenderDispose(rtpSenderId, result);
-    }
-  }
-
   public void getSenders(String peerConnectionId, Result result) {
     PeerConnectionObserver pco = mPeerConnectionObservers.get(peerConnectionId);
     if (pco == null || pco.getPeerConnection() == null) {
@@ -1709,12 +2089,24 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
     if (pco == null || pco.getPeerConnection() == null) {
       resultError("rtpSenderSetTrack", "peerConnection is null", result);
     } else {
-      MediaStreamTrack track = localTracks.get(trackId);
-      if (track == null) {
-        resultError("rtpSenderSetTrack", "track is null", result);
-        return;
+      MediaStreamTrack track = null;
+      if (trackId.length() > 0) {
+        track = localTracks.get(trackId);
+        if (track == null) {
+          resultError("rtpSenderSetTrack", "track is null", result);
+          return;
+        }
       }
       pco.rtpSenderSetTrack(rtpSenderId, track, result, replace);
+    }
+  }
+
+  public void rtpSenderSetStreams(String peerConnectionId, String rtpSenderId, List<String> streamIds, Result result) {
+    PeerConnectionObserver pco = mPeerConnectionObservers.get(peerConnectionId);
+    if (pco == null || pco.getPeerConnection() == null) {
+      resultError("rtpSenderSetStreams", "peerConnection is null", result);
+    } else {
+      pco.rtpSenderSetStreams(rtpSenderId, streamIds, result);
     }
   }
 
@@ -1732,5 +2124,47 @@ public class MethodCallHandlerImpl implements MethodCallHandler, StateProvider {
         return localTracks.get(id).enabled();
       }
     });
+  }
+
+  @RequiresApi(api = Build.VERSION_CODES.M)
+  void requestPermissions(
+          final ArrayList<String> permissions,
+          final Callback successCallback,
+          final Callback errorCallback) {
+    PermissionUtils.Callback callback =
+            (permissions_, grantResults) -> {
+              List<String> grantedPermissions = new ArrayList<>();
+              List<String> deniedPermissions = new ArrayList<>();
+
+              for (int i = 0; i < permissions_.length; ++i) {
+                String permission = permissions_[i];
+                int grantResult = grantResults[i];
+
+                if (grantResult == PackageManager.PERMISSION_GRANTED) {
+                  grantedPermissions.add(permission);
+                } else {
+                  deniedPermissions.add(permission);
+                }
+              }
+
+              // Success means that all requested permissions were granted.
+              for (String p : permissions) {
+                if (!grantedPermissions.contains(p)) {
+                  // According to step 6 of the getUserMedia() algorithm
+                  // "if the result is denied, jump to the step Permission
+                  // Failure."
+                  errorCallback.invoke(deniedPermissions);
+                  return;
+                }
+              }
+              successCallback.invoke(grantedPermissions);
+            };
+
+    final Activity activity = getActivity();
+    final Context context = getApplicationContext();
+    PermissionUtils.requestPermissions(
+            context,
+            activity,
+            permissions.toArray(new String[permissions.size()]), callback);
   }
 }
