@@ -1,11 +1,10 @@
 #ifndef FLUTTER_WEBRTC_RTC_PEER_CONNECTION_HXX
 #define FLUTTER_WEBRTC_RTC_PEER_CONNECTION_HXX
 
+#include "flutter_common.h"
 #include "flutter_webrtc_base.h"
 
 namespace flutter_webrtc_plugin {
-
-using namespace flutter;
 
 class FlutterPeerConnectionObserver : public RTCPeerConnectionObserver {
  public:
@@ -16,6 +15,7 @@ class FlutterPeerConnectionObserver : public RTCPeerConnectionObserver {
                                 std::string& peerConnectionId);
 
   virtual void OnSignalingState(RTCSignalingState state) override;
+  virtual void OnPeerConnectionState(RTCPeerConnectionState state) override;
   virtual void OnIceGatheringState(RTCIceGatheringState state) override;
   virtual void OnIceConnectionState(RTCIceConnectionState state) override;
   virtual void OnIceCandidate(
@@ -38,8 +38,7 @@ class FlutterPeerConnectionObserver : public RTCPeerConnectionObserver {
   void RemoveStreamForId(const std::string& id);
 
  private:
-  std::unique_ptr<EventChannel<EncodableValue>> event_channel_;
-  std::unique_ptr<EventSink<EncodableValue>> event_sink_;
+  std::unique_ptr<EventChannelProxy> event_channel_;
   scoped_refptr<RTCPeerConnection> peerconnection_;
   std::map<std::string, scoped_refptr<RTCMediaStream>> remote_streams_;
   FlutterWebRTCBase* base_;
@@ -50,46 +49,39 @@ class FlutterPeerConnection {
  public:
   FlutterPeerConnection(FlutterWebRTCBase* base) : base_(base) {}
 
-  void CreateRTCPeerConnection(
-      const EncodableMap& configuration,
-      const EncodableMap& constraints,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+  void CreateRTCPeerConnection(const EncodableMap& configuration,
+                               const EncodableMap& constraints,
+                               std::unique_ptr<MethodResultProxy> result);
 
-  void RTCPeerConnectionClose(
-      RTCPeerConnection* pc,
-      const std::string& uuid,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+  void RTCPeerConnectionClose(RTCPeerConnection* pc,
+                              const std::string& uuid,
+                              std::unique_ptr<MethodResultProxy> result);
 
-  void RTCPeerConnectionDispose(
-      RTCPeerConnection* pc,
-      const std::string& uuid,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+  void RTCPeerConnectionDispose(RTCPeerConnection* pc,
+                                const std::string& uuid,
+                                std::unique_ptr<MethodResultProxy> result);
 
   void CreateOffer(const EncodableMap& constraints,
                    RTCPeerConnection* pc,
-                   std::unique_ptr<MethodResult<EncodableValue>> result);
+                   std::unique_ptr<MethodResultProxy> result);
 
   void CreateAnswer(const EncodableMap& constraints,
                     RTCPeerConnection* pc,
-                    std::unique_ptr<MethodResult<EncodableValue>> result);
+                    std::unique_ptr<MethodResultProxy> result);
 
-  void SetLocalDescription(
-      RTCSessionDescription* sdp,
-      RTCPeerConnection* pc,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+  void SetLocalDescription(RTCSessionDescription* sdp,
+                           RTCPeerConnection* pc,
+                           std::unique_ptr<MethodResultProxy> result);
 
-  void SetRemoteDescription(
-      RTCSessionDescription* sdp,
-      RTCPeerConnection* pc,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+  void SetRemoteDescription(RTCSessionDescription* sdp,
+                            RTCPeerConnection* pc,
+                            std::unique_ptr<MethodResultProxy> result);
 
-  void GetLocalDescription(
-      RTCPeerConnection* pc,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+  void GetLocalDescription(RTCPeerConnection* pc,
+                           std::unique_ptr<MethodResultProxy> result);
 
-  void GetRemoteDescription(
-      RTCPeerConnection* pc,
-      std::unique_ptr<MethodResult<EncodableValue>> resulte);
+  void GetRemoteDescription(RTCPeerConnection* pc,
+                            std::unique_ptr<MethodResultProxy> result);
 
   scoped_refptr<RTCRtpTransceiverInit> mapToRtpTransceiverInit(
       const EncodableMap& transceiverInit);
@@ -104,103 +96,113 @@ class FlutterPeerConnection {
                       const std::string& trackId,
                       const std::string& mediaType,
                       const EncodableMap& transceiverInit,
-                      std::unique_ptr<MethodResult<EncodableValue>> resulte);
+                      std::unique_ptr<MethodResultProxy> result);
 
   void GetTransceivers(RTCPeerConnection* pc,
-                       std::unique_ptr<MethodResult<EncodableValue>> resulte);
+                       std::unique_ptr<MethodResultProxy> result);
 
   void GetReceivers(RTCPeerConnection* pc,
-                    std::unique_ptr<MethodResult<EncodableValue>> resulte);
-
-  void RtpSenderDispose(RTCPeerConnection* pc,
-                        std::string rtpSenderId,
-                        std::unique_ptr<MethodResult<EncodableValue>> resulte);
+                    std::unique_ptr<MethodResultProxy> result);
 
   void RtpSenderSetTrack(RTCPeerConnection* pc,
                          RTCMediaTrack* track,
                          std::string rtpSenderId,
-                         std::unique_ptr<MethodResult<EncodableValue>> resulte);
+                         std::unique_ptr<MethodResultProxy> result);
 
-  void RtpSenderReplaceTrack(
-      RTCPeerConnection* pc,
-      RTCMediaTrack* track,
-      std::string rtpSenderId,
-      std::unique_ptr<MethodResult<EncodableValue>> resulte);
+  void RtpSenderSetStream(RTCPeerConnection* pc,
+                          std::vector<std::string> streamIds,
+                          std::string rtpSenderId,
+                          std::unique_ptr<MethodResultProxy> result);
+
+  void RtpSenderReplaceTrack(RTCPeerConnection* pc,
+                             RTCMediaTrack* track,
+                             std::string rtpSenderId,
+                             std::unique_ptr<MethodResultProxy> result);
 
   scoped_refptr<RTCRtpParameters> updateRtpParameters(
       EncodableMap newParameters,
       scoped_refptr<RTCRtpParameters> parameters);
 
-  void RtpSenderSetParameters(
-      RTCPeerConnection* pc,
-      std::string rtpSenderId,
-      const EncodableMap& parameters,
-      std::unique_ptr<MethodResult<EncodableValue>> resulte);
+  void RtpSenderSetParameters(RTCPeerConnection* pc,
+                              std::string rtpSenderId,
+                              const EncodableMap& parameters,
+                              std::unique_ptr<MethodResultProxy> result);
 
-  void RtpTransceiverStop(
-      RTCPeerConnection* pc,
-      std::string rtpTransceiverId,
-      std::unique_ptr<MethodResult<EncodableValue>> resulte);
+  void RtpTransceiverStop(RTCPeerConnection* pc,
+                          std::string transceiverId,
+                          std::unique_ptr<MethodResultProxy> result);
 
   void RtpTransceiverGetCurrentDirection(
       RTCPeerConnection* pc,
-      std::string rtpTransceiverId,
-      std::unique_ptr<MethodResult<EncodableValue>> resulte);
+      std::string transceiverId,
+      std::unique_ptr<MethodResultProxy> result);
 
   void SetConfiguration(RTCPeerConnection* pc,
                         const EncodableMap& configuration,
-                        std::unique_ptr<MethodResult<EncodableValue>> resulte);
+                        std::unique_ptr<MethodResultProxy> result);
 
   void CaptureFrame(RTCVideoTrack* track,
                     std::string path,
-                    std::unique_ptr<MethodResult<EncodableValue>> resulte);
+                    std::unique_ptr<MethodResultProxy> result);
 
   scoped_refptr<RTCRtpTransceiver> getRtpTransceiverById(RTCPeerConnection* pc,
                                                          std::string id);
 
-  void RtpTransceiverSetDirection(
+  void RtpTransceiverSetDirection(RTCPeerConnection* pc,
+                                  std::string transceiverId,
+                                  std::string direction,
+                                  std::unique_ptr<MethodResultProxy> result);
+
+  void RtpTransceiverSetCodecPreferences(
       RTCPeerConnection* pc,
-      std::string rtpTransceiverId,
-      std::string direction,
-      std::unique_ptr<MethodResult<EncodableValue>> resulte);
+      std::string transceiverId,
+      const EncodableList codecs,
+      std::unique_ptr<MethodResultProxy> result);
 
   void GetSenders(RTCPeerConnection* pc,
-                  std::unique_ptr<MethodResult<EncodableValue>> resulte);
+                  std::unique_ptr<MethodResultProxy> result);
 
   void AddIceCandidate(RTCIceCandidate* candidate,
                        RTCPeerConnection* pc,
-                       std::unique_ptr<MethodResult<EncodableValue>> result);
+                       std::unique_ptr<MethodResultProxy> result);
 
   void GetStats(const std::string& track_id,
                 RTCPeerConnection* pc,
-                std::unique_ptr<MethodResult<EncodableValue>> result);
+                std::unique_ptr<MethodResultProxy> result);
 
-  void MediaStreamAddTrack(
-      scoped_refptr<RTCMediaStream> stream,
-      scoped_refptr<RTCMediaTrack> track,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+  void MediaStreamAddTrack(scoped_refptr<RTCMediaStream> stream,
+                           scoped_refptr<RTCMediaTrack> track,
+                           std::unique_ptr<MethodResultProxy> result);
 
-  void MediaStreamRemoveTrack(
-      scoped_refptr<RTCMediaStream> stream,
-      scoped_refptr<RTCMediaTrack> track,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+  void MediaStreamRemoveTrack(scoped_refptr<RTCMediaStream> stream,
+                              scoped_refptr<RTCMediaTrack> track,
+                              std::unique_ptr<MethodResultProxy> result);
 
   void AddTrack(RTCPeerConnection* pc,
                 scoped_refptr<RTCMediaTrack> track,
-                std::list<std::string> streamIds,
-                std::unique_ptr<MethodResult<EncodableValue>> result);
-
-  libwebrtc::scoped_refptr<libwebrtc::RTCRtpSender> GetRtpSenderById(
-      RTCPeerConnection* pc,
-      std::string id);
+                std::vector<std::string> streamIds,
+                std::unique_ptr<MethodResultProxy> result);
 
   void RemoveTrack(RTCPeerConnection* pc,
                    std::string senderId,
-                   std::unique_ptr<MethodResult<EncodableValue>> result);
+                   std::unique_ptr<MethodResultProxy> result);
 
  private:
   FlutterWebRTCBase* base_;
 };
+
+std::string RTCMediaTypeToString(RTCMediaType type);
+
+std::string transceiverDirectionString(RTCRtpTransceiverDirection direction);
+
+const char* iceConnectionStateString(RTCIceConnectionState state);
+
+const char* signalingStateString(RTCSignalingState state);
+
+const char* peerConnectionStateString(RTCPeerConnectionState state);
+
+const char* iceGatheringStateString(RTCIceGatheringState state);
+
 }  // namespace flutter_webrtc_plugin
 
 #endif  // !FLUTTER_WEBRTC_RTC_PEER_CONNECTION_HXX
