@@ -1,24 +1,32 @@
 #ifndef FLUTTER_WEBRTC_RTC_VIDEO_RENDERER_HXX
 #define FLUTTER_WEBRTC_RTC_VIDEO_RENDERER_HXX
 
-#include <mutex>
-
+#include "flutter_common.h"
 #include "flutter_webrtc_base.h"
+
 #include "rtc_video_frame.h"
 #include "rtc_video_renderer.h"
+
+#include <mutex>
 
 namespace flutter_webrtc_plugin {
 
 using namespace libwebrtc;
-using namespace flutter;
 
-class FlutterVideoRenderer: public RTCVideoRenderer<scoped_refptr<RTCVideoFrame>> {
+class FlutterVideoRenderer
+    : public RTCVideoRenderer<scoped_refptr<RTCVideoFrame>>,
+      public RefCountInterface {
  public:
-  FlutterVideoRenderer(TextureRegistrar *registrar, BinaryMessenger *messenger);
+  FlutterVideoRenderer() = default;
+  ~FlutterVideoRenderer();
 
-  virtual const FlutterDesktopPixelBuffer* CopyPixelBuffer(
-      size_t width,
-      size_t height) const;
+  void initialize(TextureRegistrar* registrar,
+                  BinaryMessenger* messenger,
+                  std::unique_ptr<flutter::TextureVariant> texture,
+                  int64_t texture_id);
+
+  virtual const FlutterDesktopPixelBuffer* CopyPixelBuffer(size_t width,
+                                                           size_t height) const;
 
   virtual void OnFrame(scoped_refptr<RTCVideoFrame> frame) override;
 
@@ -31,6 +39,7 @@ class FlutterVideoRenderer: public RTCVideoRenderer<scoped_refptr<RTCVideoFrame>
   bool CheckVideoTrack(std::string mediaId);
 
   std::string media_stream_id;
+
  private:
   struct FrameSize {
     size_t width;
@@ -38,9 +47,8 @@ class FlutterVideoRenderer: public RTCVideoRenderer<scoped_refptr<RTCVideoFrame>
   };
   FrameSize last_frame_size_ = {0, 0};
   bool first_frame_rendered = false;
-  TextureRegistrar *registrar_ = nullptr;
-  std::unique_ptr<EventChannel<EncodableValue>> event_channel_;
-  std::unique_ptr<EventSink<EncodableValue>> event_sink_;
+  TextureRegistrar* registrar_ = nullptr;
+  std::unique_ptr<EventChannelProxy> event_channel_;
   int64_t texture_id_ = -1;
   scoped_refptr<RTCVideoTrack> track_ = nullptr;
   scoped_refptr<RTCVideoFrame> frame_;
@@ -49,26 +57,25 @@ class FlutterVideoRenderer: public RTCVideoRenderer<scoped_refptr<RTCVideoFrame>
   mutable std::shared_ptr<uint8_t> rgb_buffer_;
   mutable std::mutex mutex_;
   RTCVideoFrame::VideoRotation rotation_ = RTCVideoFrame::kVideoRotation_0;
-
 };
 
 class FlutterVideoRendererManager {
  public:
-  FlutterVideoRendererManager(FlutterWebRTCBase *base);
+  FlutterVideoRendererManager(FlutterWebRTCBase* base);
 
-  void CreateVideoRendererTexture(
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+  void CreateVideoRendererTexture(std::unique_ptr<MethodResultProxy> result);
 
-  void SetMediaStream(int64_t texture_id,
-                      const std::string& stream_id);
+  void VideoRendererSetSrcObject(int64_t texture_id,
+                                 const std::string& stream_id,
+                                 const std::string& owner_tag,
+                                 const std::string& track_id);
 
-  void VideoRendererDispose(
-      int64_t texture_id,
-      std::unique_ptr<MethodResult<EncodableValue>> result);
+  void VideoRendererDispose(int64_t texture_id,
+                            std::unique_ptr<MethodResultProxy> result);
 
  private:
-  FlutterWebRTCBase *base_;
-  std::map<int64_t, std::unique_ptr<FlutterVideoRenderer>> renderers_;
+  FlutterWebRTCBase* base_;
+  std::map<int64_t, scoped_refptr<FlutterVideoRenderer>> renderers_;
 };
 
 }  // namespace flutter_webrtc_plugin
