@@ -2,38 +2,19 @@
 
 namespace flutter_webrtc_plugin {
 
-FlutterVideoRenderer::FlutterVideoRenderer(TextureRegistrar *registrar,
-                                           BinaryMessenger *messenger)
-    : registrar_(registrar) {
-    texture_ =
-      std::make_unique<flutter::TextureVariant>(flutter::PixelBufferTexture(
-          [this](size_t width,
-                 size_t height) -> const FlutterDesktopPixelBuffer* {
-            return this->CopyPixelBuffer(width, height);
-          }));
+FlutterVideoRenderer::~FlutterVideoRenderer() {}
 
-  texture_id_ = registrar_->RegisterTexture(texture_.get());
-
-  std::string event_channel =
+void FlutterVideoRenderer::initialize(
+    TextureRegistrar* registrar,
+    BinaryMessenger* messenger,
+    std::unique_ptr<flutter::TextureVariant> texture,
+    int64_t trxture_id) {
+  registrar_ = registrar;
+  texture_ = std::move(texture);
+  texture_id_ = trxture_id;
+  std::string channel_name =
       "FlutterWebRTC/Texture" + std::to_string(texture_id_);
-  event_channel_.reset(new EventChannel<EncodableValue>(
-      messenger, event_channel, &StandardMethodCodec::GetInstance()));
-
-  auto handler = std::make_unique<StreamHandlerFunctions<EncodableValue>>(
-		  [&](
-			  const flutter::EncodableValue* arguments,
-			  std::unique_ptr<flutter::EventSink<flutter::EncodableValue>>&& events)
-		  -> std::unique_ptr<StreamHandlerError<flutter::EncodableValue>> {
-	  event_sink_ = std::move(events);
-	  return nullptr;
-  },
-		  [&](const flutter::EncodableValue* arguments)
-	  -> std::unique_ptr<StreamHandlerError<flutter::EncodableValue>> {
-	  event_sink_ = nullptr;
-	  return nullptr;
-  });
-
-  event_channel_->SetStreamHandler(std::move(handler));
+  event_channel_ = EventChannelProxy::Create(messenger, channel_name);
 }
 
 const FlutterDesktopPixelBuffer* FlutterVideoRenderer::CopyPixelBuffer(
@@ -43,15 +24,16 @@ const FlutterDesktopPixelBuffer* FlutterVideoRenderer::CopyPixelBuffer(
   if (pixel_buffer_.get() && frame_.get()) {
     if (pixel_buffer_->width != frame_->width() ||
         pixel_buffer_->height != frame_->height()) {
-      size_t buffer_size = (frame_->width() * frame_->height()) * (32 >> 3);
+      size_t buffer_size =
+          (size_t(frame_->width()) * size_t(frame_->height())) * (32 >> 3);
       rgb_buffer_.reset(new uint8_t[buffer_size]);
       pixel_buffer_->width = frame_->width();
       pixel_buffer_->height = frame_->height();
     }
 
     frame_->ConvertToARGB(RTCVideoFrame::Type::kABGR, rgb_buffer_.get(), 0,
-                          (int)pixel_buffer_->width,
-                          (int)pixel_buffer_->height);
+                          static_cast<int>(pixel_buffer_->width),
+                          static_cast<int>(pixel_buffer_->height));
 
     pixel_buffer_->buffer = rgb_buffer_.get();
     mutex_.unlock();
@@ -63,37 +45,33 @@ const FlutterDesktopPixelBuffer* FlutterVideoRenderer::CopyPixelBuffer(
 
 void FlutterVideoRenderer::OnFrame(scoped_refptr<RTCVideoFrame> frame) {
   if (!first_frame_rendered) {
-    if (event_sink_) {
-      EncodableMap params;
-      params[EncodableValue("event")] = "didFirstFrameRendered";
-      params[EncodableValue("id")] = EncodableValue(texture_id_);
-      event_sink_->Success(EncodableValue(params));
-    }
+    EncodableMap params;
+    params[EncodableValue("event")] = "didFirstFrameRendered";
+    params[EncodableValue("id")] = EncodableValue(texture_id_);
+    event_channel_->Success(EncodableValue(params));
     pixel_buffer_.reset(new FlutterDesktopPixelBuffer());
     pixel_buffer_->width = 0;
     pixel_buffer_->height = 0;
     first_frame_rendered = true;
   }
   if (rotation_ != frame->rotation()) {
-    if (event_sink_) {
-      EncodableMap params;
-      params[EncodableValue("event")] = "didTextureChangeRotation";
-      params[EncodableValue("id")] = EncodableValue(texture_id_);
-      params[EncodableValue("rotation")] = EncodableValue((int32_t)frame->rotation());
-      event_sink_->Success(EncodableValue(params));
-    }
+    EncodableMap params;
+    params[EncodableValue("event")] = "didTextureChangeRotation";
+    params[EncodableValue("id")] = EncodableValue(texture_id_);
+    params[EncodableValue("rotation")] =
+        EncodableValue((int32_t)frame->rotation());
+    event_channel_->Success(EncodableValue(params));
     rotation_ = frame->rotation();
   }
   if (last_frame_size_.width != frame->width() ||
       last_frame_size_.height != frame->height()) {
-    if (event_sink_) {
-      EncodableMap params;
-      params[EncodableValue("event")] = "didTextureChangeVideoSize";
-      params[EncodableValue("id")] = EncodableValue(texture_id_);
-      params[EncodableValue("width")] = EncodableValue((int32_t)frame->width());
-      params[EncodableValue("height")] = EncodableValue((int32_t)frame->height());
-      event_sink_->Success(EncodableValue(params));
-    }
+    EncodableMap params;
+    params[EncodableValue("event")] = "didTextureChangeVideoSize";
+    params[EncodableValue("id")] = EncodableValue(texture_id_);
+    params[EncodableValue("width")] = EncodableValue((int32_t)frame->width());
+    params[EncodableValue("height")] = EncodableValue((int32_t)frame->height());
+    event_channel_->Success(EncodableValue(params));
+
     last_frame_size_ = {(size_t)frame->width(), (size_t)frame->height()};
   }
   mutex_.lock();
@@ -104,14 +82,15 @@ void FlutterVideoRenderer::OnFrame(scoped_refptr<RTCVideoFrame> frame) {
 
 void FlutterVideoRenderer::SetVideoTrack(scoped_refptr<RTCVideoTrack> track) {
   if (track_ != track) {
-    if (track_) track_->RemoveRenderer(this);
+    if (track_)
+      track_->RemoveRenderer(this);
     track_ = track;
     last_frame_size_ = {0, 0};
     first_frame_rendered = false;
-    if (track_) track_->AddRenderer(this);
+    if (track_)
+      track_->AddRenderer(this);
   }
 }
-
 
 bool FlutterVideoRenderer::CheckMediaStream(std::string mediaId) {
   if (0 == mediaId.size() || 0 == media_stream_id.size()) {
@@ -128,31 +107,52 @@ bool FlutterVideoRenderer::CheckVideoTrack(std::string mediaId) {
 }
 
 FlutterVideoRendererManager::FlutterVideoRendererManager(
-    FlutterWebRTCBase *base)
+    FlutterWebRTCBase* base)
     : base_(base) {}
 
 void FlutterVideoRendererManager::CreateVideoRendererTexture(
-    std::unique_ptr<MethodResult<EncodableValue>> result) {
-  std::unique_ptr<FlutterVideoRenderer> texture(
-      new FlutterVideoRenderer(base_->textures_, base_->messenger_));
-  int64_t texture_id = texture->texture_id();
-  renderers_[texture_id] = std::move(texture);
+    std::unique_ptr<MethodResultProxy> result) {
+  auto texture = new RefCountedObject<FlutterVideoRenderer>();
+  auto textureVariant =
+      std::make_unique<flutter::TextureVariant>(flutter::PixelBufferTexture(
+          [texture](size_t width,
+                    size_t height) -> const FlutterDesktopPixelBuffer* {
+            return texture->CopyPixelBuffer(width, height);
+          }));
+
+  auto texture_id = base_->textures_->RegisterTexture(textureVariant.get());
+  texture->initialize(base_->textures_, base_->messenger_,
+                      std::move(textureVariant), texture_id);
+  renderers_[texture_id] = texture;
   EncodableMap params;
   params[EncodableValue("textureId")] = EncodableValue(texture_id);
   result->Success(EncodableValue(params));
 }
 
-void FlutterVideoRendererManager::SetMediaStream(int64_t texture_id,
-                                                 const std::string& stream_id) {
+void FlutterVideoRendererManager::VideoRendererSetSrcObject(
+    int64_t texture_id,
+    const std::string& stream_id,
+    const std::string& owner_tag,
+    const std::string& track_id) {
   scoped_refptr<RTCMediaStream> stream =
-      base_->MediaStreamForId(stream_id);
+      base_->MediaStreamForId(stream_id, owner_tag);
+
   auto it = renderers_.find(texture_id);
   if (it != renderers_.end()) {
-    FlutterVideoRenderer *renderer = it->second.get();
+    FlutterVideoRenderer* renderer = it->second.get();
     if (stream.get()) {
       auto video_tracks = stream->video_tracks();
       if (video_tracks.size() > 0) {
-        renderer->SetVideoTrack(video_tracks[0]);
+        if (track_id == std::string()) {
+          renderer->SetVideoTrack(video_tracks[0]);
+        } else {
+          for (auto track : video_tracks.std_vector()) {
+            if (track->id().std_string() == track_id) {
+              renderer->SetVideoTrack(track);
+              break;
+            }
+          }
+        }
         renderer->media_stream_id = stream_id;
       }
     } else {
@@ -162,17 +162,23 @@ void FlutterVideoRendererManager::SetMediaStream(int64_t texture_id,
 }
 
 void FlutterVideoRendererManager::VideoRendererDispose(
-    int64_t texture_id, std::unique_ptr<MethodResult<EncodableValue>> result) {
+    int64_t texture_id,
+    std::unique_ptr<MethodResultProxy> result) {
   auto it = renderers_.find(texture_id);
   if (it != renderers_.end()) {
+    it->second->SetVideoTrack(nullptr);
+#if defined(_WINDOWS)
+    base_->textures_->UnregisterTexture(texture_id,
+                                        [&, it] { renderers_.erase(it); });
+#else
     base_->textures_->UnregisterTexture(texture_id);
     renderers_.erase(it);
+#endif
     result->Success();
     return;
   }
   result->Error("VideoRendererDisposeFailed",
                 "VideoRendererDispose() texture not found!");
 }
-
 
 }  // namespace flutter_webrtc_plugin
