@@ -10,14 +10,15 @@ FlutterWebRTC::FlutterWebRTC(FlutterWebRTCPlugin* plugin)
       FlutterVideoRendererManager::FlutterVideoRendererManager(this),
       FlutterMediaStream::FlutterMediaStream(this),
       FlutterPeerConnection::FlutterPeerConnection(this),
-      FlutterDataChannel::FlutterDataChannel(this) {}
+      FlutterScreenCapture::FlutterScreenCapture(this),
+      FlutterDataChannel::FlutterDataChannel(this),
+      FlutterFrameCryptor::FlutterFrameCryptor(this) {}
 
 FlutterWebRTC::~FlutterWebRTC() {}
 
 void FlutterWebRTC::HandleMethodCall(
-    const flutter::MethodCall<EncodableValue>& method_call,
-    std::unique_ptr<flutter::MethodResult<EncodableValue>> result) {
-
+    const MethodCallProxy& method_call,
+    std::unique_ptr<MethodResultProxy> result) {
   if (method_call.method_name().compare("initialize") == 0) {
     const EncodableMap params =
         GetValue<EncodableMap>(*method_call.arguments());
@@ -50,9 +51,72 @@ void FlutterWebRTC::HandleMethodCall(
     const EncodableMap params =
         GetValue<EncodableMap>(*method_call.arguments());
     const EncodableMap constraints = findMap(params, "constraints");
-    result->NotImplemented();
+
+    GetDisplayMedia(constraints, std::move(result));
+  } else if (method_call.method_name().compare("getDesktopSources") == 0) {
+    // types: ["screen", "window"]
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Bad arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+
+    const EncodableList types = findList(params, "types");
+    if (types.empty()) {
+      result->Error("Bad Arguments", "Types is required");
+      return;
+    }
+    GetDesktopSources(types, std::move(result));
+  } else if (method_call.method_name().compare("updateDesktopSources") == 0) {
+    // types: ["screen", "window"]
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Bad arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+
+    const EncodableList types = findList(params, "types");
+    if (types.empty()) {
+      result->Error("Bad Arguments", "Types is required");
+      return;
+    }
+    UpdateDesktopSources(types, std::move(result));
+  } else if (method_call.method_name().compare("getDesktopSourceThumbnail") ==
+             0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Bad arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+
+    std::string sourceId = findString(params, "sourceId");
+    if (sourceId.empty()) {
+      result->Error("Bad Arguments", "Incorrect sourceId");
+      return;
+    }
+    const EncodableMap thumbnailSize = findMap(params, "thumbnailSize");
+    if (!thumbnailSize.empty()) {
+      int width = 0;
+      int height = 0;
+      GetDesktopSourceThumbnail(sourceId, width, height, std::move(result));
+    } else {
+      result->Error("Bad Arguments", "Bad arguments received");
+    }
   } else if (method_call.method_name().compare("getSources") == 0) {
     GetSources(std::move(result));
+  } else if (method_call.method_name().compare("selectAudioInput") == 0) {
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+    const std::string deviceId = findString(params, "deviceId");
+    SelectAudioInput(deviceId, std::move(result));
+  } else if (method_call.method_name().compare("selectAudioOutput") == 0) {
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+    const std::string deviceId = findString(params, "deviceId");
+    SelectAudioOutput(deviceId, std::move(result));
   } else if (method_call.method_name().compare("mediaStreamGetTracks") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -161,7 +225,11 @@ void FlutterWebRTC::HandleMethodCall(
                                       findString(constraints, "sdp").c_str(),
                                       &error);
 
-    SetLocalDescription(description.get(), pc, std::move(result));
+    if (description.get() != nullptr) {
+      SetLocalDescription(description.get(), pc, std::move(result));
+    } else {
+      result->Error("setLocalDescriptionFailed", "Invalid type or sdp");
+    }
   } else if (method_call.method_name().compare("setRemoteDescription") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -184,7 +252,11 @@ void FlutterWebRTC::HandleMethodCall(
                                       findString(constraints, "sdp").c_str(),
                                       &error);
 
-    SetRemoteDescription(description.get(), pc, std::move(result));
+    if (description.get() != nullptr) {
+      SetRemoteDescription(description.get(), pc, std::move(result));
+    } else {
+      result->Error("setRemoteDescriptionFailed", "Invalid type or sdp");
+    }
   } else if (method_call.method_name().compare("addCandidate") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -202,13 +274,37 @@ void FlutterWebRTC::HandleMethodCall(
     }
 
     SdpParseError error;
-    scoped_refptr<RTCIceCandidate> rtc_candidate =
-        RTCIceCandidate::Create(findString(constraints, "candidate").c_str(),
-                                findString(constraints, "sdpMid").c_str(),
-                                findInt(constraints, "sdpMLineIndex"), &error);
+    std::string candidate = findString(constraints, "candidate");
+    if (candidate.empty()) {
+      // received the end-of-candidates
+      result->Success();
+      return;
+    }
+    int sdpMLineIndex = findInt(constraints, "sdpMLineIndex");
+    scoped_refptr<RTCIceCandidate> rtc_candidate = RTCIceCandidate::Create(
+        candidate.c_str(), findString(constraints, "sdpMid").c_str(),
+        sdpMLineIndex == -1 ? 0 : sdpMLineIndex, &error);
 
-    AddIceCandidate(rtc_candidate.get(), pc, std::move(result));
+    if (rtc_candidate.get() != nullptr) {
+      AddIceCandidate(rtc_candidate.get(), pc, std::move(result));
+    } else {
+      result->Error("addCandidateFailed", "Invalid candidate");
+    }
   } else if (method_call.method_name().compare("getStats") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null constraints arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+    const std::string peerConnectionId = findString(params, "peerConnectionId");
+    const std::string track_id = findString(params, "trackId");
+    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
+    if (pc == nullptr) {
+      result->Error("getStatsFailed", "getStats() peerConnection is null");
+      return;
+    }
+    GetStats(track_id, pc, std::move(result));
   } else if (method_call.method_name().compare("createDataChannel") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -245,10 +341,10 @@ void FlutterWebRTC::HandleMethodCall(
       return;
     }
 
-    int dataChannelId = findInt(params, "dataChannelId");
+    const std::string dataChannelId = findString(params, "dataChannelId");
     const std::string type = findString(params, "type");
     const EncodableValue data = findEncodableValue(params, "data");
-    RTCDataChannel* data_channel = DataChannelFormId(dataChannelId);
+    RTCDataChannel* data_channel = DataChannelForId(dataChannelId);
     if (data_channel == nullptr) {
       result->Error("dataChannelSendFailed",
                     "dataChannelSend() data_channel is null");
@@ -270,14 +366,14 @@ void FlutterWebRTC::HandleMethodCall(
       return;
     }
 
-    int dataChannelId = findInt(params, "dataChannelId");
-    RTCDataChannel* data_channel = DataChannelFormId(dataChannelId);
+    const std::string dataChannelId = findString(params, "dataChannelId");
+    RTCDataChannel* data_channel = DataChannelForId(dataChannelId);
     if (data_channel == nullptr) {
       result->Error("dataChannelCloseFailed",
                     "dataChannelClose() data_channel is null");
       return;
     }
-    DataChannelClose(data_channel, std::move(result));
+    DataChannelClose(data_channel, dataChannelId, std::move(result));
   } else if (method_call.method_name().compare("streamDispose") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -298,11 +394,9 @@ void FlutterWebRTC::HandleMethodCall(
     const std::string track_id = findString(params, "trackId");
     const EncodableValue enable = findEncodableValue(params, "enabled");
     RTCMediaTrack* track = MediaTrackForId(track_id);
-    if (nullptr == track) {
-      result->Error("mediaStreamTrackSetEnableFailed", "mediaStreamTrackSetEnable() track is null");
-      return;
+    if (track != nullptr) {
+      track->set_enabled(GetValue<bool>(enable));
     }
-    track->set_enabled(GetValue<bool>(enable));
     result->Success();
   } else if (method_call.method_name().compare("trackDispose") == 0) {
     if (!method_call.arguments()) {
@@ -313,6 +407,21 @@ void FlutterWebRTC::HandleMethodCall(
         GetValue<EncodableMap>(*method_call.arguments());
     const std::string track_id = findString(params, "trackId");
     MediaStreamTrackDispose(track_id, std::move(result));
+  } else if (method_call.method_name().compare("restartIce") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null constraints arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+    const std::string peerConnectionId = findString(params, "peerConnectionId");
+    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
+    if (pc == nullptr) {
+      result->Error("restartIceFailed", "restartIce() peerConnection is null");
+      return;
+    }
+    pc->RestartIce();
+    result->Success();
   } else if (method_call.method_name().compare("peerConnectionClose") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -338,8 +447,7 @@ void FlutterWebRTC::HandleMethodCall(
     const std::string peerConnectionId = findString(params, "peerConnectionId");
     RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
     if (pc == nullptr) {
-      result->Error("peerConnectionDisposeFailed",
-                    "peerConnectionDisposeClose() peerConnection is null");
+      result->Success();
       return;
     }
     RTCPeerConnectionDispose(pc, peerConnectionId, std::move(result));
@@ -364,8 +472,10 @@ void FlutterWebRTC::HandleMethodCall(
         GetValue<EncodableMap>(*method_call.arguments());
     const std::string stream_id = findString(params, "streamId");
     int64_t texture_id = findLongInt(params, "textureId");
+    const std::string owner_tag = findString(params, "ownerTag");
+    const std::string track_id = findString(params, "trackId");
 
-    SetMediaStream(texture_id, stream_id);
+    VideoRendererSetSrcObject(texture_id, stream_id, owner_tag, track_id);
     result->Success();
   } else if (method_call.method_name().compare(
                  "mediaStreamTrackSwitchCamera") == 0) {
@@ -378,6 +488,48 @@ void FlutterWebRTC::HandleMethodCall(
     const std::string track_id = findString(params, "trackId");
     MediaStreamTrackSwitchCamera(track_id, std::move(result));
   } else if (method_call.method_name().compare("setVolume") == 0) {
+    auto args = method_call.arguments();
+    if (!args) {
+      result->Error("Bad Arguments", "setVolume() Null arguments received");
+      return;
+    }
+
+    const EncodableMap params = GetValue<EncodableMap>(*args);
+    const std::string trackId = findString(params, "trackId");
+    const std::optional<double> volume = maybeFindDouble(params, "volume");
+
+    if (trackId.empty()) {
+      result->Error("Bad Arguments", "setVolume() Empty track provided");
+      return;
+    }
+
+    if (!volume.has_value()) {
+      result->Error("Bad Arguments", "setVolume() No volume provided");
+      return;
+    }
+
+    if (volume.value() < 0) {
+      result->Error("Bad Arguments", "setVolume() Volume must be positive");
+      return;
+    }
+
+    RTCMediaTrack* track = MediaTrackForId(trackId);
+    if (nullptr == track) {
+      result->Error("setVolume", "setVolume() Unable to find provided track");
+      return;
+    }
+
+    std::string kind = track->kind().std_string();
+    if (0 != kind.compare("audio")) {
+      result->Error("setVolume",
+                    "setVolume() Only audio tracks can have volume set");
+      return;
+    }
+
+    auto audioTrack = static_cast<RTCAudioTrack*>(track);
+    audioTrack->SetVolume(volume.value());
+
+    result->Success();
   } else if (method_call.method_name().compare("getLocalDescription") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -444,7 +596,7 @@ void FlutterWebRTC::HandleMethodCall(
       if (renderer->CheckMediaStream(streamId) && 0 == kind.compare("video")) {
         renderer->SetVideoTrack(static_cast<RTCVideoTrack*>(track.get()));
       }
-    };
+    }
   } else if (method_call.method_name().compare("mediaStreamRemoveTrack") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -477,7 +629,7 @@ void FlutterWebRTC::HandleMethodCall(
       if (renderer->CheckVideoTrack(streamId)) {
         renderer->SetVideoTrack(nullptr);
       }
-    };
+    }
   } else if (method_call.method_name().compare("addTrack") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -500,13 +652,13 @@ void FlutterWebRTC::HandleMethodCall(
     if (track == nullptr) {
       result->Error("AddTrack", "AddTrack() track is null");
       return;
-    };
-    std::list<std::string> listId;
+    }
+    std::vector<std::string> ids;
     for (EncodableValue value : streamIds) {
-      listId.push_back(GetValue<std::string>(value));
+      ids.push_back(GetValue<std::string>(value));
     }
 
-    AddTrack(pc, track, listId, std::move(result));
+    AddTrack(pc, track, ids, std::move(result));
 
   } else if (method_call.method_name().compare("removeTrack") == 0) {
     if (!method_call.arguments()) {
@@ -596,32 +748,6 @@ void FlutterWebRTC::HandleMethodCall(
     }
 
     GetSenders(pc, std::move(result));
-  } else if (method_call.method_name().compare("rtpSenderDispose") == 0) {
-    if (!method_call.arguments()) {
-      result->Error("Bad Arguments", "Null constraints arguments received");
-      return;
-    }
-    const EncodableMap params =
-        GetValue<EncodableMap>(*method_call.arguments());
-    const std::string peerConnectionId = findString(params, "peerConnectionId");
-
-    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
-    if (pc == nullptr) {
-      result->Error("rtpSenderDispose",
-                    "rtpSenderDispose() peerConnection is null");
-      return;
-    }
-
-    const std::string rtpSenderId = findString(params, "rtpSenderId");
-    if (0 < rtpSenderId.size()) {
-      if (pc == nullptr) {
-        result->Error("rtpSenderDispose",
-                      "rtpSenderDispose() rtpSenderId is null or empty");
-        return;
-      }
-    }
-    RtpSenderDispose(pc, rtpSenderId, std::move(result));
-
   } else if (method_call.method_name().compare("rtpSenderSetTrack") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -640,20 +766,48 @@ void FlutterWebRTC::HandleMethodCall(
 
     const std::string trackId = findString(params, "trackId");
     RTCMediaTrack* track = MediaTrackForId(trackId);
-    if (nullptr == track) {
-      result->Error("rtpSenderSetTrack", "rtpSenderSetTrack() track is null");
+
+    const std::string rtpSenderId = findString(params, "rtpSenderId");
+    if (rtpSenderId.empty()) {
+      result->Error("rtpSenderSetTrack",
+                    "rtpSenderSetTrack() rtpSenderId is null or empty");
+      return;
+    }
+    RtpSenderSetTrack(pc, track, rtpSenderId, std::move(result));
+  } else if (method_call.method_name().compare("rtpSenderSetStreams") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null constraints arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+    const std::string peerConnectionId = findString(params, "peerConnectionId");
+
+    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
+    if (pc == nullptr) {
+      result->Error("rtpSenderSetStream",
+                    "rtpSenderSetStream() peerConnection is null");
       return;
     }
 
-    const std::string rtpSenderId = findString(params, "rtpSenderId");
-    if (0 < rtpSenderId.size()) {
-      if (pc == nullptr) {
-        result->Error("rtpSenderSetTrack",
-                      "rtpSenderSetTrack() rtpSenderId is null or empty");
-        return;
-      }
+    const EncodableList encodableStreamIds = findList(params, "streamIds");
+    if (encodableStreamIds.empty()) {
+      result->Error("rtpSenderSetStream",
+                    "rtpSenderSetStream() streamId is null or empty");
+      return;
     }
-    RtpSenderSetTrack(pc, track, rtpSenderId, std::move(result));
+    std::vector<std::string> streamIds{};
+    for (EncodableValue value : encodableStreamIds) {
+      streamIds.push_back(GetValue<std::string>(value));
+    }
+
+    const std::string rtpSenderId = findString(params, "rtpSenderId");
+    if (rtpSenderId.empty()) {
+      result->Error("rtpSenderSetStream",
+                    "rtpSenderSetStream() rtpSenderId is null or empty");
+      return;
+    }
+    RtpSenderSetStream(pc, streamIds, rtpSenderId, std::move(result));
   } else if (method_call.method_name().compare("rtpSenderReplaceTrack") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -672,19 +826,12 @@ void FlutterWebRTC::HandleMethodCall(
 
     const std::string trackId = findString(params, "trackId");
     RTCMediaTrack* track = MediaTrackForId(trackId);
-    if (nullptr == track) {
-      result->Error("rtpSenderReplaceTrack",
-                    "rtpSenderReplaceTrack() track is null");
-      return;
-    }
 
     const std::string rtpSenderId = findString(params, "rtpSenderId");
-    if (0 < rtpSenderId.size()) {
-      if (pc == nullptr) {
-        result->Error("rtpSenderReplaceTrack",
-                      "rtpSenderReplaceTrack() rtpSenderId is null or empty");
-        return;
-      }
+    if (rtpSenderId.empty()) {
+      result->Error("rtpSenderReplaceTrack",
+                    "rtpSenderReplaceTrack() rtpSenderId is null or empty");
+      return;
     }
     RtpSenderReplaceTrack(pc, track, rtpSenderId, std::move(result));
   } else if (method_call.method_name().compare("rtpSenderSetParameters") == 0) {
@@ -704,16 +851,14 @@ void FlutterWebRTC::HandleMethodCall(
     }
 
     const std::string rtpSenderId = findString(params, "rtpSenderId");
-    if (0 < rtpSenderId.size()) {
-      if (pc == nullptr) {
-        result->Error("rtpSenderSetParameters",
-                      "rtpSenderSetParameters() rtpSenderId is null or empty");
-        return;
-      }
+    if (rtpSenderId.empty()) {
+      result->Error("rtpSenderSetParameters",
+                    "rtpSenderSetParameters() rtpSenderId is null or empty");
+      return;
     }
 
     const EncodableMap parameters = findMap(params, "parameters");
-    if (0 < parameters.size()) {
+    if (0 == parameters.size()) {
       result->Error("rtpSenderSetParameters",
                     "rtpSenderSetParameters() parameters is null or empty");
       return;
@@ -736,16 +881,14 @@ void FlutterWebRTC::HandleMethodCall(
       return;
     }
 
-    const std::string rtpTransceiverId = findString(params, "rtpTransceiverId");
-    if (0 < rtpTransceiverId.size()) {
-      if (pc == nullptr) {
-        result->Error("rtpTransceiverStop",
-                      "rtpTransceiverStop() rtpTransceiverId is null or empty");
-        return;
-      }
+    const std::string transceiverId = findString(params, "transceiverId");
+    if (transceiverId.empty()) {
+      result->Error("rtpTransceiverStop",
+                    "rtpTransceiverStop() transceiverId is null or empty");
+      return;
     }
 
-    RtpTransceiverStop(pc, rtpTransceiverId, std::move(result));
+    RtpTransceiverStop(pc, transceiverId, std::move(result));
   } else if (method_call.method_name().compare(
                  "rtpTransceiverGetCurrentDirection") == 0) {
     if (!method_call.arguments()) {
@@ -764,17 +907,15 @@ void FlutterWebRTC::HandleMethodCall(
       return;
     }
 
-    const std::string rtpTransceiverId = findString(params, "rtpTransceiverId");
-    if (0 < rtpTransceiverId.size()) {
-      if (pc == nullptr) {
-        result->Error("rtpTransceiverGetCurrentDirection",
-                      "rtpTransceiverGetCurrentDirection() rtpTransceiverId is "
-                      "null or empty");
-        return;
-      }
+    const std::string transceiverId = findString(params, "transceiverId");
+    if (transceiverId.empty()) {
+      result->Error("rtpTransceiverGetCurrentDirection",
+                    "rtpTransceiverGetCurrentDirection() transceiverId is "
+                    "null or empty");
+      return;
     }
 
-    RtpTransceiverGetCurrentDirection(pc, rtpTransceiverId, std::move(result));
+    RtpTransceiverGetCurrentDirection(pc, transceiverId, std::move(result));
   } else if (method_call.method_name().compare("rtpTransceiverSetDirection") ==
              0) {
     if (!method_call.arguments()) {
@@ -792,28 +933,22 @@ void FlutterWebRTC::HandleMethodCall(
       return;
     }
 
-    const std::string rtpTransceiverId = findString(params, "rtpTransceiverId");
-    if (0 < rtpTransceiverId.size()) {
-      if (pc == nullptr) {
-        result->Error("rtpTransceiverGetCurrentDirection",
-                      "rtpTransceiverGetCurrentDirection() rtpTransceiverId is "
-                      "null or empty");
-        return;
-      }
+    const std::string transceiverId = findString(params, "transceiverId");
+    if (transceiverId.empty()) {
+      result->Error("rtpTransceiverSetDirection",
+                    "rtpTransceiverSetDirection() transceiverId is "
+                    "null or empty");
+      return;
     }
 
     const std::string direction = findString(params, "direction");
-    if (0 < rtpTransceiverId.size()) {
-      if (pc == nullptr) {
-        result->Error(
-            "rtpTransceiverGetCurrentDirection",
-            "rtpTransceiverGetCurrentDirection() direction is null or empty");
-        return;
-      }
+    if (transceiverId.empty()) {
+      result->Error("rtpTransceiverSetDirection",
+                    "rtpTransceiverSetDirection() direction is null or empty");
+      return;
     }
 
-    RtpTransceiverSetDirection(pc, rtpTransceiverId, direction,
-                               std::move(result));
+    RtpTransceiverSetDirection(pc, transceiverId, direction, std::move(result));
   } else if (method_call.method_name().compare("setConfiguration") == 0) {
     if (!method_call.arguments()) {
       result->Error("Bad Arguments", "Null constraints arguments received");
@@ -831,12 +966,10 @@ void FlutterWebRTC::HandleMethodCall(
     }
 
     const EncodableMap configuration = findMap(params, "configuration");
-    if (0 < configuration.size()) {
-      if (pc == nullptr) {
-        result->Error("setConfiguration",
-                      "setConfiguration() configuration is null or empty");
-        return;
-      }
+    if (configuration.empty()) {
+      result->Error("setConfiguration",
+                    "setConfiguration() configuration is null or empty");
+      return;
     }
     SetConfiguration(pc, configuration, std::move(result));
   } else if (method_call.method_name().compare("captureFrame") == 0) {
@@ -848,7 +981,7 @@ void FlutterWebRTC::HandleMethodCall(
         GetValue<EncodableMap>(*method_call.arguments());
 
     const std::string path = findString(params, "path");
-    if (0 < path.size()) {
+    if (path.empty()) {
       result->Error("captureFrame", "captureFrame() path is null or empty");
       return;
     }
@@ -864,10 +997,256 @@ void FlutterWebRTC::HandleMethodCall(
       result->Error("captureFrame", "captureFrame() track not is video track");
       return;
     }
-    CaptureFrame((RTCVideoTrack*)track, path, std::move(result));
+    CaptureFrame(reinterpret_cast<RTCVideoTrack*>(track), path,
+                 std::move(result));
 
+  } else if (method_call.method_name().compare("createLocalMediaStream") == 0) {
+    CreateLocalMediaStream(std::move(result));
+  } else if (method_call.method_name().compare("canInsertDtmf") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null constraints arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+    const std::string peerConnectionId = findString(params, "peerConnectionId");
+    const std::string rtpSenderId = findString(params, "rtpSenderId");
+
+    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
+    if (pc == nullptr) {
+      result->Error("canInsertDtmf", "canInsertDtmf() peerConnection is null");
+      return;
+    }
+
+    auto rtpSender = GetRtpSenderById(pc, rtpSenderId);
+
+    if (rtpSender == nullptr) {
+      result->Error("sendDtmf", "sendDtmf() rtpSender is null");
+      return;
+    }
+    auto dtmfSender = rtpSender->dtmf_sender();
+    bool canInsertDtmf = dtmfSender->CanInsertDtmf();
+
+    result->Success(EncodableValue(canInsertDtmf));
+  } else if (method_call.method_name().compare("sendDtmf") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null constraints arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+    const std::string peerConnectionId = findString(params, "peerConnectionId");
+    const std::string rtpSenderId = findString(params, "rtpSenderId");
+    const std::string tone = findString(params, "tone");
+    int duration = findInt(params, "duration");
+    int gap = findInt(params, "gap");
+
+    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
+    if (pc == nullptr) {
+      result->Error("sendDtmf", "sendDtmf() peerConnection is null");
+      return;
+    }
+
+    auto rtpSender = GetRtpSenderById(pc, rtpSenderId);
+
+    if (rtpSender == nullptr) {
+      result->Error("sendDtmf", "sendDtmf() rtpSender is null");
+      return;
+    }
+
+    auto dtmfSender = rtpSender->dtmf_sender();
+    dtmfSender->InsertDtmf(tone, duration, gap);
+
+    result->Success();
+  } else if (method_call.method_name().compare("getRtpSenderCapabilities") ==
+             0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+
+    RTCMediaType mediaType = RTCMediaType::AUDIO;
+    const std::string kind = findString(params, "kind");
+    if (0 == kind.compare("video")) {
+      mediaType = RTCMediaType::VIDEO;
+    } else if (0 == kind.compare("audio")) {
+      mediaType = RTCMediaType::AUDIO;
+    } else {
+      result->Error("getRtpSenderCapabilities",
+                    "getRtpSenderCapabilities() kind is null or empty");
+      return;
+    }
+    auto capabilities = factory_->GetRtpSenderCapabilities(mediaType);
+    EncodableMap map;
+    EncodableList codecsList;
+    for (auto codec : capabilities->codecs().std_vector()) {
+      EncodableMap codecMap;
+      codecMap[EncodableValue("mimeType")] =
+          EncodableValue(codec->mime_type().std_string());
+      codecMap[EncodableValue("clockRate")] =
+          EncodableValue(codec->clock_rate());
+      codecMap[EncodableValue("channels")] = EncodableValue(codec->channels());
+      codecMap[EncodableValue("sdpFmtpLine")] =
+          EncodableValue(codec->sdp_fmtp_line().std_string());
+      codecsList.push_back(EncodableValue(codecMap));
+    }
+    map[EncodableValue("codecs")] = EncodableValue(codecsList);
+    map[EncodableValue("headerExtensions")] = EncodableValue(EncodableList());
+    map[EncodableValue("fecMechanisms")] = EncodableValue(EncodableList());
+
+    result->Success(EncodableValue(map));
+  } else if (method_call.method_name().compare("getRtpReceiverCapabilities") ==
+             0) {
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+
+    RTCMediaType mediaType = RTCMediaType::AUDIO;
+    const std::string kind = findString(params, "kind");
+    if (0 == kind.compare("video")) {
+      mediaType = RTCMediaType::VIDEO;
+    } else if (0 == kind.compare("audio")) {
+      mediaType = RTCMediaType::AUDIO;
+    } else {
+      result->Error("getRtpSenderCapabilities",
+                    "getRtpSenderCapabilities() kind is null or empty");
+      return;
+    }
+    auto capabilities = factory_->GetRtpReceiverCapabilities(mediaType);
+    EncodableMap map;
+    EncodableList codecsList;
+    for (auto codec : capabilities->codecs().std_vector()) {
+      EncodableMap codecMap;
+      codecMap[EncodableValue("mimeType")] =
+          EncodableValue(codec->mime_type().std_string());
+      codecMap[EncodableValue("clockRate")] =
+          EncodableValue(codec->clock_rate());
+      codecMap[EncodableValue("channels")] = EncodableValue(codec->channels());
+      codecMap[EncodableValue("sdpFmtpLine")] =
+          EncodableValue(codec->sdp_fmtp_line().std_string());
+      codecsList.push_back(EncodableValue(codecMap));
+    }
+    map[EncodableValue("codecs")] = EncodableValue(codecsList);
+    map[EncodableValue("headerExtensions")] = EncodableValue(EncodableList());
+    map[EncodableValue("fecMechanisms")] = EncodableValue(EncodableList());
+
+    result->Success(EncodableValue(map));
+  } else if (method_call.method_name().compare("setCodecPreferences") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+    const std::string peerConnectionId = findString(params, "peerConnectionId");
+    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
+    if (pc == nullptr) {
+      result->Error("setCodecPreferences",
+                    "setCodecPreferences() peerConnection is null");
+      return;
+    }
+
+    const std::string transceiverId = findString(params, "transceiverId");
+    if (transceiverId.empty()) {
+      result->Error("setCodecPreferences",
+                    "setCodecPreferences() transceiverId is null or empty");
+      return;
+    }
+
+    const EncodableList codecs = findList(params, "codecs");
+    if (codecs.empty()) {
+      result->Error("Bad Arguments", "Codecs is required");
+      return;
+    }
+    RtpTransceiverSetCodecPreferences(pc, transceiverId, codecs,
+                                      std::move(result));
+  } else if (method_call.method_name().compare("getSignalingState") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null constraints arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+
+    const std::string peerConnectionId = findString(params, "peerConnectionId");
+
+    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
+    if (pc == nullptr) {
+      result->Error("getSignalingState",
+                    "getSignalingState() peerConnection is null");
+      return;
+    }
+    EncodableMap state;
+    state[EncodableValue("state")] =
+        signalingStateString(pc->signaling_state());
+    result->Success(EncodableValue(state));
+  } else if (method_call.method_name().compare("getIceGatheringState") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null constraints arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+
+    const std::string peerConnectionId = findString(params, "peerConnectionId");
+
+    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
+    if (pc == nullptr) {
+      result->Error("getIceGatheringState",
+                    "getIceGatheringState() peerConnection is null");
+      return;
+    }
+    EncodableMap state;
+    state[EncodableValue("state")] =
+        iceGatheringStateString(pc->ice_gathering_state());
+    result->Success(EncodableValue(state));
+  } else if (method_call.method_name().compare("getIceConnectionState") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null constraints arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+
+    const std::string peerConnectionId = findString(params, "peerConnectionId");
+
+    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
+    if (pc == nullptr) {
+      result->Error("getIceConnectionState",
+                    "getIceConnectionState() peerConnection is null");
+      return;
+    }
+    EncodableMap state;
+    state[EncodableValue("state")] =
+        iceConnectionStateString(pc->ice_connection_state());
+    result->Success(EncodableValue(state));
+  } else if (method_call.method_name().compare("getConnectionState") == 0) {
+    if (!method_call.arguments()) {
+      result->Error("Bad Arguments", "Null constraints arguments received");
+      return;
+    }
+    const EncodableMap params =
+        GetValue<EncodableMap>(*method_call.arguments());
+
+    const std::string peerConnectionId = findString(params, "peerConnectionId");
+
+    RTCPeerConnection* pc = PeerConnectionForId(peerConnectionId);
+    if (pc == nullptr) {
+      result->Error("getConnectionState",
+                    "getConnectionState() peerConnection is null");
+      return;
+    }
+    EncodableMap state;
+    state[EncodableValue("state")] =
+        peerConnectionStateString(pc->peer_connection_state());
+    result->Success(EncodableValue(state));
   } else {
-    result->NotImplemented();
+    if (HandleFrameCryptorMethodCall(method_call, std::move(result), &result)) {
+      return;
+    } else {
+      result->NotImplemented();
+    }
   }
 }
 
