@@ -5,35 +5,45 @@
 
 namespace flutter_webrtc_plugin {
 
-FlutterWebRTCBase::FlutterWebRTCBase(BinaryMessenger *messenger,
-                                     TextureRegistrar *textures)
+const char* kEventChannelName = "FlutterWebRTC.Event";
+
+FlutterWebRTCBase::FlutterWebRTCBase(BinaryMessenger* messenger,
+                                     TextureRegistrar* textures)
     : messenger_(messenger), textures_(textures) {
   LibWebRTC::Initialize();
   factory_ = LibWebRTC::CreateRTCPeerConnectionFactory();
   audio_device_ = factory_->GetAudioDevice();
   video_device_ = factory_->GetVideoDevice();
+  desktop_device_ = factory_->GetDesktopDevice();
+  event_channel_ = EventChannelProxy::Create(messenger_, kEventChannelName);
 }
 
 FlutterWebRTCBase::~FlutterWebRTCBase() {
   LibWebRTC::Terminate();
 }
 
+EventChannelProxy* FlutterWebRTCBase::event_channel() {
+  return event_channel_ ? event_channel_.get() : nullptr;
+}
+
 std::string FlutterWebRTCBase::GenerateUUID() {
   return uuidxx::uuid::Generate().ToString(false);
 }
 
-RTCPeerConnection *FlutterWebRTCBase::PeerConnectionForId(
-    const std::string &id) {
+RTCPeerConnection* FlutterWebRTCBase::PeerConnectionForId(
+    const std::string& id) {
   auto it = peerconnections_.find(id);
 
-  if (it != peerconnections_.end()) return (*it).second.get();
+  if (it != peerconnections_.end())
+    return (*it).second.get();
 
   return nullptr;
 }
 
-void FlutterWebRTCBase::RemovePeerConnectionForId(const std::string &id) {
+void FlutterWebRTCBase::RemovePeerConnectionForId(const std::string& id) {
   auto it = peerconnections_.find(id);
-  if (it != peerconnections_.end()) peerconnections_.erase(it);
+  if (it != peerconnections_.end())
+    peerconnections_.erase(it);
 }
 
 RTCMediaTrack* FlutterWebRTCBase ::MediaTrackForId(const std::string& id) {
@@ -43,9 +53,10 @@ RTCMediaTrack* FlutterWebRTCBase ::MediaTrackForId(const std::string& id) {
     return (*it).second.get();
 
   for (auto kv : peerconnection_observers_) {
-      auto pco = kv.second.get();
-      auto track = pco->MediaTrackForId(id);
-      if (track != nullptr) return track;
+    auto pco = kv.second.get();
+    auto track = pco->MediaTrackForId(id);
+    if (track != nullptr)
+      return track;
   }
 
   return nullptr;
@@ -75,34 +86,46 @@ void FlutterWebRTCBase::RemovePeerConnectionObserversForId(
 }
 
 scoped_refptr<RTCMediaStream> FlutterWebRTCBase::MediaStreamForId(
-    const std::string& id) {
+    const std::string& id, std::string ownerTag) {
+  if (!ownerTag.empty()) {
+    if (ownerTag == "local") {
+      auto it = local_streams_.find(id);
+      if (it != local_streams_.end()) {
+        return (*it).second;
+      }
+    } else {
+      auto pco = peerconnection_observers_.find(ownerTag);
+      if (peerconnection_observers_.end() != pco) {
+        auto stream = pco->second->MediaStreamForId(id);
+        if (stream != nullptr) {
+          return stream;
+        }
+      }
+    }
+  }
+
   auto it = local_streams_.find(id);
   if (it != local_streams_.end()) {
     return (*it).second;
   }
 
-  for (auto kv : peerconnection_observers_) {
-    auto pco = kv.second.get();
-    auto stream = pco->MediaStreamForId(id);
-    if (stream != nullptr) return stream;
-  }
-
   return nullptr;
 }
 
-void FlutterWebRTCBase::RemoveStreamForId(const std::string &id) {
+void FlutterWebRTCBase::RemoveStreamForId(const std::string& id) {
   auto it = local_streams_.find(id);
-  if (it != local_streams_.end()) local_streams_.erase(it);
+  if (it != local_streams_.end())
+    local_streams_.erase(it);
 }
 
-bool FlutterWebRTCBase::ParseConstraints(const EncodableMap &constraints,
-                                         RTCConfiguration *configuration) {
+bool FlutterWebRTCBase::ParseConstraints(const EncodableMap& constraints,
+                                         RTCConfiguration* configuration) {
   memset(&configuration->ice_servers, 0, sizeof(configuration->ice_servers));
   return false;
 }
 
 void FlutterWebRTCBase::ParseConstraints(
-    const EncodableMap &src,
+    const EncodableMap& src,
     scoped_refptr<RTCMediaConstraints> mediaConstraints,
     ParseConstraintType type /*= kMandatory*/) {
   for (auto kv : src) {
@@ -137,7 +160,7 @@ void FlutterWebRTCBase::ParseConstraints(
 }
 
 scoped_refptr<RTCMediaConstraints> FlutterWebRTCBase::ParseMediaConstraints(
-    const EncodableMap &constraints) {
+    const EncodableMap& constraints) {
   scoped_refptr<RTCMediaConstraints> media_constraints =
       RTCMediaConstraints::Create();
 
@@ -158,7 +181,8 @@ scoped_refptr<RTCMediaConstraints> FlutterWebRTCBase::ParseMediaConstraints(
     } else if (TypeIs<EncodableList>(optional)) {
       const EncodableList list = GetValue<EncodableList>(optional);
       for (size_t i = 0; i < list.size(); i++) {
-        ParseConstraints(GetValue<EncodableMap>(list[i]), media_constraints, kOptional);
+        ParseConstraints(GetValue<EncodableMap>(list[i]), media_constraints,
+                         kOptional);
       }
     }
   } else {
@@ -168,49 +192,30 @@ scoped_refptr<RTCMediaConstraints> FlutterWebRTCBase::ParseMediaConstraints(
   return media_constraints;
 }
 
-bool FlutterWebRTCBase::CreateIceServers(const EncodableList &iceServersArray,
-                                         IceServer *ice_servers) {
+bool FlutterWebRTCBase::CreateIceServers(const EncodableList& iceServersArray,
+                                         IceServer* ice_servers) {
   size_t size = iceServersArray.size();
   for (size_t i = 0; i < size; i++) {
-    IceServer &ice_server = ice_servers[i];
+    IceServer& ice_server = ice_servers[i];
     EncodableMap iceServerMap = GetValue<EncodableMap>(iceServersArray[i]);
-    bool hasUsernameAndCredential =
-        iceServerMap.find(EncodableValue("username")) != iceServerMap.end() &&
-        iceServerMap.find(EncodableValue("credential")) != iceServerMap.end();
+
+    if (iceServerMap.find(EncodableValue("username")) != iceServerMap.end()) {
+      ice_server.username = GetValue<std::string>(
+          iceServerMap.find(EncodableValue("username"))->second);
+    }
+    if (iceServerMap.find(EncodableValue("credential")) != iceServerMap.end()) {
+      ice_server.password = GetValue<std::string>(
+          iceServerMap.find(EncodableValue("credential"))->second);
+    }
+
     auto it = iceServerMap.find(EncodableValue("url"));
     if (it != iceServerMap.end() && TypeIs<std::string>(it->second)) {
-      if (hasUsernameAndCredential) {
-        std::string username =
-             GetValue<std::string>(iceServerMap.find(EncodableValue("username"))->second);
-        std::string credential =
-             GetValue<std::string>(iceServerMap.find(EncodableValue("credential"))
-                ->second);
-        std::string uri =  GetValue<std::string>(it->second);
-        ice_server.username = username;
-        ice_server.password = credential;
-        ice_server.uri = uri;
-      } else {
-        std::string uri = GetValue<std::string>(it->second);
-        ice_server.uri = uri;
-      }
+      ice_server.uri = GetValue<std::string>(it->second);
     }
     it = iceServerMap.find(EncodableValue("urls"));
     if (it != iceServerMap.end()) {
       if (TypeIs<std::string>(it->second)) {
-        if (hasUsernameAndCredential) {
-          std::string username =  GetValue<std::string>(iceServerMap.find(EncodableValue("username"))
-                                     ->second);
-          std::string credential =
-               GetValue<std::string>(iceServerMap.find(EncodableValue("credential"))
-                  ->second);
-          std::string uri =  GetValue<std::string>(it->second);
-          ice_server.username = username;
-          ice_server.password = credential;
-          ice_server.uri = uri;
-        } else {
-          std::string uri =  GetValue<std::string>(it->second);
-          ice_server.uri = uri;
-        }
+        ice_server.uri = GetValue<std::string>(it->second);
       }
       if (TypeIs<EncodableList>(it->second)) {
         const EncodableList urls = GetValue<EncodableList>(it->second);
@@ -220,25 +225,10 @@ bool FlutterWebRTCBase::CreateIceServers(const EncodableList &iceServersArray,
             std::string value;
             auto it2 = map.find(EncodableValue("url"));
             if (it2 != map.end()) {
-              value =  GetValue<std::string>(it2->second);
-              if (hasUsernameAndCredential) {
-                std::string username =
-                     GetValue<std::string>(iceServerMap.find(EncodableValue("username"))
-                        ->second);
-                std::string credential =
-                    GetValue<std::string>(iceServerMap.find(EncodableValue("credential"))
-                        ->second);
-                ice_server.username = username;
-                ice_server.password = credential;
-                ice_server.uri = value;
-              } else {
-                ice_server.uri = value;
-              }
+              ice_server.uri = GetValue<std::string>(it2->second);
             }
-          }
-          else if (TypeIs<std::string>(url)) {
-              std::string urlString = GetValue<std::string>(url);
-              ice_server.uri = urlString;
+          } else if (TypeIs<std::string>(url)) {
+            ice_server.uri = GetValue<std::string>(url);
           }
         }
       }
@@ -247,8 +237,8 @@ bool FlutterWebRTCBase::CreateIceServers(const EncodableList &iceServersArray,
   return size > 0;
 }
 
-bool FlutterWebRTCBase::ParseRTCConfiguration(const EncodableMap &map,
-                                              RTCConfiguration &conf) {
+bool FlutterWebRTCBase::ParseRTCConfiguration(const EncodableMap& map,
+                                              RTCConfiguration& conf) {
   auto it = map.find(EncodableValue("iceServers"));
   if (it != map.end()) {
     const EncodableList iceServersArray = GetValue<EncodableList>(it->second);
@@ -306,16 +296,31 @@ bool FlutterWebRTCBase::ParseRTCConfiguration(const EncodableMap &map,
       conf.sdp_semantics = SdpSemantics::kPlanB;
     else if (v == "unified-plan")  // public
       conf.sdp_semantics = SdpSemantics::kUnifiedPlan;
+  } else {
+    conf.sdp_semantics = SdpSemantics::kUnifiedPlan;
+  }
+
+  // maxIPv6Networks
+  it = map.find(EncodableValue("maxIPv6Networks"));
+  if (it != map.end()) {
+    conf.max_ipv6_networks = GetValue<int>(it->second);
   }
   return true;
 }
-
 
 scoped_refptr<RTCMediaTrack> FlutterWebRTCBase::MediaTracksForId(
     const std::string& id) {
   auto it = local_tracks_.find(id);
   if (it != local_tracks_.end()) {
     return (*it).second;
+  }
+
+  for (auto it2 : peerconnection_observers_) {
+    auto pco = it2.second;
+    auto t = pco->MediaTrackForId(id);
+    if (t != nullptr) {
+      return t;
+    }
   }
 
   return nullptr;
@@ -327,5 +332,31 @@ void FlutterWebRTCBase::RemoveTracksForId(const std::string& id) {
     local_tracks_.erase(it);
 }
 
+libwebrtc::scoped_refptr<libwebrtc::RTCRtpSender>
+FlutterWebRTCBase::GetRtpSenderById(RTCPeerConnection* pc, std::string id) {
+  libwebrtc::scoped_refptr<libwebrtc::RTCRtpSender> result;
+  auto senders = pc->senders();
+  for (scoped_refptr<RTCRtpSender> item : senders.std_vector()) {
+    std::string itemId = item->id().std_string();
+    if (nullptr == result.get() && 0 == id.compare(itemId)) {
+      result = item;
+    }
+  }
+  return result;
+}
+
+libwebrtc::scoped_refptr<libwebrtc::RTCRtpReceiver>
+FlutterWebRTCBase::GetRtpReceiverById(RTCPeerConnection* pc,
+                                          std::string id) {
+  libwebrtc::scoped_refptr<libwebrtc::RTCRtpReceiver> result;
+  auto receivers = pc->receivers();
+  for (scoped_refptr<RTCRtpReceiver> item : receivers.std_vector()) {
+    std::string itemId = item->id().std_string();
+    if (nullptr == result.get() && 0 == id.compare(itemId)) {
+      result = item;
+    }
+  }
+  return result;
+}
 
 }  // namespace flutter_webrtc_plugin
