@@ -1,7 +1,15 @@
-@interface AudioUtils : NSObject
+#if TARGET_OS_IPHONE
 
+#import <WebRTC/WebRTC.h>
+
+@interface AudioUtils : NSObject
 + (void)ensureAudioSessionWithRecording:(BOOL)recording;
 // needed for wired headphones to use headphone mic
-+ (void)setPreferHeadphoneInput;
-
++ (BOOL)selectAudioInput:(AVAudioSessionPort)type;
++ (void)setSpeakerphoneOn:(BOOL)enable;
++ (void)setSpeakerphoneOnButPreferBluetooth;
++ (void)deactiveRtcAudioSession;
++ (void) setAppleAudioConfiguration:(NSDictionary*)configuration;
 @end
+
+#endif
