@@ -1,13 +1,13 @@
-#import <WebRTC/WebRTC.h>
 #if TARGET_OS_IPHONE
+#import <WebRTC/WebRTC.h>
 @interface FlutterRPScreenRecorder : RTCVideoCapturer
 
--(void)startCapture;
+- (void)startCapture;
 
 // Stops the capture session asynchronously and notifies callback on completion.
 - (void)stopCaptureWithCompletionHandler:(nullable void (^)(void))completionHandler;
 
--(void)stopCapture;
+- (void)stopCapture;
 
 @end
 #endif
