@@ -1,29 +1,35 @@
 #import <Foundation/Foundation.h>
 #import "FlutterWebRTCPlugin.h"
 
+@interface RTCMediaStreamTrack (Flutter)
+@property(nonatomic, strong, nonnull) id settings;
+@end
+
 @interface FlutterWebRTCPlugin (RTCMediaStream)
 
--(void)getUserMedia:(NSDictionary *)constraints
-             result:(FlutterResult)result;
-#if TARGET_OS_IPHONE
--(void)getDisplayMedia:(NSDictionary *)constraints
-             result:(FlutterResult)result;
-#endif
--(void)createLocalMediaStream:(FlutterResult)result;
+- (void)getUserMedia:(nonnull NSDictionary*)constraints result:(nonnull FlutterResult)result;
 
--(void)getSources:(FlutterResult)result;
+- (void)createLocalMediaStream:(nonnull FlutterResult)result;
 
--(void)mediaStreamTrackHasTorch:(RTCMediaStreamTrack *)track
-                         result:(FlutterResult) result;
+- (void)getSources:(nonnull FlutterResult)result;
 
--(void)mediaStreamTrackSetTorch:(RTCMediaStreamTrack *)track
-                          torch:(BOOL) torch
-                         result:(FlutterResult) result;
+- (void)mediaStreamTrackHasTorch:(nonnull RTCMediaStreamTrack*)track result:(nonnull FlutterResult)result;
 
--(void)mediaStreamTrackSwitchCamera:(RTCMediaStreamTrack *)track
-                             result:(FlutterResult) result;
+- (void)mediaStreamTrackSetTorch:(nonnull RTCMediaStreamTrack*)track
+                           torch:(BOOL)torch
+                          result:(nonnull FlutterResult)result;
 
--(void)mediaStreamTrackCaptureFrame:(RTCMediaStreamTrack *)track
-                             toPath:(NSString *) path
-                             result:(FlutterResult) result;
+- (void)mediaStreamTrackSetZoom:(nonnull RTCMediaStreamTrack*)track
+                           zoomLevel:(double)zoomLevel
+                          result:(nonnull FlutterResult)result;
+
+- (void)mediaStreamTrackSwitchCamera:(nonnull RTCMediaStreamTrack*)track result:(nonnull FlutterResult)result;
+
+- (void)mediaStreamTrackCaptureFrame:(nonnull RTCMediaStreamTrack*)track
+                              toPath:(nonnull NSString*)path
+                              result:(nonnull FlutterResult)result;
+
+- (void)selectAudioInput:(nonnull NSString*)deviceId result:(nullable FlutterResult)result;
+
+- (void)selectAudioOutput:(nonnull NSString*)deviceId result:(nullable FlutterResult)result;
 @end
