@@ -1,51 +1,57 @@
 #import <objc/runtime.h>
-
-#import <WebRTC/WebRTC.h>
-
+#import "AudioUtils.h"
 #import "FlutterRTCFrameCapturer.h"
 #import "FlutterRTCMediaStream.h"
 #import "FlutterRTCPeerConnection.h"
-#import "AudioUtils.h"
 
-#if TARGET_OS_IPHONE
-#import "FlutterRPScreenRecorder.h"
-#endif
+@implementation RTCMediaStreamTrack (Flutter)
+
+- (id)settings {
+  return objc_getAssociatedObject(self, _cmd);
+}
+
+- (void)setSettings:(id)settings {
+  objc_setAssociatedObject(self, @selector(settings), settings, OBJC_ASSOCIATION_RETAIN_NONATOMIC);
+}
+@end
 
 @implementation AVCaptureDevice (Flutter)
 
 - (NSString*)positionString {
   switch (self.position) {
-    case AVCaptureDevicePositionUnspecified: return @"unspecified";
-    case AVCaptureDevicePositionBack: return @"back";
-    case AVCaptureDevicePositionFront: return @"front";
+    case AVCaptureDevicePositionUnspecified:
+      return @"unspecified";
+    case AVCaptureDevicePositionBack:
+      return @"back";
+    case AVCaptureDevicePositionFront:
+      return @"front";
   }
   return nil;
 }
 
 @end
 
-@implementation  FlutterWebRTCPlugin (RTCMediaStream)
+@implementation FlutterWebRTCPlugin (RTCMediaStream)
 
 /**
  * {@link https://www.w3.org/TR/mediacapture-streams/#navigatorusermediaerrorcallback}
  */
-typedef void (^NavigatorUserMediaErrorCallback)(NSString *errorType, NSString *errorMessage);
+typedef void (^NavigatorUserMediaErrorCallback)(NSString* errorType, NSString* errorMessage);
 
 /**
  * {@link https://www.w3.org/TR/mediacapture-streams/#navigatorusermediasuccesscallback}
  */
-typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream *mediaStream);
+typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream* mediaStream);
 
-- (RTCMediaConstraints *)defaultMediaStreamConstraints {
-    NSDictionary *mandatoryConstraints
-    = @{ @"minWidth"     : @"1280",
-         @"minHeight"    : @"720",
-         @"minFrameRate" : @"30" };
-    RTCMediaConstraints* constraints =
-    [[RTCMediaConstraints alloc]
-     initWithMandatoryConstraints:mandatoryConstraints
-     optionalConstraints:nil];
-    return constraints;
+- (NSDictionary*)defaultVideoConstraints {
+    return @{@"minWidth" : @"1280", @"minHeight" : @"720", @"minFrameRate" : @"30"};
+}
+
+- (RTCMediaConstraints*)defaultMediaStreamConstraints {
+  RTCMediaConstraints* constraints =
+      [[RTCMediaConstraints alloc] initWithMandatoryConstraints:[self defaultVideoConstraints]
+                                            optionalConstraints:nil];
+  return constraints;
 }
 
 /**
@@ -65,69 +71,119 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream *mediaStream);
  * new {@code RTCAudioTrack} is to be added, and which is to be reported to
  * {@code successCallback} upon success.
  */
-- (void)getUserAudio:(NSDictionary *)constraints
+- (void)getUserAudio:(NSDictionary*)constraints
      successCallback:(NavigatorUserMediaSuccessCallback)successCallback
        errorCallback:(NavigatorUserMediaErrorCallback)errorCallback
-         mediaStream:(RTCMediaStream *)mediaStream {
-    NSString *trackId = [[NSUUID UUID] UUIDString];
+         mediaStream:(RTCMediaStream*)mediaStream {
+  id audioConstraints = constraints[@"audio"];
+  NSString* audioDeviceId = @"";
 
-    id audioConstraints = constraints[@"audio"];
-    if([audioConstraints isKindOfClass:[NSNumber class]] && [audioConstraints boolValue]) {
-        RTCAudioTrack *audioTrack = [self.peerConnectionFactory audioTrackWithTrackId:trackId];
-        [mediaStream addAudioTrack:audioTrack];
-    } else if([audioConstraints isKindOfClass:[NSDictionary class]]) {
-        RTCAudioSource *audioSource = [self.peerConnectionFactory
-                                       audioSourceWithConstraints:[self parseMediaConstraints:constraints]];
-        RTCAudioTrack *audioTrack
-        = [self.peerConnectionFactory audioTrackWithSource:audioSource trackId:trackId];
-        [mediaStream addAudioTrack:audioTrack];
+  if ([audioConstraints isKindOfClass:[NSDictionary class]]) {
+    // constraints.audio.deviceId
+    NSString* deviceId = audioConstraints[@"deviceId"];
+
+    if (deviceId) {
+      audioDeviceId = deviceId;
     }
 
-    // allow audio capture
-    [AudioUtils ensureAudioSessionWithRecording:YES];
+    // constraints.audio.optional.sourceId
+    id optionalVideoConstraints = audioConstraints[@"optional"];
+    if (optionalVideoConstraints && [optionalVideoConstraints isKindOfClass:[NSArray class]] &&
+        !deviceId) {
+      NSArray* options = optionalVideoConstraints;
+      for (id item in options) {
+        if ([item isKindOfClass:[NSDictionary class]]) {
+          NSString* sourceId = ((NSDictionary*)item)[@"sourceId"];
+          if (sourceId) {
+            audioDeviceId = sourceId;
+          }
+        }
+      }
+    }
+  }
 
-    successCallback(mediaStream);
+#if !defined(TARGET_OS_IPHONE)
+  if (audioDeviceId != nil) {
+    [self selectAudioInput:audioDeviceId result:nil];
+  }
+#endif
+
+  NSString* trackId = [[NSUUID UUID] UUIDString];
+  RTCAudioTrack* audioTrack = [self.peerConnectionFactory audioTrackWithTrackId:trackId];
+
+  audioTrack.settings = @{
+    @"deviceId" : audioDeviceId,
+    @"kind" : @"audioinput",
+    @"autoGainControl" : @YES,
+    @"echoCancellation" : @YES,
+    @"noiseSuppression" : @YES,
+    @"channelCount" : @1,
+    @"latency" : @0,
+  };
+
+  [mediaStream addAudioTrack:audioTrack];
+
+  [self.localTracks setObject:audioTrack forKey:trackId];
+
+  [self ensureAudioSession];
+
+  successCallback(mediaStream);
 }
 
 // TODO: Use RCTConvert for constraints ...
--(void)getUserMedia:(NSDictionary *)constraints
-             result:(FlutterResult) result  {
-    // Initialize RTCMediaStream with a unique label in order to allow multiple
-    // RTCMediaStream instances initialized by multiple getUserMedia calls to be
-    // added to 1 RTCPeerConnection instance. As suggested by
-    // https://www.w3.org/TR/mediacapture-streams/#mediastream to be a good
-    // practice, use a UUID (conforming to RFC4122).
-    NSString *mediaStreamId = [[NSUUID UUID] UUIDString];
-    RTCMediaStream *mediaStream
-    = [self.peerConnectionFactory mediaStreamWithStreamId:mediaStreamId];
+- (void)getUserMedia:(NSDictionary*)constraints result:(FlutterResult)result {
+  // Initialize RTCMediaStream with a unique label in order to allow multiple
+  // RTCMediaStream instances initialized by multiple getUserMedia calls to be
+  // added to 1 RTCPeerConnection instance. As suggested by
+  // https://www.w3.org/TR/mediacapture-streams/#mediastream to be a good
+  // practice, use a UUID (conforming to RFC4122).
+  NSString* mediaStreamId = [[NSUUID UUID] UUIDString];
+  RTCMediaStream* mediaStream = [self.peerConnectionFactory mediaStreamWithStreamId:mediaStreamId];
 
-    [self
-     getUserMedia:constraints
-     successCallback:^ (RTCMediaStream *mediaStream) {
-         NSString *mediaStreamId = mediaStream.streamId;
+  [self getUserMedia:constraints
+      successCallback:^(RTCMediaStream* mediaStream) {
+        NSString* mediaStreamId = mediaStream.streamId;
 
-         NSMutableArray *audioTracks = [NSMutableArray array];
-         NSMutableArray *videoTracks = [NSMutableArray array];
+        NSMutableArray* audioTracks = [NSMutableArray array];
+        NSMutableArray* videoTracks = [NSMutableArray array];
 
-         for (RTCAudioTrack *track in mediaStream.audioTracks) {
-             [self.localTracks setObject:track forKey:track.trackId];
-             [audioTracks addObject:@{@"id": track.trackId, @"kind": track.kind, @"label": track.trackId, @"enabled": @(track.isEnabled), @"remote": @(YES), @"readyState": @"live"}];
-         }
+        for (RTCAudioTrack* track in mediaStream.audioTracks) {
+          [audioTracks addObject:@{
+            @"id" : track.trackId,
+            @"kind" : track.kind,
+            @"label" : track.trackId,
+            @"enabled" : @(track.isEnabled),
+            @"remote" : @(YES),
+            @"readyState" : @"live",
+            @"settings" : track.settings
+          }];
+        }
 
-         for (RTCVideoTrack *track in mediaStream.videoTracks) {
-             [self.localTracks setObject:track forKey:track.trackId];
-             [videoTracks addObject:@{@"id": track.trackId, @"kind": track.kind, @"label": track.trackId, @"enabled": @(track.isEnabled), @"remote": @(YES), @"readyState": @"live"}];
-         }
+        for (RTCVideoTrack* track in mediaStream.videoTracks) {
+          [videoTracks addObject:@{
+            @"id" : track.trackId,
+            @"kind" : track.kind,
+            @"label" : track.trackId,
+            @"enabled" : @(track.isEnabled),
+            @"remote" : @(YES),
+            @"readyState" : @"live",
+            @"settings" : track.settings
+          }];
+        }
 
-         self.localStreams[mediaStreamId] = mediaStream;
-         result(@{@"streamId": mediaStreamId, @"audioTracks" : audioTracks, @"videoTracks" : videoTracks });
-     }
-     errorCallback:^ (NSString *errorType, NSString *errorMessage) {
-         result([FlutterError errorWithCode:[NSString stringWithFormat:@"Error %@", errorType]
-                                    message:errorMessage
-                                    details:nil]);
-     }
-     mediaStream:mediaStream];
+        self.localStreams[mediaStreamId] = mediaStream;
+        result(@{
+          @"streamId" : mediaStreamId,
+          @"audioTracks" : audioTracks,
+          @"videoTracks" : videoTracks
+        });
+      }
+      errorCallback:^(NSString* errorType, NSString* errorMessage) {
+        result([FlutterError errorWithCode:[NSString stringWithFormat:@"Error %@", errorType]
+                                   message:errorMessage
+                                   details:nil]);
+      }
+      mediaStream:mediaStream];
 }
 
 /**
@@ -153,82 +209,79 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream *mediaStream);
  * @param mediaStream The {@link RTCMediaStream} which is being initialized as
  * part of the execution of the {@code getUserMedia()} algorithm.
  */
-- (void)getUserMedia:(NSDictionary *)constraints
+- (void)getUserMedia:(NSDictionary*)constraints
      successCallback:(NavigatorUserMediaSuccessCallback)successCallback
        errorCallback:(NavigatorUserMediaErrorCallback)errorCallback
-         mediaStream:(RTCMediaStream *)mediaStream {
-    // If mediaStream contains no audioTracks and the constraints request such a
-    // track, then run an iteration of the getUserMedia() algorithm to obtain
-    // local audio content.
-    if (mediaStream.audioTracks.count == 0) {
-        // constraints.audio
-        id audioConstraints = constraints[@"audio"];
-        BOOL constraintsIsDictionary = [audioConstraints isKindOfClass:[NSDictionary class]];
-        if (audioConstraints && (constraintsIsDictionary || [audioConstraints boolValue])) {
-            [self requestAccessForMediaType:AVMediaTypeAudio
-                                constraints:constraints
-                            successCallback:successCallback
-                              errorCallback:errorCallback
-                                mediaStream:mediaStream];
-            return;
-        }
+         mediaStream:(RTCMediaStream*)mediaStream {
+  // If mediaStream contains no audioTracks and the constraints request such a
+  // track, then run an iteration of the getUserMedia() algorithm to obtain
+  // local audio content.
+  if (mediaStream.audioTracks.count == 0) {
+    // constraints.audio
+    id audioConstraints = constraints[@"audio"];
+    BOOL constraintsIsDictionary = [audioConstraints isKindOfClass:[NSDictionary class]];
+    if (audioConstraints && (constraintsIsDictionary || [audioConstraints boolValue])) {
+      [self requestAccessForMediaType:AVMediaTypeAudio
+                          constraints:constraints
+                      successCallback:successCallback
+                        errorCallback:errorCallback
+                          mediaStream:mediaStream];
+      return;
     }
+  }
 
-    // If mediaStream contains no videoTracks and the constraints request such a
-    // track, then run an iteration of the getUserMedia() algorithm to obtain
-    // local video content.
-    if (mediaStream.videoTracks.count == 0) {
-        // constraints.video
-        id videoConstraints = constraints[@"video"];
-        if (videoConstraints) {
-            BOOL requestAccessForVideo
-            = [videoConstraints isKindOfClass:[NSNumber class]]
-            ? [videoConstraints boolValue]
-            : [videoConstraints isKindOfClass:[NSDictionary class]];
+  // If mediaStream contains no videoTracks and the constraints request such a
+  // track, then run an iteration of the getUserMedia() algorithm to obtain
+  // local video content.
+  if (mediaStream.videoTracks.count == 0) {
+    // constraints.video
+    id videoConstraints = constraints[@"video"];
+    if (videoConstraints) {
+      BOOL requestAccessForVideo = [videoConstraints isKindOfClass:[NSNumber class]]
+                                       ? [videoConstraints boolValue]
+                                       : [videoConstraints isKindOfClass:[NSDictionary class]];
 #if !TARGET_IPHONE_SIMULATOR
-            if (requestAccessForVideo) {
-                [self requestAccessForMediaType:AVMediaTypeVideo
-                                    constraints:constraints
-                                successCallback:successCallback
-                                  errorCallback:errorCallback
-                                    mediaStream:mediaStream];
-                return;
-            }
+      if (requestAccessForVideo) {
+        [self requestAccessForMediaType:AVMediaTypeVideo
+                            constraints:constraints
+                        successCallback:successCallback
+                          errorCallback:errorCallback
+                            mediaStream:mediaStream];
+        return;
+      }
 #endif
-        }
     }
+  }
 
-    // There are audioTracks and/or videoTracks in mediaStream as requested by
-    // constraints so the getUserMedia() is to conclude with success.
-    successCallback(mediaStream);
+  // There are audioTracks and/or videoTracks in mediaStream as requested by
+  // constraints so the getUserMedia() is to conclude with success.
+  successCallback(mediaStream);
 }
 
-- (int)getConstrainInt:(NSDictionary *)constraints
-                forKey:(NSString *)key {
-    
-    if (![constraints isKindOfClass:[NSDictionary class]]) {
-        return 0;
-    }
-
-    id constraint = constraints[key];
-    if ([constraint isKindOfClass:[NSNumber class]]) {
-        return [constraint intValue];
-    } else if ([constraint isKindOfClass:[NSString class]]) {
-        int possibleValue = [constraint intValue];
-        if (possibleValue != 0) {
-            return possibleValue;
-        }
-    } else if ([constraint isKindOfClass:[NSDictionary class]]) {
-        id idealConstraint = constraint[@"ideal"];
-        if([idealConstraint isKindOfClass: [NSString class]]) {
-            int possibleValue = [idealConstraint intValue];
-            if(possibleValue != 0){
-                return possibleValue;
-            }
-        }
-    }
-    
+- (int)getConstrainInt:(NSDictionary*)constraints forKey:(NSString*)key {
+  if (![constraints isKindOfClass:[NSDictionary class]]) {
     return 0;
+  }
+
+  id constraint = constraints[key];
+  if ([constraint isKindOfClass:[NSNumber class]]) {
+    return [constraint intValue];
+  } else if ([constraint isKindOfClass:[NSString class]]) {
+    int possibleValue = [constraint intValue];
+    if (possibleValue != 0) {
+      return possibleValue;
+    }
+  } else if ([constraint isKindOfClass:[NSDictionary class]]) {
+    id idealConstraint = constraint[@"ideal"];
+    if ([idealConstraint isKindOfClass:[NSString class]]) {
+      int possibleValue = [idealConstraint intValue];
+      if (possibleValue != 0) {
+        return possibleValue;
+      }
+    }
+  }
+
+  return 0;
 }
 
 /**
@@ -248,149 +301,216 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream *mediaStream);
  * new {@code RTCVideoTrack} is to be added, and which is to be reported to
  * {@code successCallback} upon success.
  */
-- (void)getUserVideo:(NSDictionary *)constraints
+- (void)getUserVideo:(NSDictionary*)constraints
      successCallback:(NavigatorUserMediaSuccessCallback)successCallback
        errorCallback:(NavigatorUserMediaErrorCallback)errorCallback
-         mediaStream:(RTCMediaStream *)mediaStream {
-    id videoConstraints = constraints[@"video"];
-    AVCaptureDevice *videoDevice;
-    if ([videoConstraints isKindOfClass:[NSDictionary class]]) {
-        // constraints.video.optional
-        id optionalVideoConstraints = videoConstraints[@"optional"];
-        if (optionalVideoConstraints
-            && [optionalVideoConstraints isKindOfClass:[NSArray class]]) {
-            NSArray *options = optionalVideoConstraints;
-            for (id item in options) {
-                if ([item isKindOfClass:[NSDictionary class]]) {
-                    NSString *sourceId = ((NSDictionary *)item)[@"sourceId"];
-                    if (sourceId) {
-                        videoDevice = [AVCaptureDevice deviceWithUniqueID:sourceId];
-                        if (videoDevice) {
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-        if (!videoDevice) {
-            // constraints.video.facingMode
-            //
-            // https://www.w3.org/TR/mediacapture-streams/#def-constraint-facingMode
-            id facingMode = videoConstraints[@"facingMode"];
-            if (facingMode && [facingMode isKindOfClass:[NSString class]]) {
-                AVCaptureDevicePosition position;
-                if ([facingMode isEqualToString:@"environment"]) {
-                    self._usingFrontCamera = NO;
-                    position = AVCaptureDevicePositionBack;
-                } else if ([facingMode isEqualToString:@"user"]) {
-                    self._usingFrontCamera = YES;
-                    position = AVCaptureDevicePositionFront;
-                } else {
-                    // If the specified facingMode value is not supported, fall back to
-                    // the default video device.
-                    self._usingFrontCamera = NO;
-                    position = AVCaptureDevicePositionUnspecified;
-                }
-                videoDevice = [self findDeviceForPosition:position];
-            }
-        }
-        if (!videoDevice) {
-            videoDevice = [AVCaptureDevice defaultDeviceWithMediaType:AVMediaTypeVideo];
-        }
+         mediaStream:(RTCMediaStream*)mediaStream {
+  id videoConstraints = constraints[@"video"];
+  AVCaptureDevice* videoDevice;
+  NSString* videoDeviceId = nil;
+  NSString* facingMode = nil;
+
+  if ([videoConstraints isKindOfClass:[NSDictionary class]]) {
+    // constraints.video.deviceId
+    NSString* deviceId = videoConstraints[@"deviceId"];
+
+    if (deviceId) {
+      videoDevice = [AVCaptureDevice deviceWithUniqueID:deviceId];
+      videoDeviceId = deviceId;
     }
 
-    //TODO(rostopira): refactor to separate function and add support for max
-
-    self._targetWidth = 1280;
-    self._targetHeight = 720;
-    self._targetFps = 30;
-
-    if (!videoDevice && [constraints[@"video"] boolValue] == YES) {
-        videoDevice = [AVCaptureDevice defaultDeviceWithMediaType:AVMediaTypeVideo];
+    // constraints.video.optional
+    id optionalVideoConstraints = videoConstraints[@"optional"];
+    if (optionalVideoConstraints && [optionalVideoConstraints isKindOfClass:[NSArray class]] &&
+        !videoDevice) {
+      NSArray* options = optionalVideoConstraints;
+      for (id item in options) {
+        if ([item isKindOfClass:[NSDictionary class]]) {
+          NSString* sourceId = ((NSDictionary*)item)[@"sourceId"];
+          if (sourceId) {
+            videoDevice = [AVCaptureDevice deviceWithUniqueID:sourceId];
+            if (videoDevice) {
+              videoDeviceId = sourceId;
+              break;
+            }
+          }
+        }
+      }
     }
 
-    id mandatory = [videoConstraints isKindOfClass:[NSDictionary class]]? videoConstraints[@"mandatory"] : nil ;
-
-    // constraints.video.mandatory
-    if(mandatory && [mandatory isKindOfClass:[NSDictionary class]])
-    {
-        id widthConstraint = mandatory[@"minWidth"];
-        if ([widthConstraint isKindOfClass:[NSString class]]) {
-            int possibleWidth = [widthConstraint intValue];
-            if (possibleWidth != 0) {
-                self._targetWidth = possibleWidth;
-            }
+    if (!videoDevice) {
+      // constraints.video.facingMode
+      // https://www.w3.org/TR/mediacapture-streams/#def-constraint-facingMode
+      facingMode = videoConstraints[@"facingMode"];
+      if (facingMode && [facingMode isKindOfClass:[NSString class]]) {
+        AVCaptureDevicePosition position;
+        if ([facingMode isEqualToString:@"environment"]) {
+          self._usingFrontCamera = NO;
+          position = AVCaptureDevicePositionBack;
+        } else if ([facingMode isEqualToString:@"user"]) {
+          self._usingFrontCamera = YES;
+          position = AVCaptureDevicePositionFront;
+        } else {
+          // If the specified facingMode value is not supported, fall back to
+          // the default video device.
+          self._usingFrontCamera = NO;
+          position = AVCaptureDevicePositionUnspecified;
         }
-        id heightConstraint = mandatory[@"minHeight"];
-        if ([heightConstraint isKindOfClass:[NSString class]]) {
-            int possibleHeight = [heightConstraint intValue];
-            if (possibleHeight != 0) {
-                self._targetHeight = possibleHeight;
-            }
-        }
-        id fpsConstraint = mandatory[@"minFrameRate"];
-        if ([fpsConstraint isKindOfClass:[NSString class]]) {
-            int possibleFps = [fpsConstraint intValue];
-            if (possibleFps != 0) {
-                self._targetFps = possibleFps;
-            }
-        }
+        videoDevice = [self findDeviceForPosition:position];
+      }
     }
+  }
 
-    int possibleWidth = [self getConstrainInt:videoConstraints forKey:@"width"];
-    if(possibleWidth != 0){
-        self._targetWidth = possibleWidth;
+  if ([videoConstraints isKindOfClass:[NSNumber class]]) {
+    videoConstraints = @{@"mandatory": [self defaultVideoConstraints]};
+  }
+
+  NSInteger targetWidth = 0;
+  NSInteger targetHeight = 0;
+  NSInteger targetFps = 0;
+
+  if (!videoDevice) {
+    videoDevice = [AVCaptureDevice defaultDeviceWithMediaType:AVMediaTypeVideo];
+  }
+
+  int possibleWidth = [self getConstrainInt:videoConstraints forKey:@"width"];
+  if (possibleWidth != 0) {
+    targetWidth = possibleWidth;
+  }
+
+  int possibleHeight = [self getConstrainInt:videoConstraints forKey:@"height"];
+  if (possibleHeight != 0) {
+    targetHeight = possibleHeight;
+  }
+
+  int possibleFps = [self getConstrainInt:videoConstraints forKey:@"frameRate"];
+  if (possibleFps != 0) {
+    targetFps = possibleFps;
+  }
+
+  id mandatory =
+      [videoConstraints isKindOfClass:[NSDictionary class]] ? videoConstraints[@"mandatory"] : nil;
+
+  // constraints.video.mandatory
+  if (mandatory && [mandatory isKindOfClass:[NSDictionary class]]) {
+    id widthConstraint = mandatory[@"minWidth"];
+    if ([widthConstraint isKindOfClass:[NSString class]] ||
+        [widthConstraint isKindOfClass:[NSNumber class]]) {
+      int possibleWidth = [widthConstraint intValue];
+      if (possibleWidth != 0) {
+        targetWidth = possibleWidth;
+      }
     }
+    id heightConstraint = mandatory[@"minHeight"];
+    if ([heightConstraint isKindOfClass:[NSString class]] ||
+        [heightConstraint isKindOfClass:[NSNumber class]]) {
+      int possibleHeight = [heightConstraint intValue];
+      if (possibleHeight != 0) {
+        targetHeight = possibleHeight;
+      }
+    }
+    id fpsConstraint = mandatory[@"minFrameRate"];
+    if ([fpsConstraint isKindOfClass:[NSString class]] ||
+        [fpsConstraint isKindOfClass:[NSNumber class]]) {
+      int possibleFps = [fpsConstraint intValue];
+      if (possibleFps != 0) {
+        targetFps = possibleFps;
+      }
+    }
+  }
+
+  if (videoDevice) {
+    RTCVideoSource* videoSource = [self.peerConnectionFactory videoSource];
+    if (self.videoCapturer) {
+      [self.videoCapturer stopCapture];
+    }
+    self.videoCapturer = [[RTCCameraVideoCapturer alloc] initWithDelegate:videoSource];
+    AVCaptureDeviceFormat* selectedFormat = [self selectFormatForDevice:videoDevice
+                                                            targetWidth:targetWidth
+                                                           targetHeight:targetHeight];
+
+    CMVideoDimensions selectedDimension = CMVideoFormatDescriptionGetDimensions(selectedFormat.formatDescription);
+    NSInteger selectedWidth = (NSInteger) selectedDimension.width;
+    NSInteger selectedHeight = (NSInteger) selectedDimension.height;
+    NSInteger selectedFps = [self selectFpsForFormat:selectedFormat targetFps:targetFps];
+
+    self._lastTargetFps = selectedFps;
+    self._lastTargetWidth = targetWidth;
+    self._lastTargetHeight = targetHeight;
     
-    int possibleHeight = [self getConstrainInt:videoConstraints forKey:@"height"];
-    if(possibleHeight != 0){
-        self._targetHeight = possibleHeight;
-    }
-    
-    int possibleFps = [self getConstrainInt:videoConstraints forKey:@"frameRate"];
-    if(possibleFps != 0){
-        self._targetFps = possibleFps;
+    NSLog(@"target format %ldx%ld, targetFps: %ld, selected format: %ldx%ld, selected fps %ld", targetWidth, targetHeight, targetFps, selectedWidth, selectedHeight, selectedFps);
+
+    if ([videoDevice lockForConfiguration:NULL]) {
+      @try {
+        videoDevice.activeVideoMaxFrameDuration = CMTimeMake(1, (int32_t)selectedFps);
+        videoDevice.activeVideoMinFrameDuration = CMTimeMake(1, (int32_t)selectedFps);
+      } @catch (NSException* exception) {
+        NSLog(@"Failed to set active frame rate!\n User info:%@", exception.userInfo);
+      }
+      [videoDevice unlockForConfiguration];
     }
 
-    if (videoDevice) {
-        RTCVideoSource *videoSource = [self.peerConnectionFactory videoSource];
-        if (self.videoCapturer) {
-            [self.videoCapturer stopCapture];
-        }
-        self.videoCapturer = [[RTCCameraVideoCapturer alloc] initWithDelegate:videoSource];
-        AVCaptureDeviceFormat *selectedFormat = [self selectFormatForDevice:videoDevice];
-        NSInteger selectedFps = [self selectFpsForFormat:selectedFormat];
-        [self.videoCapturer startCaptureWithDevice:videoDevice format:selectedFormat fps:selectedFps completionHandler:^(NSError *error) {
-            if (error) {
-                NSLog(@"Start capture error: %@", [error localizedDescription]);
-            }
-        }];
+    [self.videoCapturer startCaptureWithDevice:videoDevice
+                                        format:selectedFormat
+                                           fps:selectedFps
+                             completionHandler:^(NSError* error) {
+                               if (error) {
+                                 NSLog(@"Start capture error: %@", [error localizedDescription]);
+                               }
+                             }];
 
-        NSString *trackUUID = [[NSUUID UUID] UUIDString];
-        RTCVideoTrack *videoTrack = [self.peerConnectionFactory videoTrackWithSource:videoSource trackId:trackUUID];
-        [mediaStream addVideoTrack:videoTrack];
+    NSString* trackUUID = [[NSUUID UUID] UUIDString];
+    RTCVideoTrack* videoTrack = [self.peerConnectionFactory videoTrackWithSource:videoSource
+                                                                         trackId:trackUUID];
 
-        successCallback(mediaStream);
-    } else {
-        // According to step 6.2.3 of the getUserMedia() algorithm, if there is no
-        // source, fail with a new OverconstrainedError.
-        errorCallback(@"OverconstrainedError", /* errorMessage */ nil);
+    __weak RTCCameraVideoCapturer* capturer = self.videoCapturer;
+    self.videoCapturerStopHandlers[videoTrack.trackId] = ^(CompletionHandler handler) {
+      NSLog(@"Stop video capturer, trackID %@", videoTrack.trackId);
+      [capturer stopCaptureWithCompletionHandler:handler];
+    };
+
+    if (!videoDeviceId) {
+      videoDeviceId = videoDevice.uniqueID;
     }
+
+    if (!facingMode) {
+      facingMode = videoDevice.position == AVCaptureDevicePositionBack    ? @"environment"
+                   : videoDevice.position == AVCaptureDevicePositionFront ? @"user"
+                                                                          : @"unspecified";
+    }
+
+    videoTrack.settings = @{
+      @"deviceId" : videoDeviceId,
+      @"kind" : @"videoinput",
+      @"width" : [NSNumber numberWithInteger:selectedWidth],
+      @"height" : [NSNumber numberWithInteger:selectedHeight],
+      @"frameRate" : [NSNumber numberWithInteger:selectedFps],
+      @"facingMode" : facingMode,
+    };
+
+    [mediaStream addVideoTrack:videoTrack];
+
+    [self.localTracks setObject:videoTrack forKey:trackUUID];
+
+    successCallback(mediaStream);
+  } else {
+    // According to step 6.2.3 of the getUserMedia() algorithm, if there is no
+    // source, fail with a new OverconstrainedError.
+    errorCallback(@"OverconstrainedError", /* errorMessage */ nil);
+  }
 }
 
--(void)mediaStreamRelease:(RTCMediaStream *)stream
-{
-    if (stream) {
-        for (RTCVideoTrack *track in stream.videoTracks) {
-            [self.localTracks removeObjectForKey:track.trackId];
-        }
-        for (RTCAudioTrack *track in stream.audioTracks) {
-            [self.localTracks removeObjectForKey:track.trackId];
-        }
-        [self.localStreams removeObjectForKey:stream.streamId];
+- (void)mediaStreamRelease:(RTCMediaStream*)stream {
+  if (stream) {
+    for (RTCVideoTrack* track in stream.videoTracks) {
+      [self.localTracks removeObjectForKey:track.trackId];
     }
+    for (RTCAudioTrack* track in stream.audioTracks) {
+      [self.localTracks removeObjectForKey:track.trackId];
+    }
+    [self.localStreams removeObjectForKey:stream.streamId];
+  }
 }
-
 
 /**
  * Obtains local media content of a specific type. Requests access for the
@@ -408,291 +528,419 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream *mediaStream);
  * @param mediaStream The {@link RTCMediaStream} which is to collect the
  * obtained local media content of the specified {@code mediaType}.
  */
-- (void)requestAccessForMediaType:(NSString *)mediaType
-                      constraints:(NSDictionary *)constraints
+- (void)requestAccessForMediaType:(NSString*)mediaType
+                      constraints:(NSDictionary*)constraints
                   successCallback:(NavigatorUserMediaSuccessCallback)successCallback
                     errorCallback:(NavigatorUserMediaErrorCallback)errorCallback
-                      mediaStream:(RTCMediaStream *)mediaStream {
-    // According to step 6.2.1 of the getUserMedia() algorithm, if there is no
-    // source, fail "with a new DOMException object whose name attribute has the
-    // value NotFoundError."
-    // XXX The following approach does not work for audio in Simulator. That is
-    // because audio capture is done using AVAudioSession which does not use
-    // AVCaptureDevice there. Anyway, Simulator will not (visually) request access
-    // for audio.
-    if (mediaType == AVMediaTypeVideo
-        && [AVCaptureDevice devicesWithMediaType:mediaType].count == 0) {
-        // Since successCallback and errorCallback are asynchronously invoked
-        // elsewhere, make sure that the invocation here is consistent.
-        dispatch_async(dispatch_get_main_queue(), ^ {
-            errorCallback(@"DOMException", @"NotFoundError");
-        });
-        return;
-    }
+                      mediaStream:(RTCMediaStream*)mediaStream {
+  // According to step 6.2.1 of the getUserMedia() algorithm, if there is no
+  // source, fail "with a new DOMException object whose name attribute has the
+  // value NotFoundError."
+  // XXX The following approach does not work for audio in Simulator. That is
+  // because audio capture is done using AVAudioSession which does not use
+  // AVCaptureDevice there. Anyway, Simulator will not (visually) request access
+  // for audio.
+  if (mediaType == AVMediaTypeVideo &&
+      [AVCaptureDevice devicesWithMediaType:mediaType].count == 0) {
+    // Since successCallback and errorCallback are asynchronously invoked
+    // elsewhere, make sure that the invocation here is consistent.
+    dispatch_async(dispatch_get_main_queue(), ^{
+      errorCallback(@"DOMException", @"NotFoundError");
+    });
+    return;
+  }
 
 #if TARGET_OS_OSX
-    if (@available(macOS 10.14, *)) {
+  if (@available(macOS 10.14, *)) {
 #endif
-        [AVCaptureDevice
-         requestAccessForMediaType:mediaType
-         completionHandler:^ (BOOL granted) {
-            dispatch_async(dispatch_get_main_queue(), ^ {
-                if (granted) {
-                    NavigatorUserMediaSuccessCallback scb
-                    = ^ (RTCMediaStream *mediaStream) {
-                        [self getUserMedia:constraints
-                           successCallback:successCallback
-                             errorCallback:errorCallback
-                               mediaStream:mediaStream];
-                    };
+    [AVCaptureDevice requestAccessForMediaType:mediaType
+                             completionHandler:^(BOOL granted) {
+                               dispatch_async(dispatch_get_main_queue(), ^{
+                                 if (granted) {
+                                   NavigatorUserMediaSuccessCallback scb =
+                                       ^(RTCMediaStream* mediaStream) {
+                                         [self getUserMedia:constraints
+                                             successCallback:successCallback
+                                               errorCallback:errorCallback
+                                                 mediaStream:mediaStream];
+                                       };
 
-                    if (mediaType == AVMediaTypeAudio) {
-                        [self getUserAudio:constraints
-                           successCallback:scb
-                             errorCallback:errorCallback
-                               mediaStream:mediaStream];
-                    } else if (mediaType == AVMediaTypeVideo) {
-                        [self getUserVideo:constraints
-                           successCallback:scb
-                             errorCallback:errorCallback
-                               mediaStream:mediaStream];
-                    }
-                } else {
-                    // According to step 10 Permission Failure of the getUserMedia()
-                    // algorithm, if the user has denied permission, fail "with a new
-                    // DOMException object whose name attribute has the value
-                    // NotAllowedError."
-                    errorCallback(@"DOMException", @"NotAllowedError");
-                }
-            });
-        }];
+                                   if (mediaType == AVMediaTypeAudio) {
+                                     [self getUserAudio:constraints
+                                         successCallback:scb
+                                           errorCallback:errorCallback
+                                             mediaStream:mediaStream];
+                                   } else if (mediaType == AVMediaTypeVideo) {
+                                     [self getUserVideo:constraints
+                                         successCallback:scb
+                                           errorCallback:errorCallback
+                                             mediaStream:mediaStream];
+                                   }
+                                 } else {
+                                   // According to step 10 Permission Failure of the getUserMedia()
+                                   // algorithm, if the user has denied permission, fail "with a new
+                                   // DOMException object whose name attribute has the value
+                                   // NotAllowedError."
+                                   errorCallback(@"DOMException", @"NotAllowedError");
+                                 }
+                               });
+                             }];
 #if TARGET_OS_OSX
-    } else {
-        // Fallback on earlier versions
-        NavigatorUserMediaSuccessCallback scb
-        = ^ (RTCMediaStream *mediaStream) {
-            [self getUserMedia:constraints
-               successCallback:successCallback
-                 errorCallback:errorCallback
-                   mediaStream:mediaStream];
-        };
-        if (mediaType == AVMediaTypeAudio) {
-            [self getUserAudio:constraints
-               successCallback:scb
-                 errorCallback:errorCallback
-                   mediaStream:mediaStream];
-        } else if (mediaType == AVMediaTypeVideo) {
-            [self getUserVideo:constraints
-               successCallback:scb
-                 errorCallback:errorCallback
-                   mediaStream:mediaStream];
-        }
+  } else {
+    // Fallback on earlier versions
+    NavigatorUserMediaSuccessCallback scb = ^(RTCMediaStream* mediaStream) {
+      [self getUserMedia:constraints
+          successCallback:successCallback
+            errorCallback:errorCallback
+              mediaStream:mediaStream];
+    };
+    if (mediaType == AVMediaTypeAudio) {
+      [self getUserAudio:constraints
+          successCallback:scb
+            errorCallback:errorCallback
+              mediaStream:mediaStream];
+    } else if (mediaType == AVMediaTypeVideo) {
+      [self getUserVideo:constraints
+          successCallback:scb
+            errorCallback:errorCallback
+              mediaStream:mediaStream];
     }
+  }
 #endif
 }
 
+- (void)createLocalMediaStream:(FlutterResult)result {
+  NSString* mediaStreamId = [[NSUUID UUID] UUIDString];
+  RTCMediaStream* mediaStream = [self.peerConnectionFactory mediaStreamWithStreamId:mediaStreamId];
+
+  self.localStreams[mediaStreamId] = mediaStream;
+  result(@{@"streamId" : [mediaStream streamId]});
+}
+
+- (void)getSources:(FlutterResult)result {
+  NSMutableArray* sources = [NSMutableArray array];
+  NSArray* videoDevices = [AVCaptureDevice devicesWithMediaType:AVMediaTypeVideo];
+  for (AVCaptureDevice* device in videoDevices) {
+    [sources addObject:@{
+      @"facing" : device.positionString,
+      @"deviceId" : device.uniqueID,
+      @"label" : device.localizedName,
+      @"kind" : @"videoinput",
+    }];
+  }
 #if TARGET_OS_IPHONE
--(void)getDisplayMedia:(NSDictionary *)constraints
-                result:(FlutterResult)result {
-    NSString *mediaStreamId = [[NSUUID UUID] UUIDString];
-    RTCMediaStream *mediaStream = [self.peerConnectionFactory mediaStreamWithStreamId:mediaStreamId];
 
-    RTCVideoSource *videoSource = [self.peerConnectionFactory videoSource];
-    FlutterRPScreenRecorder *screenCapturer = [[FlutterRPScreenRecorder alloc] initWithDelegate:videoSource];
+  RTCAudioSession* session = [RTCAudioSession sharedInstance];
+  for (AVAudioSessionPortDescription* port in session.session.availableInputs) {
+    // NSLog(@"input portName: %@, type %@", port.portName,port.portType);
+    [sources addObject:@{
+      @"deviceId" : port.UID,
+      @"label" : port.portName,
+      @"groupId" : port.portType,
+      @"kind" : @"audioinput",
+    }];
+  }
 
-    [screenCapturer startCapture];
-
-    //TODO:
-    self.videoCapturer = screenCapturer;
-
-    NSString *trackUUID = [[NSUUID UUID] UUIDString];
-    RTCVideoTrack *videoTrack = [self.peerConnectionFactory videoTrackWithSource:videoSource trackId:trackUUID];
-    [mediaStream addVideoTrack:videoTrack];
-
-    NSMutableArray *audioTracks = [NSMutableArray array];
-    NSMutableArray *videoTracks = [NSMutableArray array];
-
-    for (RTCVideoTrack *track in mediaStream.videoTracks) {
-        [self.localTracks setObject:track forKey:track.trackId];
-        [videoTracks addObject:@{@"id": track.trackId, @"kind": track.kind, @"label": track.trackId, @"enabled": @(track.isEnabled), @"remote": @(YES), @"readyState": @"live"}];
+  for (AVAudioSessionPortDescription* port in session.currentRoute.outputs) {
+    // NSLog(@"output portName: %@, type %@", port.portName,port.portType);
+    if (session.currentRoute.outputs.count == 1 && ![port.UID isEqualToString:@"Speaker"]) {
+      [sources addObject:@{
+        @"deviceId" : @"Speaker",
+        @"label" : @"Speaker",
+        @"groupId" : @"Speaker",
+        @"kind" : @"audiooutput",
+      }];
     }
-
-    self.localStreams[mediaStreamId] = mediaStream;
-    result(@{@"streamId": mediaStreamId, @"audioTracks" : audioTracks, @"videoTracks" : videoTracks });
-}
+    [sources addObject:@{
+      @"deviceId" : port.UID,
+      @"label" : port.portName,
+      @"groupId" : port.portType,
+      @"kind" : @"audiooutput",
+    }];
+  }
 #endif
--(void)createLocalMediaStream:(FlutterResult)result{
-    NSString *mediaStreamId = [[NSUUID UUID] UUIDString];
-    RTCMediaStream *mediaStream = [self.peerConnectionFactory mediaStreamWithStreamId:mediaStreamId];
+#if TARGET_OS_OSX
+  RTCAudioDeviceModule* audioDeviceModule = [self.peerConnectionFactory audioDeviceModule];
 
-    self.localStreams[mediaStreamId] = mediaStream;
-    result(@{@"streamId": [mediaStream streamId] });
+  NSArray* inputDevices = [audioDeviceModule inputDevices];
+  for (RTCIODevice* device in inputDevices) {
+    [sources addObject:@{
+      @"deviceId" : device.deviceId,
+      @"label" : device.name,
+      @"kind" : @"audioinput",
+    }];
+  }
+
+  NSArray* outputDevices = [audioDeviceModule outputDevices];
+  for (RTCIODevice* device in outputDevices) {
+    [sources addObject:@{
+      @"deviceId" : device.deviceId,
+      @"label" : device.name,
+      @"kind" : @"audiooutput",
+    }];
+  }
+#endif
+  result(@{@"sources" : sources});
 }
 
--(void)getSources:(FlutterResult)result{
-  NSMutableArray *sources = [NSMutableArray array];
-  NSArray *videoDevices = [AVCaptureDevice devicesWithMediaType:AVMediaTypeVideo];
-  for (AVCaptureDevice *device in videoDevices) {
-    [sources addObject:@{
-                         @"facing": device.positionString,
-                         @"deviceId": device.uniqueID,
-                         @"label": device.localizedName,
-                         @"kind": @"videoinput",
-                         }];
+- (void)selectAudioInput:(NSString*)deviceId result:(FlutterResult)result {
+#if TARGET_OS_OSX
+  RTCAudioDeviceModule* audioDeviceModule = [self.peerConnectionFactory audioDeviceModule];
+  NSArray* inputDevices = [audioDeviceModule inputDevices];
+  for (RTCIODevice* device in inputDevices) {
+    if ([deviceId isEqualToString:device.deviceId]) {
+      [audioDeviceModule setInputDevice:device];
+      if (result)
+        result(nil);
+      return;
+    }
   }
-  NSArray *audioDevices = [AVCaptureDevice devicesWithMediaType:AVMediaTypeAudio];
-  for (AVCaptureDevice *device in audioDevices) {
-    [sources addObject:@{
-                         @"facing": @"",
-                         @"deviceId": device.uniqueID,
-                         @"label": device.localizedName,
-                         @"kind": @"audioinput",
-                         }];
+#endif
+#if TARGET_OS_IPHONE
+  RTCAudioSession* session = [RTCAudioSession sharedInstance];
+  for (AVAudioSessionPortDescription* port in session.session.availableInputs) {
+    if ([port.UID isEqualToString:deviceId]) {
+      if (self.preferredInput != port.portType) {
+        self.preferredInput = port.portType;
+        [AudioUtils selectAudioInput:self.preferredInput];
+      }
+      break;
+    }
   }
-    result(@{@"sources": sources});
+  if (result)
+    result(nil);
+#endif
+  if (result)
+    result([FlutterError errorWithCode:@"selectAudioInputFailed"
+                               message:[NSString stringWithFormat:@"Error: deviceId not found!"]
+                               details:nil]);
 }
 
--(void)mediaStreamTrackRelease:(RTCMediaStream *)mediaStream  track:(RTCMediaStreamTrack *)track
-{
+- (void)selectAudioOutput:(NSString*)deviceId result:(FlutterResult)result {
+#if TARGET_OS_OSX
+  RTCAudioDeviceModule* audioDeviceModule = [self.peerConnectionFactory audioDeviceModule];
+  NSArray* outputDevices = [audioDeviceModule outputDevices];
+  for (RTCIODevice* device in outputDevices) {
+    if ([deviceId isEqualToString:device.deviceId]) {
+      [audioDeviceModule setOutputDevice:device];
+      result(nil);
+      return;
+    }
+  }
+#endif
+#if TARGET_OS_IPHONE
+  RTCAudioSession* session = [RTCAudioSession sharedInstance];
+  NSError* setCategoryError = nil;
+
+  if ([deviceId isEqualToString:@"Speaker"]) {
+    [session.session overrideOutputAudioPort:kAudioSessionOverrideAudioRoute_Speaker
+                                       error:&setCategoryError];
+  } else {
+    [session.session overrideOutputAudioPort:kAudioSessionOverrideAudioRoute_None
+                                       error:&setCategoryError];
+  }
+
+  if (setCategoryError == nil) {
+    result(nil);
+    return;
+  }
+
+  result([FlutterError
+      errorWithCode:@"selectAudioOutputFailed"
+            message:[NSString
+                        stringWithFormat:@"Error: %@", [setCategoryError localizedFailureReason]]
+            details:nil]);
+
+#endif
+  result([FlutterError errorWithCode:@"selectAudioOutputFailed"
+                             message:[NSString stringWithFormat:@"Error: deviceId not found!"]
+                             details:nil]);
+}
+
+- (void)mediaStreamTrackRelease:(RTCMediaStream*)mediaStream track:(RTCMediaStreamTrack*)track {
   // what's different to mediaStreamTrackStop? only call mediaStream explicitly?
   if (mediaStream && track) {
     track.isEnabled = NO;
     // FIXME this is called when track is removed from the MediaStream,
     // but it doesn't mean it can not be added back using MediaStream.addTrack
-    //TODO: [self.localTracks removeObjectForKey:trackID];
+    // TODO: [self.localTracks removeObjectForKey:trackID];
     if ([track.kind isEqualToString:@"audio"]) {
-      [mediaStream removeAudioTrack:(RTCAudioTrack *)track];
-    } else if([track.kind isEqualToString:@"video"]) {
-      [mediaStream removeVideoTrack:(RTCVideoTrack *)track];
+      [mediaStream removeAudioTrack:(RTCAudioTrack*)track];
+    } else if ([track.kind isEqualToString:@"video"]) {
+      [mediaStream removeVideoTrack:(RTCVideoTrack*)track];
     }
   }
 }
 
--(void)mediaStreamTrackSetEnabled:(RTCMediaStreamTrack *)track : (BOOL)enabled
-{
-  if (track && track.isEnabled != enabled) {
-    track.isEnabled = enabled;
+- (void)mediaStreamTrackHasTorch:(RTCMediaStreamTrack*)track result:(FlutterResult)result {
+  if (!self.videoCapturer) {
+    result(@NO);
+    return;
+  }
+  if (self.videoCapturer.captureSession.inputs.count == 0) {
+    result(@NO);
+    return;
+  }
+
+  AVCaptureDeviceInput* deviceInput = [self.videoCapturer.captureSession.inputs objectAtIndex:0];
+  AVCaptureDevice* device = deviceInput.device;
+
+  result(@([device isTorchModeSupported:AVCaptureTorchModeOn]));
+}
+
+- (void)mediaStreamTrackSetTorch:(RTCMediaStreamTrack*)track
+                           torch:(BOOL)torch
+                          result:(FlutterResult)result {
+  if (!self.videoCapturer) {
+    NSLog(@"Video capturer is null. Can't set torch");
+    return;
+  }
+  if (self.videoCapturer.captureSession.inputs.count == 0) {
+    NSLog(@"Video capturer is missing an input. Can't set torch");
+    return;
+  }
+
+  AVCaptureDeviceInput* deviceInput = [self.videoCapturer.captureSession.inputs objectAtIndex:0];
+  AVCaptureDevice* device = deviceInput.device;
+
+  if (![device isTorchModeSupported:AVCaptureTorchModeOn]) {
+    NSLog(@"Current capture device does not support torch. Can't set torch");
+    return;
+  }
+
+  NSError* error;
+  if ([device lockForConfiguration:&error] == NO) {
+    NSLog(@"Failed to aquire configuration lock. %@", error.localizedDescription);
+    return;
+  }
+
+  device.torchMode = torch ? AVCaptureTorchModeOn : AVCaptureTorchModeOff;
+  [device unlockForConfiguration];
+
+  result(nil);
+}
+
+- (void)mediaStreamTrackSetZoom:(RTCMediaStreamTrack*)track
+                           zoomLevel:(double)zoomLevel
+                          result:(FlutterResult)result {
+#if TARGET_OS_OSX
+  NSLog(@"Not supported on macOS. Can't set zoom");
+  return;
+#endif
+#if TARGET_OS_IPHONE
+  if (!self.videoCapturer) {
+    NSLog(@"Video capturer is null. Can't set zoom");
+    return;
+  }
+  if (self.videoCapturer.captureSession.inputs.count == 0) {
+    NSLog(@"Video capturer is missing an input. Can't set zoom");
+    return;
+  }
+
+  AVCaptureDeviceInput* deviceInput = [self.videoCapturer.captureSession.inputs objectAtIndex:0];
+  AVCaptureDevice* device = deviceInput.device;
+
+  NSError* error;
+  if ([device lockForConfiguration:&error] == NO) {
+    NSLog(@"Failed to acquire configuration lock. %@", error.localizedDescription);
+    return;
+  }
+  
+  CGFloat desiredZoomFactor = (CGFloat)zoomLevel;
+  device.videoZoomFactor = MAX(1.0, MIN(desiredZoomFactor, device.activeFormat.videoMaxZoomFactor));
+  [device unlockForConfiguration];
+
+  result(nil);
+#endif
+}
+
+- (void)mediaStreamTrackSwitchCamera:(RTCMediaStreamTrack*)track result:(FlutterResult)result {
+  if (!self.videoCapturer) {
+    NSLog(@"Video capturer is null. Can't switch camera");
+    return;
+  }
+  self._usingFrontCamera = !self._usingFrontCamera;
+  AVCaptureDevicePosition position =
+      self._usingFrontCamera ? AVCaptureDevicePositionFront : AVCaptureDevicePositionBack;
+  AVCaptureDevice* videoDevice = [self findDeviceForPosition:position];
+  AVCaptureDeviceFormat* selectedFormat = [self selectFormatForDevice:videoDevice
+                                                          targetWidth:self._lastTargetWidth
+                                                         targetHeight:self._lastTargetHeight];
+  [self.videoCapturer startCaptureWithDevice:videoDevice
+                                      format:selectedFormat
+                                         fps:[self selectFpsForFormat:selectedFormat
+                                                            targetFps:self._lastTargetFps]
+                           completionHandler:^(NSError* error) {
+                             if (error != nil) {
+                               result([FlutterError errorWithCode:@"Error while switching camera"
+                                                          message:@"Error while switching camera"
+                                                          details:error]);
+                             } else {
+                               result([NSNumber numberWithBool:self._usingFrontCamera]);
+                             }
+                           }];
+}
+
+- (void)mediaStreamTrackCaptureFrame:(RTCVideoTrack*)track
+                              toPath:(NSString*)path
+                              result:(FlutterResult)result {
+  self.frameCapturer = [[FlutterRTCFrameCapturer alloc] initWithTrack:track
+                                                               toPath:path
+                                                               result:result];
+}
+
+- (void)mediaStreamTrackStop:(RTCMediaStreamTrack*)track {
+  if (track) {
+    track.isEnabled = NO;
+    [self.localTracks removeObjectForKey:track.trackId];
   }
 }
 
--(void)mediaStreamTrackHasTorch:(RTCMediaStreamTrack *)track result:(FlutterResult) result
-{
-    if (!self.videoCapturer) {
-        result(@NO);
-        return;
+- (AVCaptureDevice*)findDeviceForPosition:(AVCaptureDevicePosition)position {
+  if (position == AVCaptureDevicePositionUnspecified) {
+    return [AVCaptureDevice defaultDeviceWithMediaType:AVMediaTypeVideo];
+  }
+  NSArray<AVCaptureDevice*>* captureDevices = [RTCCameraVideoCapturer captureDevices];
+  for (AVCaptureDevice* device in captureDevices) {
+    if (device.position == position) {
+      return device;
     }
-    if (self.videoCapturer.captureSession.inputs.count == 0) {
-        result(@NO);
-        return;
-    }
-
-    AVCaptureDeviceInput *deviceInput = [self.videoCapturer.captureSession.inputs objectAtIndex:0];
-    AVCaptureDevice *device = deviceInput.device;
-
-    result(@([device isTorchModeSupported:AVCaptureTorchModeOn]));
-}
-
--(void)mediaStreamTrackSetTorch:(RTCMediaStreamTrack *)track torch:(BOOL)torch result:(FlutterResult)result
-{
-    if (!self.videoCapturer) {
-        NSLog(@"Video capturer is null. Can't set torch");
-        return;
-    }
-    if (self.videoCapturer.captureSession.inputs.count == 0) {
-        NSLog(@"Video capturer is missing an input. Can't set torch");
-        return;
-    }
-
-    AVCaptureDeviceInput *deviceInput = [self.videoCapturer.captureSession.inputs objectAtIndex:0];
-    AVCaptureDevice *device = deviceInput.device;
-
-    if (![device isTorchModeSupported:AVCaptureTorchModeOn]) {
-        NSLog(@"Current capture device does not support torch. Can't set torch");
-        return;
-    }
-
-    NSError *error;
-    if ([device lockForConfiguration:&error] == NO) {
-        NSLog(@"Failed to aquire configuration lock. %@", error.localizedDescription);
-        return;
-    }
-
-    device.torchMode = torch ? AVCaptureTorchModeOn : AVCaptureTorchModeOff;
-    [device unlockForConfiguration];
-
-    result(nil);
-}
-
--(void)mediaStreamTrackSwitchCamera:(RTCMediaStreamTrack *)track result:(FlutterResult)result
-{
-    if (!self.videoCapturer) {
-        NSLog(@"Video capturer is null. Can't switch camera");
-        return;
-    }
-    self._usingFrontCamera = !self._usingFrontCamera;
-    AVCaptureDevicePosition position = self._usingFrontCamera ? AVCaptureDevicePositionFront : AVCaptureDevicePositionBack;
-    AVCaptureDevice *videoDevice = [self findDeviceForPosition:position];
-    AVCaptureDeviceFormat *selectedFormat = [self selectFormatForDevice:videoDevice];
-    [self.videoCapturer startCaptureWithDevice:videoDevice format:selectedFormat fps:[self selectFpsForFormat:selectedFormat] completionHandler:^(NSError* error){
-        if (error != nil) {
-            result([FlutterError errorWithCode:@"Error while switching camera" message:@"Error while switching camera" details:error]);
-        } else {
-            result([NSNumber numberWithBool:self._usingFrontCamera]);
-        }
-    }];
-}
-
--(void)mediaStreamTrackCaptureFrame:(RTCVideoTrack *)track toPath:(NSString *) path result:(FlutterResult)result
-{
-    self.frameCapturer = [[FlutterRTCFrameCapturer alloc] initWithTrack:track toPath:path result:result];
-}
-
--(void)mediaStreamTrackStop:(RTCMediaStreamTrack *)track
-{
-    if (track) {
-        track.isEnabled = NO;
-        [self.localTracks removeObjectForKey:track.trackId];
-    }
-}
-
-- (AVCaptureDevice *)findDeviceForPosition:(AVCaptureDevicePosition)position {
-    if (position == AVCaptureDevicePositionUnspecified) {
-        return [AVCaptureDevice defaultDeviceWithMediaType:AVMediaTypeVideo];
-    }
-    NSArray<AVCaptureDevice *> *captureDevices = [RTCCameraVideoCapturer captureDevices];
-    for (AVCaptureDevice *device in captureDevices) {
-        if (device.position == position) {
-            return device;
-        }
-    }
+  }
+  if(captureDevices.count > 0) {
     return captureDevices[0];
+  }
+  return nil;
 }
 
-- (AVCaptureDeviceFormat *)selectFormatForDevice:(AVCaptureDevice *)device {
-    NSArray<AVCaptureDeviceFormat *> *formats =
-    [RTCCameraVideoCapturer supportedFormatsForDevice:device];
-    AVCaptureDeviceFormat *selectedFormat = nil;
-    int currentDiff = INT_MAX;
-    for (AVCaptureDeviceFormat *format in formats) {
-        CMVideoDimensions dimension = CMVideoFormatDescriptionGetDimensions(format.formatDescription);
-        FourCharCode pixelFormat = CMFormatDescriptionGetMediaSubType(format.formatDescription);
-        int diff = abs(self._targetWidth - dimension.width) + abs(self._targetHeight - dimension.height);
-        if (diff < currentDiff) {
-            selectedFormat = format;
-            currentDiff = diff;
-        } else if (diff == currentDiff && pixelFormat == [self.videoCapturer preferredOutputPixelFormat]) {
-            selectedFormat = format;
-        }
+- (AVCaptureDeviceFormat*)selectFormatForDevice:(AVCaptureDevice*)device
+                                    targetWidth:(NSInteger)targetWidth
+                                   targetHeight:(NSInteger)targetHeight {
+  NSArray<AVCaptureDeviceFormat*>* formats =
+      [RTCCameraVideoCapturer supportedFormatsForDevice:device];
+  AVCaptureDeviceFormat* selectedFormat = nil;
+  long currentDiff = INT_MAX;
+  for (AVCaptureDeviceFormat* format in formats) {
+    CMVideoDimensions dimension = CMVideoFormatDescriptionGetDimensions(format.formatDescription);
+    FourCharCode pixelFormat = CMFormatDescriptionGetMediaSubType(format.formatDescription);
+    //NSLog(@"AVCaptureDeviceFormats,fps %d, dimension: %dx%d", format.videoSupportedFrameRateRanges, dimension.width, dimension.height);
+      long diff = labs(targetWidth - dimension.width) + labs(targetHeight - dimension.height);
+    if (diff < currentDiff) {
+      selectedFormat = format;
+      currentDiff = diff;
+    } else if (diff == currentDiff &&
+               pixelFormat == [self.videoCapturer preferredOutputPixelFormat]) {
+      selectedFormat = format;
     }
-    return selectedFormat;
+  }
+  return selectedFormat;
 }
 
-- (NSInteger)selectFpsForFormat:(AVCaptureDeviceFormat *)format {
-    Float64 maxSupportedFramerate = 0;
-    for (AVFrameRateRange *fpsRange in format.videoSupportedFrameRateRanges) {
-        maxSupportedFramerate = fmax(maxSupportedFramerate, fpsRange.maxFrameRate);
-    }
-    return fmin(maxSupportedFramerate, self._targetFps);
+- (NSInteger)selectFpsForFormat:(AVCaptureDeviceFormat*)format targetFps:(NSInteger)targetFps {
+  Float64 maxSupportedFramerate = 0;
+  for (AVFrameRateRange* fpsRange in format.videoSupportedFrameRateRanges) {
+    maxSupportedFramerate = fmax(maxSupportedFramerate, fpsRange.maxFrameRate);
+  }
+  return fmin(maxSupportedFramerate, targetFps);
 }
 
 @end
