@@ -1,19 +1,20 @@
 #import "FlutterWebRTCPlugin.h"
 
-#import <WebRTC/RTCVideoRenderer.h>
 #import <WebRTC/RTCMediaStream.h>
 #import <WebRTC/RTCVideoFrame.h>
+#import <WebRTC/RTCVideoRenderer.h>
 #import <WebRTC/RTCVideoTrack.h>
 
-@interface FlutterRTCVideoRenderer : NSObject <FlutterTexture, RTCVideoRenderer, FlutterStreamHandler>
+@interface FlutterRTCVideoRenderer
+    : NSObject <FlutterTexture, RTCVideoRenderer, FlutterStreamHandler>
 
 /**
  * The {@link RTCVideoTrack}, if any, which this instance renders.
  */
-@property (nonatomic, strong) RTCVideoTrack *videoTrack;
-@property (nonatomic) int64_t textureId;
-@property (nonatomic, weak) id<FlutterTextureRegistry> registry;
-@property (nonatomic, strong) FlutterEventSink eventSink;
+@property(nonatomic, strong) RTCVideoTrack* videoTrack;
+@property(nonatomic) int64_t textureId;
+@property(nonatomic, weak) id<FlutterTextureRegistry> registry;
+@property(nonatomic, strong) FlutterEventSink eventSink;
 
 - (instancetype)initWithTextureRegistry:(id<FlutterTextureRegistry>)registry
                               messenger:(NSObject<FlutterBinaryMessenger>*)messenger;
@@ -22,12 +23,11 @@
 
 @end
 
-
 @interface FlutterWebRTCPlugin (FlutterVideoRendererManager)
 
-- (FlutterRTCVideoRenderer *)createWithTextureRegistry:(id<FlutterTextureRegistry>)registry
-                       messenger:(NSObject<FlutterBinaryMessenger>*)messenger;
+- (FlutterRTCVideoRenderer*)createWithTextureRegistry:(id<FlutterTextureRegistry>)registry
+                                            messenger:(NSObject<FlutterBinaryMessenger>*)messenger;
 
--(void)rendererSetSrcObject:(FlutterRTCVideoRenderer*)renderer stream:(RTCVideoTrack*)videoTrack;
+- (void)rendererSetSrcObject:(FlutterRTCVideoRenderer*)renderer stream:(RTCVideoTrack*)videoTrack;
 
 @end
