@@ -1,9 +1,10 @@
-import 'dart:async';
 import 'dart:core';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_background/flutter_background.dart';
 import 'package:flutter_webrtc/flutter_webrtc.dart';
+import 'package:flutter_webrtc_example/src/widgets/screen_select_dialog.dart';
 
 /*
  * getDisplayMedia sample
@@ -19,8 +20,7 @@ class _GetDisplayMediaSampleState extends State<GetDisplayMediaSample> {
   MediaStream? _localStream;
   final RTCVideoRenderer _localRenderer = RTCVideoRenderer();
   bool _inCalling = false;
-  Timer? _timer;
-  var _counter = 0;
+  DesktopCapturerSource? selected_source_;
 
   @override
   void initState() {
@@ -34,7 +34,6 @@ class _GetDisplayMediaSampleState extends State<GetDisplayMediaSample> {
     if (_inCalling) {
       _stop();
     }
-    _timer?.cancel();
     _localRenderer.dispose();
   }
 
@@ -42,19 +41,68 @@ class _GetDisplayMediaSampleState extends State<GetDisplayMediaSample> {
     await _localRenderer.initialize();
   }
 
-  void handleTimer(Timer timer) async {
-    setState(() {
-      _counter++;
-    });
+  Future<void> selectScreenSourceDialog(BuildContext context) async {
+    if (WebRTC.platformIsDesktop) {
+      final source = await showDialog<DesktopCapturerSource>(
+        context: context,
+        builder: (context) => ScreenSelectDialog(),
+      );
+      if (source != null) {
+        await _makeCall(source);
+      }
+    } else {
+      if (WebRTC.platformIsAndroid) {
+        // Android specific
+        Future<void> requestBackgroundPermission([bool isRetry = false]) async {
+          // Required for android screenshare.
+          try {
+            var hasPermissions = await FlutterBackground.hasPermissions;
+            if (!isRetry) {
+              const androidConfig = FlutterBackgroundAndroidConfig(
+                notificationTitle: 'Screen Sharing',
+                notificationText: 'LiveKit Example is sharing the screen.',
+                notificationImportance: AndroidNotificationImportance.normal,
+                notificationIcon: AndroidResource(
+                    name: 'livekit_ic_launcher', defType: 'mipmap'),
+              );
+              hasPermissions = await FlutterBackground.initialize(
+                  androidConfig: androidConfig);
+            }
+            if (hasPermissions &&
+                !FlutterBackground.isBackgroundExecutionEnabled) {
+              await FlutterBackground.enableBackgroundExecution();
+            }
+          } catch (e) {
+            if (!isRetry) {
+              return await Future<void>.delayed(const Duration(seconds: 1),
+                  () => requestBackgroundPermission(true));
+            }
+            print('could not publish video: $e');
+          }
+        }
+
+        await requestBackgroundPermission();
+      }
+      await _makeCall(null);
+    }
   }
 
   // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> _makeCall() async {
-    final mediaConstraints = <String, dynamic>{'audio': true, 'video': true};
+  Future<void> _makeCall(DesktopCapturerSource? source) async {
+    setState(() {
+      selected_source_ = source;
+    });
 
     try {
       var stream =
-          await navigator.mediaDevices.getDisplayMedia(mediaConstraints);
+          await navigator.mediaDevices.getDisplayMedia(<String, dynamic>{
+        'video': selected_source_ == null
+            ? true
+            : {
+                'deviceId': {'exact': selected_source_!.id},
+                'mandatory': {'frameRate': 30.0}
+              }
+      });
       stream.getVideoTracks()[0].onEnded = () {
         print(
             'By adding a listener on onEnded you can: 1) catch stop video sharing on Web');
@@ -70,8 +118,6 @@ class _GetDisplayMediaSampleState extends State<GetDisplayMediaSample> {
     setState(() {
       _inCalling = true;
     });
-
-    _timer = Timer.periodic(Duration(milliseconds: 100), handleTimer);
   }
 
   Future<void> _stop() async {
@@ -92,36 +138,39 @@ class _GetDisplayMediaSampleState extends State<GetDisplayMediaSample> {
     setState(() {
       _inCalling = false;
     });
-    _timer?.cancel();
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('GetUserMedia API Test'),
+        title: Text('GetDisplayMedia source: ' +
+            (selected_source_ != null ? selected_source_!.name : '')),
         actions: [],
       ),
       body: OrientationBuilder(
         builder: (context, orientation) {
           return Center(
+              child: Container(
+            width: MediaQuery.of(context).size.width,
+            color: Colors.white10,
             child: Stack(children: <Widget>[
-              Center(
-                child: Text('counter: ' + _counter.toString()),
-              ),
-              Container(
-                margin: EdgeInsets.fromLTRB(0.0, 0.0, 0.0, 0.0),
-                width: MediaQuery.of(context).size.width,
-                height: MediaQuery.of(context).size.height,
-                decoration: BoxDecoration(color: Colors.black54),
-                child: RTCVideoView(_localRenderer),
-              )
+              if (_inCalling)
+                Container(
+                  margin: EdgeInsets.fromLTRB(0.0, 0.0, 0.0, 0.0),
+                  width: MediaQuery.of(context).size.width,
+                  height: MediaQuery.of(context).size.height,
+                  decoration: BoxDecoration(color: Colors.black54),
+                  child: RTCVideoView(_localRenderer),
+                )
             ]),
-          );
+          ));
         },
       ),
       floatingActionButton: FloatingActionButton(
-        onPressed: _inCalling ? _hangUp : _makeCall,
+        onPressed: () {
+          _inCalling ? _hangUp() : selectScreenSourceDialog(context);
+        },
         tooltip: _inCalling ? 'Hangup' : 'Call',
         child: Icon(_inCalling ? Icons.call_end : Icons.phone),
       ),
