@@ -1,11 +1,10 @@
-import 'package:webrtc_interface/src/media_recorder.dart' as _interface;
-import 'package:webrtc_interface/webrtc_interface.dart';
+import 'package:webrtc_interface/webrtc_interface.dart' as rtc;
 
 import '../flutter_webrtc.dart';
 
-class MediaRecorder extends _interface.MediaRecorder {
+class MediaRecorder extends rtc.MediaRecorder {
   MediaRecorder() : _delegate = mediaRecorder();
-  final _interface.MediaRecorder _delegate;
+  final rtc.MediaRecorder _delegate;
 
   @override
   Future<void> start(String path,
@@ -20,7 +19,12 @@ class MediaRecorder extends _interface.MediaRecorder {
     MediaStream stream, {
     Function(dynamic blob, bool isLastOne)? onDataChunk,
     String? mimeType,
+    int timeSlice = 1000,
   }) =>
-      _delegate.startWeb(stream,
-          onDataChunk: onDataChunk, mimeType: mimeType ?? 'video/webm');
+      _delegate.startWeb(
+        stream,
+        onDataChunk: onDataChunk,
+        mimeType: mimeType ?? 'video/webm',
+        timeSlice: timeSlice,
+      );
 }
