@@ -3,6 +3,7 @@ import 'dart:math';
 
 import 'package:webrtc_interface/webrtc_interface.dart';
 
+import 'media_stream_track_impl.dart';
 import 'utils.dart';
 
 class MediaRecorderNative extends MediaRecorder {
@@ -22,13 +23,18 @@ class MediaRecorderNative extends MediaRecorder {
       'path': path,
       if (audioChannel != null) 'audioChannel': audioChannel.index,
       if (videoTrack != null) 'videoTrackId': videoTrack.id,
-      'recorderId': _recorderId
+      'recorderId': _recorderId,
+      'peerConnectionId': videoTrack is MediaStreamTrackNative
+          ? videoTrack.peerConnectionId
+          : null
     });
   }
 
   @override
   void startWeb(MediaStream stream,
-      {Function(dynamic blob, bool isLastOne)? onDataChunk, String? mimeType}) {
+      {Function(dynamic blob, bool isLastOne)? onDataChunk,
+      String? mimeType,
+      int timeSlice = 1000}) {
     throw 'It\'s for Flutter Web only';
   }
 
