@@ -2,12 +2,12 @@ import 'dart:async';
 
 import 'package:webrtc_interface/webrtc_interface.dart';
 
+import 'factory_impl.dart';
 import 'media_stream_track_impl.dart';
 import 'utils.dart';
 
 class MediaStreamNative extends MediaStream {
-  MediaStreamNative(String streamId, String ownerTag)
-      : super(streamId, ownerTag);
+  MediaStreamNative(super.streamId, super.ownerTag);
 
   factory MediaStreamNative.fromMap(Map<dynamic, dynamic> map) {
     return MediaStreamNative(map['streamId'], map['ownerTag'])
@@ -20,16 +20,16 @@ class MediaStreamNative extends MediaStream {
   void setMediaTracks(List<dynamic> audioTracks, List<dynamic> videoTracks) {
     _audioTracks.clear();
 
-    audioTracks.forEach((track) {
-      _audioTracks.add(MediaStreamTrackNative(
-          track['id'], track['label'], track['kind'], track['enabled']));
-    });
+    for (var track in audioTracks) {
+      _audioTracks.add(MediaStreamTrackNative(track['id'], track['label'],
+          track['kind'], track['enabled'], ownerTag, track['settings'] ?? {}));
+    }
 
     _videoTracks.clear();
-    videoTracks.forEach((track) {
-      _videoTracks.add(MediaStreamTrackNative(
-          track['id'], track['label'], track['kind'], track['enabled']));
-    });
+    for (var track in videoTracks) {
+      _videoTracks.add(MediaStreamTrackNative(track['id'], track['label'],
+          track['kind'], track['enabled'], ownerTag, track['settings'] ?? {}));
+    }
   }
 
   @override
@@ -88,7 +88,7 @@ class MediaStreamNative extends MediaStream {
   }
 
   @override
-  Future<Null> dispose() async {
+  Future<void> dispose() async {
     await WebRTC.invokeMethod(
       'streamDispose',
       <String, dynamic>{'streamId': id},
@@ -100,8 +100,11 @@ class MediaStreamNative extends MediaStream {
   bool get active => throw UnimplementedError();
 
   @override
-  MediaStream clone() {
-    // TODO(cloudwebrtc): Implement
-    throw UnimplementedError();
+  Future<MediaStream> clone() async {
+    final cloneStream = await createLocalMediaStream(id);
+    for (var track in [..._audioTracks, ..._videoTracks]) {
+      await cloneStream.addTrack(track);
+    }
+    return cloneStream;
   }
 }
