@@ -4,10 +4,29 @@ import 'package:flutter/services.dart';
 
 import 'package:webrtc_interface/webrtc_interface.dart';
 
+import 'event_channel.dart';
 import 'media_stream_impl.dart';
 import 'utils.dart';
 
 class MediaDeviceNative extends MediaDevices {
+  MediaDeviceNative._internal() {
+    FlutterWebRTCEventChannel.instance.handleEvents.stream.listen((data) {
+      var event = data.keys.first;
+      Map<dynamic, dynamic> map = data.values.first;
+      handleEvent(event, map);
+    });
+  }
+
+  static final MediaDeviceNative instance = MediaDeviceNative._internal();
+
+  void handleEvent(String event, final Map<dynamic, dynamic> map) async {
+    switch (map['event']) {
+      case 'onDeviceChange':
+        ondevicechange?.call(null);
+        break;
+    }
+  }
+
   @override
   Future<MediaStream> getUserMedia(
       Map<String, dynamic> mediaConstraints) async {
@@ -68,8 +87,8 @@ class MediaDeviceNative extends MediaDevices {
 
   @override
   Future<List<MediaDeviceInfo>> enumerateDevices() async {
-    var _source = await getSources();
-    return _source
+    var source = await getSources();
+    return source
         .map(
           (e) => MediaDeviceInfo(
               deviceId: e['deviceId'],
@@ -78,5 +97,15 @@ class MediaDeviceNative extends MediaDevices {
               label: e['label']),
         )
         .toList();
+  }
+
+  @override
+  Future<MediaDeviceInfo> selectAudioOutput(
+      [AudioOutputOptions? options]) async {
+    await WebRTC.invokeMethod('selectAudioOutput', {
+      'deviceId': options?.deviceId,
+    });
+    // TODO(cloudwebrtc): return the selected device
+    return MediaDeviceInfo(label: 'label', deviceId: options!.deviceId);
   }
 }
