@@ -3,6 +3,10 @@ import 'package:webrtc_interface/webrtc_interface.dart';
 import 'mediadevices_impl.dart';
 
 class NavigatorNative extends Navigator {
+  NavigatorNative._internal();
+
+  static final NavigatorNative instance = NavigatorNative._internal();
+
   @override
   Future<MediaStream> getDisplayMedia(Map<String, dynamic> mediaConstraints) {
     return mediaDevices.getDisplayMedia(mediaConstraints);
@@ -19,5 +23,5 @@ class NavigatorNative extends Navigator {
   }
 
   @override
-  MediaDevices get mediaDevices => MediaDeviceNative();
+  MediaDevices get mediaDevices => MediaDeviceNative.instance;
 }
