@@ -47,13 +47,85 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
   RTCSignalingState? get signalingState => _signalingState;
 
   @override
+  Future<RTCSignalingState?> getSignalingState() async {
+    try {
+      final response =
+          await WebRTC.invokeMethod('getSignalingState', <String, dynamic>{
+        'peerConnectionId': _peerConnectionId,
+      });
+
+      if (null == response) {
+        return null;
+      }
+      _signalingState = signalingStateForString(response['state']);
+      return _signalingState;
+    } on PlatformException catch (e) {
+      throw 'Unable to RTCPeerConnection::getSignalingState: ${e.message}';
+    }
+  }
+
+  @override
   RTCIceGatheringState? get iceGatheringState => _iceGatheringState;
+
+  @override
+  Future<RTCIceGatheringState?> getIceGatheringState() async {
+    try {
+      final response =
+          await WebRTC.invokeMethod('getIceGatheringState', <String, dynamic>{
+        'peerConnectionId': _peerConnectionId,
+      });
+
+      if (null == response) {
+        return null;
+      }
+      _iceGatheringState = iceGatheringStateforString(response['state']);
+      return _iceGatheringState;
+    } on PlatformException catch (e) {
+      throw 'Unable to RTCPeerConnection::getIceGatheringState: ${e.message}';
+    }
+  }
 
   @override
   RTCIceConnectionState? get iceConnectionState => _iceConnectionState;
 
   @override
+  Future<RTCIceConnectionState?> getIceConnectionState() async {
+    try {
+      final response =
+          await WebRTC.invokeMethod('getIceConnectionState', <String, dynamic>{
+        'peerConnectionId': _peerConnectionId,
+      });
+
+      if (null == response) {
+        return null;
+      }
+      _iceConnectionState = iceConnectionStateForString(response['state']);
+      return _iceConnectionState;
+    } on PlatformException catch (e) {
+      throw 'Unable to RTCPeerConnection::getIceConnectionState: ${e.message}';
+    }
+  }
+
+  @override
   RTCPeerConnectionState? get connectionState => _connectionState;
+
+  @override
+  Future<RTCPeerConnectionState?> getConnectionState() async {
+    try {
+      final response =
+          await WebRTC.invokeMethod('getConnectionState', <String, dynamic>{
+        'peerConnectionId': _peerConnectionId,
+      });
+
+      if (null == response) {
+        return null;
+      }
+      _connectionState = peerConnectionStateForString(response['state']);
+      return _connectionState;
+    } on PlatformException catch (e) {
+      throw 'Unable to RTCPeerConnection::getConnectionState: ${e.message}';
+    }
+  }
 
   Future<RTCSessionDescription?> get localDescription => getLocalDescription();
 
@@ -118,7 +190,12 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
         Map<dynamic, dynamic> track = map['track'];
 
         var newTrack = MediaStreamTrackNative(
-            track['id'], track['label'], track['kind'], track['enabled']);
+            track['id'],
+            track['label'],
+            track['kind'],
+            track['enabled'],
+            _peerConnectionId,
+            track['settings'] ?? {});
         String kind = track['kind'];
 
         var stream =
@@ -155,8 +232,10 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
       case 'didOpenDataChannel':
         int dataChannelId = map['id'];
         String label = map['label'];
-        _dataChannel =
-            RTCDataChannelNative(_peerConnectionId, label, dataChannelId);
+        String flutterId = map['flutterId'];
+        _dataChannel = RTCDataChannelNative(
+            _peerConnectionId, label, dataChannelId, flutterId,
+            state: RTCDataChannelState.RTCDataChannelOpen);
         onDataChannel?.call(_dataChannel!);
         break;
       case 'onRenegotiationNeeded':
@@ -175,7 +254,8 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
             receiver: RTCRtpReceiverNative.fromMap(map['receiver'],
                 peerConnectionId: _peerConnectionId),
             streams: streams,
-            track: MediaStreamTrackNative.fromMap(map['track']),
+            track:
+                MediaStreamTrackNative.fromMap(map['track'], _peerConnectionId),
             transceiver: transceiver));
         break;
 
@@ -221,7 +301,7 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
   }
 
   EventChannel _eventChannelFor(String peerConnectionId) {
-    return EventChannel('FlutterWebRTC/peerConnectoinEvent$peerConnectionId');
+    return EventChannel('FlutterWebRTC/peerConnectionEvent$peerConnectionId');
   }
 
   @override
@@ -358,10 +438,14 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
 
   @override
   Future<void> addCandidate(RTCIceCandidate candidate) async {
-    await WebRTC.invokeMethod('addCandidate', <String, dynamic>{
-      'peerConnectionId': _peerConnectionId,
-      'candidate': candidate.toMap(),
-    });
+    try {
+      await WebRTC.invokeMethod('addCandidate', <String, dynamic>{
+        'peerConnectionId': _peerConnectionId,
+        'candidate': candidate.toMap(),
+      });
+    } on PlatformException catch (e) {
+      throw 'Unable to RTCPeerConnection::addCandidate: ${e.message}';
+    }
   }
 
   @override
@@ -369,16 +453,16 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
     try {
       final response = await WebRTC.invokeMethod('getStats', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
-        'track': track != null ? track.id : null
+        'trackId': track?.id
       });
 
       var stats = <StatsReport>[];
       if (response != null) {
         List<dynamic> reports = response['stats'];
-        reports.forEach((report) {
+        for (var report in reports) {
           stats.add(StatsReport(report['id'], report['type'],
-              report['timestamp'], report['values']));
-        });
+              (report['timestamp'] as num).toDouble(), report['values']));
+        }
       }
       return stats;
     } on PlatformException catch (e) {
@@ -407,8 +491,8 @@ class RTCPeerConnectionNative extends RTCPeerConnection {
         'dataChannelDict': dataChannelDict.toMap()
       });
 
-      _dataChannel =
-          RTCDataChannelNative(_peerConnectionId, label, response['id']);
+      _dataChannel = RTCDataChannelNative(
+          _peerConnectionId, label, response['id'], response['flutterId']);
       return _dataChannel!;
     } on PlatformException catch (e) {
       throw 'Unable to RTCPeerConnection::createDataChannel: ${e.message}';
