@@ -19,7 +19,7 @@ class RTCRtpSenderNative extends RTCRtpSender {
     return RTCRtpSenderNative(
         map['senderId'],
         (trackMap.isNotEmpty)
-            ? MediaStreamTrackNative.fromMap(map['track'])
+            ? MediaStreamTrackNative.fromMap(map['track'], peerConnectionId)
             : null,
         RTCDTMFSenderNative(peerConnectionId, map['senderId']),
         RTCRtpParameters.fromMap(map['rtpParameters']),
@@ -38,6 +38,7 @@ class RTCRtpSenderNative extends RTCRtpSender {
   String _peerConnectionId;
   String _id;
   MediaStreamTrack? _track;
+  final Set<MediaStream> _streams = {};
   RTCDTMFSender _dtmf;
   RTCRtpParameters _parameters;
   bool _ownsTrack = false;
@@ -47,19 +48,19 @@ class RTCRtpSenderNative extends RTCRtpSender {
     try {
       final response = await WebRTC.invokeMethod('getStats', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
-        if (track != null) 'track': track!.id,
+        if (track != null) 'trackId': track!.id,
       });
       var stats = <StatsReport>[];
       if (response != null) {
         List<dynamic> reports = response['stats'];
-        reports.forEach((report) {
+        for (var report in reports) {
           stats.add(StatsReport(report['id'], report['type'],
               report['timestamp'], report['values']));
-        });
+        }
       }
       return stats;
     } on PlatformException catch (e) {
-      throw 'Unable to RTCPeerConnection::getStats: ${e.message}';
+      throw 'Unable to RTCRtpSenderNative::getStats: ${e.message}';
     }
   }
 
@@ -75,39 +76,55 @@ class RTCRtpSenderNative extends RTCRtpSender {
       });
       return response['result'];
     } on PlatformException catch (e) {
-      throw 'Unable to RTCRtpSender::setParameters: ${e.message}';
+      throw 'Unable to RTCRtpSenderNative::setParameters: ${e.message}';
     }
   }
 
   @override
-  Future<void> replaceTrack(MediaStreamTrack track) async {
+  Future<void> replaceTrack(MediaStreamTrack? track) async {
     try {
       await WebRTC.invokeMethod('rtpSenderReplaceTrack', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
         'rtpSenderId': _id,
-        'trackId': track.id
+        'trackId': track != null ? track.id : ''
       });
 
       // change reference of associated MediaTrack
       _track = track;
     } on PlatformException catch (e) {
-      throw 'Unable to RTCRtpSender::replaceTrack: ${e.message}';
+      throw 'Unable to RTCRtpSenderNative::replaceTrack: ${e.message}';
     }
   }
 
   @override
-  Future<void> setTrack(MediaStreamTrack track,
+  Future<void> setTrack(MediaStreamTrack? track,
       {bool takeOwnership = true}) async {
     try {
       await WebRTC.invokeMethod('rtpSenderSetTrack', <String, dynamic>{
         'peerConnectionId': _peerConnectionId,
         'rtpSenderId': _id,
-        'trackId': track.id,
+        'trackId': track != null ? track.id : '',
         'takeOwnership': takeOwnership,
       });
 
       // change reference of associated MediaTrack
       _track = track;
+    } on PlatformException catch (e) {
+      throw 'Unable to RTCRtpSenderNative::setTrack: ${e.message}';
+    }
+  }
+
+  @override
+  Future<void> setStreams(List<MediaStream> streams) async {
+    try {
+      await WebRTC.invokeMethod('rtpSenderSetStreams', <String, dynamic>{
+        'peerConnectionId': _peerConnectionId,
+        'rtpSenderId': _id,
+        'streamIds': streams.map<String>((e) => e.id).toList(),
+      });
+
+      // change reference of associated MediaTrack
+      _streams.addAll(streams);
     } on PlatformException catch (e) {
       throw 'Unable to RTCRtpSender::setTrack: ${e.message}';
     }
@@ -132,16 +149,11 @@ class RTCRtpSenderNative extends RTCRtpSender {
   @override
   RTCDTMFSender get dtmfSender => _dtmf;
 
+  String get peerConnectionId => _peerConnectionId;
+
+  @Deprecated(
+      'No need to dispose rtpSender as it is handled by peerConnection.')
   @override
   @mustCallSuper
-  Future<void> dispose() async {
-    try {
-      await WebRTC.invokeMethod('rtpSenderDispose', <String, dynamic>{
-        'peerConnectionId': _peerConnectionId,
-        'rtpSenderId': _id,
-      });
-    } on PlatformException catch (e) {
-      throw 'Unable to RTCRtpSender::dispose: ${e.message}';
-    }
-  }
+  Future<void> dispose() async {}
 }
