@@ -9,16 +9,18 @@ import 'rtc_video_renderer_impl.dart';
 class RTCVideoView extends StatelessWidget {
   RTCVideoView(
     this._renderer, {
-    Key? key,
+    super.key,
     this.objectFit = RTCVideoViewObjectFit.RTCVideoViewObjectFitContain,
     this.mirror = false,
     this.filterQuality = FilterQuality.low,
-  }) : super(key: key);
+    this.placeholderBuilder,
+  });
 
   final RTCVideoRenderer _renderer;
   final RTCVideoViewObjectFit objectFit;
   final bool mirror;
   final FilterQuality filterQuality;
+  final WidgetBuilder? placeholderBuilder;
 
   RTCVideoRenderer get videoRenderer => _renderer;
 
@@ -26,10 +28,10 @@ class RTCVideoView extends StatelessWidget {
   Widget build(BuildContext context) {
     return LayoutBuilder(
         builder: (BuildContext context, BoxConstraints constraints) =>
-            _buildVideoView(constraints));
+            _buildVideoView(context, constraints));
   }
 
-  Widget _buildVideoView(BoxConstraints constraints) {
+  Widget _buildVideoView(BuildContext context, BoxConstraints constraints) {
     return Center(
       child: Container(
         width: constraints.maxWidth,
@@ -53,13 +55,12 @@ class RTCVideoView extends StatelessWidget {
               child: Transform(
                 transform: Matrix4.identity()..rotateY(mirror ? -pi : 0.0),
                 alignment: FractionalOffset.center,
-                child: videoRenderer.textureId != null &&
-                        videoRenderer.srcObject != null
+                child: videoRenderer.renderVideo
                     ? Texture(
                         textureId: videoRenderer.textureId!,
                         filterQuality: filterQuality,
                       )
-                    : Container(),
+                    : placeholderBuilder?.call(context) ?? Container(),
               ),
             ),
           ),
